add test for masher sound decoding to allow refactor/fix release static bug
