+++ test/masher_tests.cpp
@@ -1,0 +1,1803 @@
+#include <gmock/gmock.h>
+#include "oddlib/masher.hpp"
+#include "all_colours_high_compression_30_fps.ddv.g.h"
+#include "ddv_test1.ddv.g.h"
+#include "all_colours_low_compression_30_fps.ddv.g.h"
+#include "all_colours_max_compression_30_fps.ddv.g.h"
+#include "all_colours_medium_compression_30_fps.ddv.g.h"
+#include "all_colours_min_compression_30_fps.ddv.g.h"
+#include "mono_16_high_compression_all_samples.ddv.g.h"
+#include "mono_16_low_compression_all_samples.ddv.g.h"
+#include "mono_8_high_compression_all_samples.ddv.g.h"
+#include "mono_8_low_compression_all_samples.ddv.g.h"
+#include "stereo_16_high_compression_all_samples.ddv.g.h"
+#include "stereo_16_low_compression_all_samples.ddv.g.h"
+#include "stereo_8_high_compression_all_samples.ddv.g.h"
+#include "stereo_8_low_compression_all_samples.ddv.g.h"
+#include "expected_audio.dat.g.h"
+#include "compressed_audio.dat.g.h"
+
+const static u16 kLowCompression16BitExpected[0x3281] =
+{
+    0x07FC, 0x0001, 0xFE00, 0x0002, 0x03FF, 0x0001, 0x03FF, 0x03FF,
+    0xFE00, 0x0400, 0xFE00, 0x040C, 0x03FE, 0x03FE, 0x03FE, 0x07FF,
+    0x03FF, 0x0BFF, 0xFE00, 0x0402, 0x03FE, 0x03FE, 0x0401, 0x0BFF,
+    0x03FF, 0xFE00, 0x0416, 0x03FC, 0x03F9, 0x0002, 0x0001, 0x03FF,
+    0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0FFF, 0xFE00, 0x07FE, 0x07FC,
+    0x0403, 0xFE00, 0x07FA, 0x0001, 0x03FF, 0x0401, 0xFE00, 0x0406,
+    0x03FC, 0x0C01, 0xFE00, 0x0424, 0x03FA, 0x13FF, 0x1FFF, 0xFE00,
+    0x0400, 0x0FFF, 0x0FFF, 0xFE00, 0x040A, 0x03FE, 0x0003, 0x0003,
+    0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0002, 0x0001, 0xFE00, 0x0002,
+    0x03FF, 0x0001, 0x07FF, 0xFE00, 0x07F8, 0x0003, 0x0001, 0x07FF,
+    0x03FF, 0xFE00, 0x0402, 0x03FE, 0x03FE, 0x0401, 0x0BFF, 0x03FF,
+    0xFE00, 0x0414, 0x03FC, 0x03F4, 0x0004, 0x0003, 0x07FF, 0x07FF,
+    0x0001, 0x03FF, 0x0FFF, 0xFE00, 0x0406, 0x03FC, 0x0C01, 0xFE00,
+    0x042E, 0x03F6, 0x13FF, 0x1FFE, 0xFE00, 0x0002, 0x03FF, 0x0001,
+    0x03FF, 0x03FF, 0x0C01, 0xFE00, 0x0004, 0x03FE, 0x07FF, 0xFE00,
+    0x0400, 0x0FFF, 0x0FFF, 0xFE00, 0x0418, 0x03FC, 0x0002, 0x0003,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0001, 0x0001, 0x0FFF, 0x03FF,
+    0xFE00, 0x0402, 0x03FE, 0x03FE, 0x0401, 0x0BFF, 0x03FF, 0xFE00,
+    0x0420, 0x03FA, 0x03F6, 0x0004, 0x0002, 0x03FE, 0x03FE, 0x03FF,
+    0x03FF, 0x07FF, 0x0FFE, 0x03FF, 0xFE00, 0x07F6, 0x0002, 0x03FE,
+    0x0401, 0xFE00, 0x07FE, 0xFE00, 0x0406, 0x03FC, 0x0C01, 0xFE00,
+    0x0434, 0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x0406,
+    0x03FC, 0x0C01, 0xFE00, 0x043E, 0x03F6, 0x03FF, 0x03FF, 0x07FE,
+    0x03FD, 0x0BFF, 0x13FD, 0x03FF, 0xFE00, 0x07F2, 0x0004, 0x0003,
+    0x07FE, 0xFE00, 0x0008, 0x03FC, 0x03FD, 0x03FF, 0x0002, 0xFE00,
+    0x0406, 0x03FB, 0x03FF, 0x0801, 0x0BFF, 0x03FF, 0x13FF, 0xFE00,
+    0x043A, 0x03F6, 0x0405, 0x07FE, 0x03FE, 0x03FF, 0x03FF, 0x03FA,
+    0x0002, 0x0001, 0x0BFD, 0x03FF, 0x17FF, 0x3401, 0xFE00, 0x040C,
+    0x03F7, 0x0C02, 0x23FF, 0xFE00, 0x044E, 0x03F2, 0x0FFF, 0x03FE,
+    0x1FFC, 0x03FF, 0xFE00, 0x0000, 0x07FC, 0x0402, 0x13FF, 0xFE00,
+    0x0014, 0x03F8, 0x07FE, 0x0401, 0x0801, 0xFE00, 0x0406, 0x03FA,
+    0x0BFE, 0x0001, 0x0BFF, 0x17FF, 0xFE00, 0x0436, 0x03F6, 0x0007,
+    0x0005, 0x03FE, 0x07FF, 0x07FF, 0x0001, 0x13FD, 0xFE00, 0x0404,
+    0x03FC, 0x0BFF, 0x0001, 0xFE00, 0x042A, 0x03F8, 0x0002, 0x03FD,
+    0x03FE, 0x07FF, 0x1FFE, 0xFE00, 0x07A8, 0x0004, 0x003C, 0x03E8,
+    0x03F1, 0x0002, 0x07FF, 0x0006, 0x0001, 0x0001, 0x0BFF, 0x1BFF,
+    0xFE00, 0x00BE, 0x03FF, 0x039D, 0x0027, 0x07FA, 0x03FE, 0x0007,
+    0x07FE, 0x03FD, 0x07FF, 0x0002, 0x03FF, 0x0401, 0x0C02, 0x1401,
+    0xFE00, 0x0402, 0x03FE, 0xFE00, 0x0420, 0x03FB, 0x0FFF, 0x03FE,
+    0x1FFF, 0x03FF, 0xFE00, 0x0556, 0x03C4, 0x035A, 0x03B9, 0x0026,
+    0x03E2, 0x03E2, 0x0015, 0x000C, 0x0010, 0x0016, 0x03FD, 0x0007,
+    0x0010, 0x03EC, 0x03F5, 0x000A, 0x0004, 0x03FE, 0x03FB, 0x0004,
+    0x03FB, 0x07FE, 0x03FF, 0x0002, 0x0005, 0x03FC, 0x03FF, 0x0002,
+    0x0001, 0x07FF, 0x0401, 0x03FA, 0x0001, 0x0BFF, 0xFE00, 0x0608,
+    0x03E3, 0x030E, 0x039B, 0x0010, 0x07FE, 0x03FF, 0x0005, 0x001A,
+    0x0022, 0x03FE, 0x03FF, 0x0BFF, 0x0FFE, 0x0006, 0x03F7, 0x37F7,
+    0xFE00, 0x077A, 0x000E, 0x03C5, 0x03EC, 0x000B, 0x0002, 0x0001,
+    0x0001, 0x0004, 0x0003, 0x0003, 0x0801, 0x1801, 0x03FE, 0x37FE,
+    0xFE00, 0x00F6, 0x03FF, 0x0065, 0x0026, 0x03FF, 0x03F7, 0x03FD,
+    0x03F9, 0x07F9, 0x03FA, 0x07FF, 0x03FE, 0x03FF, 0x03FF, 0x03FF,
+    0x0FFE, 0x0003, 0x13FF, 0x2003, 0xFE00, 0x0678, 0x039C, 0x039F,
+    0x0083, 0x000A, 0x03C4, 0x03C7, 0x0012, 0x03F2, 0x039C, 0x004A,
+    0x000E, 0x03EF, 0x000B, 0x03DB, 0x03EA, 0x0007, 0x03F6, 0x000D,
+    0x03F3, 0x03D5, 0x001A, 0x0006, 0x03F8, 0x0009, 0x03FB, 0x0003,
+    0x03F9, 0x03FE, 0x0001, 0x03FE, 0x0004, 0x03FB, 0x0005, 0x03FB,
+    0x03F4, 0x0003, 0x03FD, 0x0003, 0x03FD, 0x0002, 0x03FF, 0x0801,
+    0x03FE, 0x0002, 0x03FE, 0x0002, 0x0001, 0x03FF, 0x0001, 0x03FF,
+    0x0801, 0x03FF, 0x0001, 0xFE00, 0x0764, 0x03DF, 0x03BF, 0x0071,
+    0x03F2, 0x0001, 0x03FE, 0x0408, 0x0399, 0x0054, 0x03FD, 0x0FFF,
+    0x13CC, 0x0021, 0x37F0, 0xFE00, 0x0296, 0x031F, 0x0030, 0x03FA,
+    0x03F8, 0x034D, 0x035F, 0x03F8, 0x0404, 0x03FF, 0x0BFA, 0x0395,
+    0x03C1, 0x03FD, 0x27FF, 0x03EB, 0x03F9, 0xFE00, 0x0398, 0x03FC,
+    0x0030, 0x03FC, 0x0002, 0x07FF, 0x0803, 0x03FF, 0xFE00, 0x0758,
+    0x0012, 0x001D, 0x07FA, 0x0002, 0x0001, 0x03FF, 0x0401, 0xFE00,
+    0x012A, 0x03FF, 0x03CF, 0x0BF5, 0x03FC, 0x0003, 0x07FE, 0x0C01,
+    0x03FF, 0x03FF, 0x2FFF, 0xFE00, 0x0150, 0x0344, 0x0050, 0x03FE,
+    0x03F4, 0x0372, 0x037F, 0x03F3, 0x0405, 0x0FF8, 0x03AB, 0x03CE,
+    0x03FA, 0x07FF, 0x1FFE, 0x03F0, 0x03FA, 0x03FF, 0xFE00, 0x0258,
+    0x03F0, 0x004F, 0x03FF, 0x0004, 0x07FF, 0x0806, 0xFE00, 0x07CA,
+    0x0372, 0x0050, 0x07F2, 0x039E, 0x03A5, 0x03F3, 0x0406, 0x07FF,
+    0x03FF, 0x03F8, 0x03C5, 0x03DD, 0x03FA, 0x27FD, 0x03F5, 0x03FC,
+    0x03FF, 0xFE00, 0x00D6, 0x03E2, 0x004E, 0x0404, 0x07FC, 0x03FF,
+    0x0405, 0x17FF, 0xFE00, 0x06C2, 0x0021, 0x03FF, 0x0005, 0x0405,
+    0x0003, 0x07FF, 0x1401, 0x0001, 0xFE00, 0x0232, 0x03FE, 0x03FF,
+    0x03F5, 0x07EC, 0x03F9, 0x1FFD, 0x03FE, 0x2FFE, 0xFE00, 0x0682,
+    0x0398, 0x0032, 0x0004, 0x03F8, 0x03C4, 0x03C6, 0x03F8, 0x03FF,
+    0x0003, 0x0FFA, 0x03DB, 0x03EA, 0x03FC, 0x07FF, 0x1FFF, 0x03F9,
+    0x03FE, 0xFE00, 0x0792, 0x03D5, 0x0032, 0x0004, 0x0001, 0x07FC,
+    0x0803, 0x17FF, 0xFE00, 0x05FC, 0x03AA, 0x0015, 0x03FE, 0x03FA,
+    0x03D3, 0x03D3, 0x03FE, 0x0401, 0x0FFE, 0x03E3, 0x03EE, 0x03FF,
+    0x2BFA, 0x03FE, 0xFE00, 0x06FA, 0x03D5, 0x0020, 0x03FC, 0x03FE,
+    0x0001, 0x03FB, 0x0803, 0x03FF, 0x13FF, 0xFE00, 0x075A, 0x0011,
+    0x03E4, 0x0405, 0x0002, 0x0002, 0x0BFF, 0xFE00, 0x012C, 0x03FF,
+    0x0032, 0x0BF5, 0x03FC, 0x03FD, 0x0402, 0x0FFF, 0x03FF, 0x03FF,
+    0x2FFF, 0xFE00, 0x0568, 0x03C3, 0x0025, 0x07F6, 0x03E0, 0x03DF,
+    0x03FC, 0x03FF, 0x0004, 0x07FF, 0x03FF, 0x03FD, 0x03EC, 0x03F4,
+    0x03FE, 0x27FF, 0x03FC, 0x03FF, 0xFE00, 0x061A, 0x03E2, 0x0037,
+    0x07FC, 0x07FD, 0x0805, 0xFE00, 0x04B4, 0x03E2, 0x0025, 0x07F7,
+    0x03F0, 0x03F0, 0x03FB, 0x0403, 0x0FFD, 0x03F5, 0x03FA, 0x03FE,
+    0x27FF, 0x03FE, 0xFE00, 0x050E, 0x03F0, 0x0036, 0x07FC, 0x07FE,
+    0x0805, 0xFE00, 0x07F6, 0x07FB, 0x03FE, 0x0001, 0xFE00, 0x0012,
+    0x0409, 0x0004, 0x07FF, 0x0C01, 0xFE00, 0x0420, 0x03FB, 0x0012,
+    0x0004, 0x03FC, 0x03FE, 0x03FD, 0x03FE, 0x03FF, 0x0002, 0x0FFE,
+    0x03FE, 0x03FF, 0x03FF, 0xFE00, 0x0432, 0x03FD, 0x001E, 0x0005,
+    0x03FE, 0x1002, 0xFE00, 0x0400, 0xFE00, 0x0400, 0xFE00, 0x07EA,
+    0x0002, 0x0003, 0x03FF, 0x03FF, 0xFE00, 0x001A, 0x03FD, 0x0008,
+    0x03F9, 0x03FE, 0x0C01, 0x0001, 0xFE00, 0x041A, 0x03FE, 0x03FC,
+    0x03FC, 0x07FF, 0x03FF, 0x0BFE, 0x03FF, 0xFE00, 0x0428, 0x03FE,
+    0x03FB, 0x03FA, 0x07FF, 0x03FF, 0x0BFD, 0x03FE, 0x27FF, 0xFE00,
+    0x042E, 0x03FC, 0x03F3, 0x0005, 0x0002, 0x03FF, 0x03FF, 0x07FF,
+    0x0001, 0x03FE, 0x0FFF, 0xFE00, 0x0446, 0x03FC, 0x03EC, 0x0008,
+    0x0002, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x0001, 0x03FD, 0x0FFF,
+    0xFE00, 0x0002, 0x07EC, 0x0002, 0x0004, 0x1001, 0xFE00, 0x07E4,
+    0x0003, 0x03F9, 0x0401, 0xFE00, 0x044A, 0x03F9, 0x0FFF, 0x03FF,
+    0x1FFE, 0xFE00, 0x046E, 0x03F8, 0x33FE, 0xFE00, 0x0416, 0x03FE,
+    0x0007, 0x0007, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0004, 0x0002,
+    0x2401, 0xFE00, 0x0422, 0x03FE, 0x000A, 0x000B, 0x03FF, 0x03FF,
+    0x03FF, 0x07FF, 0x0006, 0x0003, 0x2401, 0xFE00, 0x001C, 0x03FD,
+    0x0007, 0x0003, 0x03FE, 0xFE00, 0x07DA, 0x0004, 0x0007, 0x0002,
+    0x03FF, 0xFE00, 0x042A, 0x03FC, 0x03E9, 0x0009, 0x0002, 0x03FF,
+    0x03FF, 0x07FF, 0x0002, 0x03FD, 0x0FFF, 0xFE00, 0x0442, 0x03FC,
+    0x03DD, 0x000E, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0003, 0x03FB,
+    0x0FFF, 0x1801, 0x37FF, 0xFE00, 0x045A, 0x03F9, 0x0FFF, 0x03FF,
+    0x1FFE, 0xFE00, 0x048A, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF,
+    0xFE00, 0x0010, 0x03FF, 0x0009, 0x03F9, 0x03FE, 0x0C01, 0x0001,
+    0xFE00, 0x0018, 0x03FE, 0x07FB, 0xFE00, 0x0432, 0x03FC, 0x0005,
+    0x0005, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0003, 0x0001, 0x0FFF,
+    0x03FF, 0xFE00, 0x0452, 0x03FC, 0x0007, 0x0008, 0x03FF, 0x03FE,
+    0x03FE, 0x07FF, 0x0004, 0x0002, 0x0FFE, 0x03FF, 0x1001, 0xFE00,
+    0x0442, 0x03FB, 0x03EE, 0x0007, 0x0002, 0x03FE, 0x03FE, 0x07FF,
+    0x0001, 0x03FE, 0x0FFE, 0x03FF, 0xFE00, 0x046A, 0x03FA, 0x03E6,
+    0x000A, 0x0002, 0x03FE, 0x03FE, 0x07FF, 0x0002, 0x03FD, 0x0FFE,
+    0x03FF, 0x4FFF, 0xFE00, 0x07DA, 0x0004, 0x03F4, 0x0001, 0x0002,
+    0x1001, 0xFE00, 0x0000, 0xFE00, 0x046A, 0x03F8, 0x0FFF, 0x03FE,
+    0x1FFE, 0x03FF, 0xFE00, 0x04A0, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD,
+    0x03FF, 0xFE00, 0x047C, 0x03F6, 0x03FE, 0x03FD, 0x07FE, 0x03FD,
+    0x0BFE, 0x03FF, 0x0FFD, 0x03FF, 0xFE00, 0x04B8, 0x03F5, 0x03FD,
+    0x03FC, 0x07FF, 0x03FE, 0x0BFE, 0x03FF, 0x0FFD, 0x03FF, 0xFE00,
+    0x07C6, 0x0005, 0x0010, 0x0004, 0x03FD, 0x0FFF, 0x03FF, 0xFE00,
+    0x002A, 0x03FC, 0x03F3, 0x03FC, 0x0002, 0xFE00, 0x0476, 0x03F6,
+    0x0002, 0x000B, 0x07FE, 0x03FE, 0x03FF, 0x03FF, 0x03F3, 0x0005,
+    0x0001, 0x0BFD, 0x03FF, 0x1001, 0x03FD, 0x3402, 0xFE00, 0x04AE,
+    0x03F5, 0x0003, 0x0010, 0x07FF, 0x03FE, 0x03FF, 0x03FF, 0x03EE,
+    0x0007, 0x0001, 0x0BFD, 0x03FF, 0x1001, 0x03FB, 0x3403, 0xFE00,
+    0x049A, 0x03F0, 0x13FE, 0x1FFC, 0x03FF, 0xFE00, 0x04E2, 0x03F1,
+    0x13FF, 0x1FFC, 0xFE00, 0x000E, 0x03FF, 0x03F0, 0x0403, 0x13FB,
+    0x2BFF, 0xFE00, 0x005A, 0x03F7, 0x0003, 0x03FA, 0x1401, 0x03FF,
+    0xFE00, 0x046C, 0x03F5, 0x000D, 0x000A, 0x03FE, 0x07FF, 0x07FF,
+    0x0003, 0x13FD, 0x17FF, 0xFE00, 0x04A0, 0x03F7, 0x0014, 0x0010,
+    0x03FE, 0x07FF, 0x07FF, 0x0004, 0x13FE, 0x17FF, 0x03FF, 0xFE00,
+    0x0458, 0x03F8, 0x0005, 0x03FC, 0x03FF, 0x07FF, 0x0801, 0x13FE,
+    0xFE00, 0x0482, 0x03F9, 0x0007, 0x03FA, 0x07FF, 0x03FF, 0x0801,
+    0x13FE, 0xFE00, 0x078A, 0x0003, 0x007B, 0x03D1, 0x03F0, 0x0C06,
+    0x0001, 0x0004, 0x27FD, 0x3801, 0xFE00, 0x008A, 0x0008, 0x03C4,
+    0x0018, 0x03F4, 0x0C05, 0x03FE, 0x03FF, 0x2401, 0xFE00, 0x0446,
+    0x03FA, 0x0FFE, 0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x046A, 0x03F9,
+    0x0FFF, 0x03FF, 0x1FFE, 0xFE00, 0x06BE, 0x03DC, 0x02C3, 0x037E,
+    0x0015, 0x03FF, 0x03FC, 0x0406, 0x0025, 0x002D, 0x03FE, 0x13FF,
+    0x0BFD, 0x0007, 0x03F4, 0x37F4, 0x0001, 0xFE00, 0x0778, 0x03DF,
+    0x0276, 0x0360, 0x0012, 0x07FD, 0x0405, 0x002F, 0x0039, 0x03FE,
+    0x0BFF, 0x13FD, 0x0009, 0x03F0, 0x37F0, 0xFE00, 0x0706, 0x000E,
+    0x0398, 0x03DA, 0x000B, 0x0C04, 0x0007, 0x0006, 0x2402, 0x03FD,
+    0x37FD, 0xFE00, 0x009E, 0x000E, 0x0042, 0x0018, 0x000B, 0x0C04,
+    0x03FC, 0x03FD, 0x27FF, 0x0002, 0x3402, 0xFE00, 0x0032, 0x03D6,
+    0x03FF, 0x003E, 0x03EE, 0x07FC, 0x040A, 0x03A8, 0x0054, 0x03FC,
+    0x07FF, 0x1BC8, 0x0025, 0x33FF, 0x03EE, 0xFE00, 0x0108, 0x03DB,
+    0x0040, 0x000A, 0x03F0, 0x07FD, 0x0409, 0x03B8, 0x0054, 0x03FD,
+    0x07FF, 0x03FF, 0x17C4, 0x002A, 0x33FF, 0x03EC, 0xFE00, 0x03AE,
+    0x03FC, 0x0025, 0x03FB, 0x0003, 0x1002, 0x03FF, 0x07FF, 0xFE00,
+    0x03C2, 0x03FC, 0x0019, 0x03FD, 0x0002, 0x1001, 0x03FF, 0xFE00,
+    0x06D0, 0x0011, 0x0033, 0x07FB, 0x1002, 0xFE00, 0x00C0, 0x0011,
+    0x03E0, 0x07FB, 0x13FF, 0xFE00, 0x02B4, 0x03ED, 0x003C, 0x0405,
+    0x07FF, 0x0804, 0xFE00, 0x0314, 0x03EE, 0x002A, 0x0404, 0x07FF,
+    0x03FF, 0x0403, 0xFE00, 0x0186, 0x03DC, 0x003D, 0x0404, 0x07FD,
+    0x0804, 0x17FF, 0xFE00, 0x0242, 0x03DE, 0x002A, 0x0403, 0x03FF,
+    0x03FC, 0x07FF, 0x0003, 0x17FF, 0xFE00, 0x05C4, 0x0020, 0x03FE,
+    0x000A, 0x0801, 0x07FF, 0xFE00, 0x016E, 0x0020, 0x03FE, 0x03F9,
+    0x0801, 0x07FF, 0xFE00, 0x008C, 0x03CE, 0x0026, 0x0004, 0x0002,
+    0x07FD, 0x0802, 0x0001, 0x13FF, 0xFE00, 0x0192, 0x03D1, 0x001A,
+    0x0003, 0x0002, 0x03FF, 0x03FB, 0x0802, 0x13FF, 0xFE00, 0x07F4,
+    0x03CE, 0x002C, 0x03FB, 0x03FD, 0x07FC, 0x0804, 0x03FF, 0x13FF,
+    0xFE00, 0x00F6, 0x03D0, 0x0038, 0x03FB, 0x03FE, 0x03FF, 0x03FB,
+    0x0805, 0x03FF, 0xFE00, 0x06D2, 0x0011, 0x03CE, 0x0405, 0x13FE,
+    0xFE00, 0x00C4, 0x0011, 0x0021, 0x0405, 0x1001, 0xFE00, 0x06CC,
+    0x03DD, 0x0049, 0x07FB, 0x07FD, 0x0806, 0x0BFF, 0x0BFF, 0xFE00,
+    0x0786, 0x03DE, 0x0059, 0x07FC, 0x03FF, 0x03FE, 0x07FF, 0x0007,
+    0xFE00, 0x056A, 0x03EE, 0x0048, 0x07FC, 0x07FE, 0x03FF, 0x0406,
+    0x0BFF, 0xFE00, 0x05CC, 0x03EE, 0x005A, 0x07FB, 0x07FE, 0x0808,
+    0xFE00, 0x07EC, 0x0001, 0x03F6, 0x03FB, 0x0001, 0x13FF, 0xFE00,
+    0x000A, 0x0001, 0x0005, 0x0002, 0x0001, 0xFE00, 0x0446, 0x03FB,
+    0x0029, 0x0005, 0x03FD, 0x07FF, 0x0803, 0x0001, 0xFE00, 0x045C,
+    0x03FC, 0x0036, 0x0007, 0x03FD, 0x1003, 0x0001, 0xFE00, 0x0400,
+    0xFE00, 0x0400, 0xFE00, 0x07D6, 0x0001, 0x0006, 0x03FE, 0x1401,
+    0xFE00, 0x0038, 0x03FC, 0x000F, 0x03F2, 0x03FF, 0x0C01, 0x0002,
+    0x03FF, 0xFE00, 0x043C, 0x03FE, 0x03F8, 0x03F7, 0x07FE, 0x03FE,
+    0x0BFB, 0x03FD, 0x0FFF, 0x03FF, 0x13FF, 0xFE00, 0x044A, 0x03FE,
+    0x03F6, 0x03F4, 0x07FF, 0x03FF, 0x0BFA, 0x03FD, 0x27FF, 0xFE00,
+    0x0462, 0x03FC, 0x03E8, 0x0009, 0x0001, 0x03FF, 0x03FF, 0x03FF,
+    0x0402, 0x03FD, 0x0FFF, 0xFE00, 0x047A, 0x03FC, 0x03E3, 0x000C,
+    0x0002, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x0002, 0x03FC, 0x0FFF,
+    0x53FF, 0xFE00, 0x0006, 0x07D9, 0x0005, 0x0003, 0x1003, 0x03FF,
+    0xFE00, 0x07CC, 0x0002, 0x03F4, 0x0001, 0x0001, 0x1001, 0xFE00,
+    0x0494, 0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x04B6, 0x03F8, 0x33FE,
+    0xFE00, 0x042C, 0x03FE, 0x000E, 0x000F, 0x03FF, 0x03FF, 0x03FF,
+    0x07FF, 0x0008, 0x0004, 0x2402, 0x0001, 0xFE00, 0x0438, 0x03FE,
+    0x0011, 0x0012, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x000A, 0x0006,
+    0x2402, 0x0001, 0xFE00, 0x003C, 0x03FD, 0x000D, 0x0005, 0x03FF,
+    0xFE00, 0x07B6, 0x0004, 0x000D, 0x0005, 0x03FF, 0xFE00, 0x0458,
+    0x03FC, 0x03D1, 0x0014, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0004,
+    0x03FA, 0x0FFF, 0x1401, 0x0001, 0x37FF, 0xFE00, 0x046E, 0x03FC,
+    0x03C6, 0x0018, 0x0003, 0x07FF, 0x07FF, 0x0005, 0x03F8, 0x0FFF,
+    0x1401, 0x0001, 0x37FE, 0xFE00, 0x04BA, 0x03F8, 0x0FFF, 0x03FE,
+    0x1FFE, 0x03FF, 0xFE00, 0x04E6, 0x03F6, 0x13FF, 0x1FFE, 0xFE00,
+    0x0024, 0x03FE, 0x0013, 0x03F2, 0x03FE, 0x0C01, 0x0003, 0x03FF,
+    0xFE00, 0x0036, 0x03FD, 0x07F7, 0x1BFF, 0xFE00, 0x0474, 0x03FC,
+    0x0009, 0x0009, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0005, 0x0003,
+    0x0FFE, 0x03FF, 0x1001, 0xFE00, 0x0490, 0x03FC, 0x000B, 0x000C,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0007, 0x0004, 0x0FFF, 0x03FF,
+    0x1001, 0xFE00, 0x0496, 0x03FA, 0x03DE, 0x000E, 0x0002, 0x03FE,
+    0x03FE, 0x03FF, 0x03FF, 0x0003, 0x03FC, 0x0FFE, 0x03FF, 0x1401,
+    0x37FF, 0xFE00, 0x04B8, 0x03FB, 0x03D5, 0x0012, 0x0002, 0x03FE,
+    0x03FE, 0x07FF, 0x0004, 0x03FA, 0x0FFE, 0x03FF, 0x1001, 0x0001,
+    0x37FF, 0xFE00, 0x07BA, 0x0003, 0x03E8, 0x0003, 0x0002, 0x1002,
+    0xFE00, 0x0006, 0x0402, 0xFE00, 0x04DC, 0x03F5, 0x0FFF, 0x03FE,
+    0x1FFD, 0x03FF, 0xFE00, 0x0512, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE,
+    0x03FF, 0xFE00, 0x04FA, 0x03F6, 0x03FB, 0x03FC, 0x07FE, 0x03FC,
+    0x0BFD, 0x03FF, 0x0FFD, 0x03FF, 0xFE00, 0x053A, 0x03F6, 0x03FB,
+    0x03FA, 0x07FE, 0x03FD, 0x0BFD, 0x03FE, 0x0FFD, 0x03FF, 0x13FF,
+    0xFE00, 0x0796, 0x0005, 0x001E, 0x0009, 0x03FD, 0x0FFF, 0x03FF,
+    0x03FF, 0x6001, 0xFE00, 0x0050, 0x03FC, 0x03E8, 0x03F9, 0x0002,
+    0x1001, 0x0001, 0xFE00, 0x04EA, 0x03F6, 0x0007, 0x0016, 0x07FE,
+    0x03FE, 0x03FF, 0x03FF, 0x03E7, 0x000A, 0x0001, 0x0BFD, 0x03FF,
+    0x1002, 0x03F9, 0x3405, 0xFE00, 0x0524, 0x03F6, 0x0009, 0x001B,
+    0x07FE, 0x03FE, 0x03FF, 0x03FF, 0x03E1, 0x000D, 0x0001, 0x0BFD,
+    0x03FF, 0x1003, 0x03F7, 0x3406, 0xFE00, 0x052E, 0x03F2, 0x0FFF,
+    0x03FE, 0x1FFC, 0x03FF, 0xFE00, 0x0576, 0x03F2, 0x0FFF, 0x03FE,
+    0x1FFC, 0x03FF, 0xFE00, 0x001C, 0x03FF, 0x03E3, 0x03FF, 0x0003,
+    0x13F7, 0x2BFD, 0x3BFF, 0xFE00, 0x00A6, 0x03F7, 0x0005, 0x03F5,
+    0x1001, 0x0001, 0x03FE, 0x2BFF, 0xFE00, 0x04D6, 0x03F6, 0x001A,
+    0x0015, 0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x0006, 0x13FD, 0x03FF,
+    0x13FF, 0x03FF, 0x37FF, 0xFE00, 0x050C, 0x03F6, 0x0020, 0x0019,
+    0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x0007, 0x13FD, 0x03FF, 0x13FE,
+    0x03FE, 0x37FF, 0xFE00, 0x04B0, 0x03F8, 0x0008, 0x03F8, 0x07FF,
+    0x03FE, 0x0802, 0x13FE, 0x03FF, 0xFE00, 0x04DC, 0x03F8, 0x000B,
+    0x03F7, 0x03FF, 0x03FF, 0x03FE, 0x0802, 0x13FE, 0x03FF, 0xFE00,
+    0x0780, 0x03FD, 0x00B0, 0x03BE, 0x03F8, 0x0001, 0x07FF, 0x0003,
+    0x0001, 0x0006, 0x27FC, 0x3802, 0xFE00, 0x0054, 0x0004, 0x03F0,
+    0x0007, 0x03F9, 0x0C02, 0x03FF, 0xFE00, 0x0492, 0x03F9, 0x0FFF,
+    0x03FF, 0x1FFE, 0xFE00, 0x04B8, 0x03FB, 0x0FFF, 0x03FE, 0x1FFF,
+    0x03FF, 0xFE00, 0x0018, 0x03E6, 0x0236, 0x0348, 0x000D, 0x03FF,
+    0x03FC, 0x0404, 0x0039, 0x0043, 0x03FE, 0x1FFE, 0x000A, 0x03ED,
+    0x37ED, 0xFE00, 0x0064, 0x0002, 0x021E, 0x0340, 0x03FC, 0x03FC,
+    0x03FF, 0x0001, 0x03FE, 0x003E, 0x0047, 0x0FFF, 0x03FF, 0x100B,
+    0x03EB, 0x37EC, 0xFE00, 0x06AC, 0x0005, 0x0374, 0x03CC, 0x0004,
+    0x0002, 0x0401, 0x0001, 0x000A, 0x0008, 0x2402, 0x03FC, 0x37FB,
+    0xFE00, 0x0040, 0x0007, 0x001C, 0x000A, 0x0006, 0x0C02, 0x03FE,
+    0x03FF, 0x6001, 0xFE00, 0x01BA, 0x03E3, 0x0078, 0x03DF, 0x03F4,
+    0x03FF, 0x03FC, 0x0407, 0x03C6, 0x0054, 0x03FE, 0x07FF, 0x1BC0,
+    0x002D, 0x37EA, 0xFE00, 0x020C, 0x0005, 0x0091, 0x03CA, 0x0002,
+    0x03FC, 0x03FF, 0x03FE, 0x03FF, 0x03CC, 0x0054, 0x0401, 0x07FF,
+    0x17BE, 0x002F, 0x37E9, 0xFE00, 0x03D6, 0x03FC, 0x000E, 0x03FC,
+    0x0003, 0x0FFF, 0x0001, 0xFE00, 0x03DE, 0x0409, 0x03FF, 0x1401,
+    0xFE00, 0x0662, 0x0007, 0x0044, 0x07FF, 0x0002, 0x0C03, 0xFE00,
+    0x004E, 0x0008, 0x03F2, 0x07FD, 0x0001, 0xFE00, 0x0366, 0x03F2,
+    0x001C, 0x0402, 0x03FF, 0x03FF, 0x0802, 0xFE00, 0x038E, 0x0413,
+    0x03FF, 0x03FF, 0x03FF, 0x0C01, 0xFE00, 0x02DE, 0x03E7, 0x001B,
+    0x03FF, 0x0003, 0x03FF, 0x03FE, 0x0803, 0x0BFF, 0xFE00, 0x032A,
+    0x0002, 0x0015, 0x0BFC, 0x03FF, 0x0802, 0x17FF, 0xFE00, 0x04FC,
+    0x000C, 0x03FE, 0x000E, 0x0405, 0xFE00, 0x009A, 0x0012, 0x07FE,
+    0x03FF, 0x0001, 0x0001, 0xFE00, 0x026E, 0x03DD, 0x000F, 0x0402,
+    0x03FE, 0x03FC, 0x0801, 0x1FFF, 0xFE00, 0x02D6, 0x0004, 0x000C,
+    0x07FF, 0x03FB, 0x03FF, 0x0801, 0x0BFF, 0x07FF, 0xFE00, 0x01D6,
+    0x03DC, 0x0041, 0x03F9, 0x03FE, 0x03FF, 0x03FC, 0x0806, 0x03FF,
+    0x13FF, 0xFE00, 0x0238, 0x0006, 0x0047, 0x03F8, 0x07FC, 0x03FF,
+    0x03FF, 0x0405, 0x03FF, 0x0FFE, 0xFE00, 0x0664, 0x0006, 0x03BC,
+    0x0401, 0x0002, 0x0FFE, 0xFE00, 0x004E, 0x000A, 0x000E, 0x0403,
+    0xFE00, 0x0026, 0x03E6, 0x006A, 0x03FF, 0x03FC, 0x03FF, 0x03FE,
+    0x03FF, 0x03FF, 0x0008, 0x0BFF, 0xFE00, 0x006C, 0x0004, 0x0071,
+    0x0BFD, 0x03FF, 0x0807, 0xFE00, 0x061E, 0x03F3, 0x006A, 0x07FC,
+    0x07FE, 0x03FF, 0x0408, 0x17FF, 0xFE00, 0x0644, 0x0002, 0x0072,
+    0x0BFF, 0x03FF, 0x03FF, 0x03FF, 0x0007, 0xFE00, 0x07E4, 0x07F1,
+    0x03F9, 0x17FF, 0xFE00, 0x0002, 0xFE00, 0x0470, 0x03FD, 0x0040,
+    0x0008, 0x03FE, 0x1005, 0x0001, 0xFE00, 0x047A, 0x0445, 0x000A,
+    0x13FF, 0x0005, 0x0001, 0xFE00, 0x0400, 0xFE00, 0x0400, 0xFE00,
+    0x07BE, 0x0002, 0x0008, 0x03FE, 0x03FF, 0x1002, 0xFE00, 0x0054,
+    0x03FD, 0x0016, 0x03EB, 0x03FE, 0x0C01, 0x0003, 0x03FE, 0xFE00,
+    0x045E, 0x03FE, 0x03F3, 0x03F2, 0x07FE, 0x03FE, 0x0BF8, 0x03FC,
+    0x0FFF, 0x03FF, 0x13FE, 0x03FF, 0xFE00, 0x046E, 0x03FE, 0x03F0,
+    0x03F0, 0x07FF, 0x03FF, 0x0BF7, 0x03FB, 0x27FE, 0x03FF, 0xFE00,
+    0x0496, 0x03FC, 0x03DE, 0x000E, 0x0001, 0x03FF, 0x03FF, 0x03FF,
+    0x0403, 0x03FC, 0x0FFF, 0x1801, 0x37FF, 0xFE00, 0x04AE, 0x03FC,
+    0x03D8, 0x0010, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0004, 0x03FB,
+    0x0FFF, 0x1401, 0x0001, 0x37FF, 0xFE00, 0x000A, 0x07C8, 0x0008,
+    0x0004, 0x1005, 0x03FF, 0xFE00, 0x07B2, 0x0004, 0x03EF, 0x0002,
+    0x0001, 0x1001, 0xFE00, 0x04DC, 0x03FA, 0x13FF, 0x1FFF, 0xFE00,
+    0x0502, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE, 0xFE00, 0x044A, 0x03FE,
+    0x0014, 0x0015, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x000C, 0x0006,
+    0x0FFF, 0x1403, 0x0001, 0xFE00, 0x0456, 0x03FE, 0x0018, 0x0019,
+    0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x000E, 0x0008, 0x2403, 0x0002,
+    0xFE00, 0x005A, 0x03FD, 0x0014, 0x0008, 0x03FF, 0xFE00, 0x0792,
+    0x0004, 0x0014, 0x0007, 0x03FF, 0xFE00, 0x0488, 0x03FC, 0x03BD,
+    0x001C, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0006, 0x03F7, 0x0FFF,
+    0x1401, 0x0002, 0x37FE, 0xFE00, 0x04A0, 0x03FC, 0x03B1, 0x0021,
+    0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0007, 0x03F5, 0x0FFF, 0x1402,
+    0x0002, 0x37FE, 0xFE00, 0x0512, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE,
+    0xFE00, 0x0542, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF, 0xFE00,
+    0x003A, 0x03FE, 0x001D, 0x03EB, 0x03FE, 0x0C01, 0x0005, 0x03FF,
+    0xFE00, 0x0052, 0x03FD, 0x07F3, 0x17FF, 0x03FF, 0xFE00, 0x04B0,
+    0x03FC, 0x000D, 0x000E, 0x07FE, 0x03FE, 0x0808, 0x0004, 0x0FFE,
+    0x03FF, 0x1002, 0x0001, 0xFE00, 0x04CC, 0x03FC, 0x0010, 0x0010,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0009, 0x0005, 0x0FFF, 0x03FF,
+    0x1002, 0x0001, 0xFE00, 0x04E0, 0x03FA, 0x03CE, 0x0014, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0005, 0x03F9, 0x0FFE, 0x03FF, 0x1001,
+    0x0001, 0x37FF, 0xFE00, 0x0506, 0x03FB, 0x03C4, 0x0019, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0005, 0x03F8, 0x0FFE, 0x03FF, 0x1001,
+    0x0002, 0x37FE, 0xFE00, 0x079C, 0x0004, 0x03DE, 0x0004, 0x0002,
+    0x1003, 0xFE00, 0x000C, 0x0404, 0xFE00, 0x0548, 0x03F7, 0x0FFE,
+    0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x0580, 0x03F7, 0x0FFE, 0x03FE,
+    0x1FFD, 0x03FF, 0xFE00, 0x0578, 0x03F6, 0x03F9, 0x03F9, 0x07FD,
+    0x03FC, 0x0BFC, 0x03FE, 0x0FFD, 0x03FF, 0x13FF, 0xFE00, 0x05B8,
+    0x03F6, 0x03F8, 0x03F8, 0x07FD, 0x03FC, 0x0BFB, 0x03FE, 0x0FFD,
+    0x03FF, 0x13FF, 0xFE00, 0x0768, 0x0005, 0x002C, 0x000D, 0x03FD,
+    0x0FFF, 0x03FE, 0x03FE, 0x2801, 0x3401, 0xFE00, 0x0074, 0x03FC,
+    0x03DE, 0x03F6, 0x0002, 0x1001, 0x0001, 0x2BFF, 0x37FF, 0xFE00,
+    0x0560, 0x03F6, 0x000B, 0x0020, 0x07FE, 0x03FE, 0x03FF, 0x03FF,
+    0x03DB, 0x0010, 0x0001, 0x0BFD, 0x03FF, 0x1003, 0x03F6, 0x3407,
+    0xFE00, 0x059A, 0x03F6, 0x000D, 0x0025, 0x07FF, 0x03FE, 0x03FF,
+    0x03FF, 0x03D5, 0x0012, 0x0001, 0x0BFD, 0x03FF, 0x1004, 0x03F4,
+    0x3409, 0xFE00, 0x05C2, 0x03F0, 0x13FE, 0x1FFC, 0x03FF, 0xFE00,
+    0x060A, 0x03F1, 0x13FF, 0x1FFC, 0xFE00, 0x002C, 0x03FE, 0x03D6,
+    0x03FF, 0x0003, 0x13F3, 0x2BFC, 0x3BFF, 0xFE00, 0x00F0, 0x03F7,
+    0x0008, 0x03F1, 0x1001, 0x0002, 0x03FE, 0x2BFF, 0xFE00, 0x0544,
+    0x03F6, 0x0026, 0x001E, 0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x0009,
+    0x13FD, 0x03FF, 0x13FE, 0x03FE, 0x37FF, 0xFE00, 0x057A, 0x03F6,
+    0x002C, 0x0023, 0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x000A, 0x13FD,
+    0x03FF, 0x13FD, 0x03FE, 0x37FF, 0xFE00, 0x050A, 0x03F8, 0x000D,
+    0x03F5, 0x03FF, 0x03FE, 0x03FE, 0x0803, 0x13FE, 0x03FF, 0x13FF,
+    0xFE00, 0x0536, 0x03F8, 0x0010, 0x03F2, 0x03FF, 0x03FF, 0x03FE,
+    0x0804, 0x13FE, 0x03FF, 0x13FF, 0x0001, 0xFE00, 0x07CA, 0x03F4,
+    0x00BA, 0x03BC, 0x1807, 0x27FC, 0x3802, 0xFE00, 0x0022, 0x0009,
+    0x0017, 0x03F8, 0x03F4, 0x0C05, 0x0401, 0xFE00, 0x04DE, 0x03FA,
+    0x0FFE, 0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x0500, 0x03FB, 0x0FFF,
+    0x03FE, 0x1FFF, 0x03FF, 0xFE00, 0x003C, 0x000B, 0x023F, 0x0352,
+    0x03F5, 0x0BFF, 0x03FC, 0x003E, 0x0044, 0x0BFF, 0x03FE, 0x140A,
+    0x03EC, 0x37ED, 0xFE00, 0x07F6, 0x000D, 0x026D, 0x0368, 0x03F4,
+    0x03FF, 0x0001, 0x07FC, 0x003C, 0x003F, 0x0BFF, 0x0BFF, 0x0801,
+    0x0008, 0x03EC, 0x37EE, 0xFE00, 0x06BE, 0x03FA, 0x037C, 0x03CF,
+    0x03FB, 0x0FFE, 0x0009, 0x0007, 0x2402, 0x03FC, 0x37FB, 0xFE00,
+    0x07EC, 0x000E, 0x03F6, 0x03FD, 0x000B, 0x0C04, 0xFE00, 0x01D0,
+    0x000F, 0x007F, 0x03D9, 0x0006, 0x0401, 0x07FC, 0x03C7, 0x0054,
+    0x0001, 0x0BFE, 0x17C0, 0x002E, 0x37EA, 0xFE00, 0x0172, 0x0011,
+    0x0061, 0x03F0, 0x0007, 0x03FF, 0x0001, 0x07FB, 0x03C0, 0x0054,
+    0x0001, 0x23C2, 0x002C, 0x37EB, 0xFE00, 0x03DC, 0x0001, 0x000B,
+    0x03FD, 0x03FE, 0x0C01, 0x0001, 0xFE00, 0x03D2, 0x0001, 0x0012,
+    0x03FC, 0x03FE, 0x1001, 0x03FF, 0xFE00, 0x0676, 0x03F8, 0x0040,
+    0x0402, 0x1002, 0xFE00, 0x07EA, 0x0010, 0x0004, 0x07FA, 0x03FF,
+    0xFE00, 0x0374, 0x0006, 0x001A, 0x07FD, 0x0401, 0x0801, 0x0801,
+    0xFE00, 0x034A, 0x0008, 0x0022, 0x07FE, 0x0BFF, 0x0403, 0xFE00,
+    0x02F4, 0x000E, 0x001A, 0x07FE, 0x03FF, 0x0C02, 0xFE00, 0x02A2,
+    0x0010, 0x0022, 0x07FE, 0x0401, 0x0803, 0xFE00, 0x0526, 0x03F1,
+    0x03FD, 0x000E, 0xFE00, 0x07CE, 0x0021, 0x0401, 0x07FF, 0x0001,
+    0x07FF, 0xFE00, 0x028A, 0x0013, 0x000F, 0x0001, 0x03FF, 0x0401,
+    0x07FF, 0x0001, 0xFE00, 0x0216, 0x0017, 0x0015, 0x0002, 0x03FF,
+    0x0401, 0x0802, 0xFE00, 0x01E8, 0x0013, 0x0045, 0x03F8, 0x0001,
+    0x0401, 0x0805, 0xFE00, 0x0176, 0x0016, 0x0040, 0x03F9, 0x0801,
+    0x0805, 0x03FF, 0xFE00, 0x067C, 0x03F8, 0x03C0, 0x07FE, 0x13FE,
+    0xFE00, 0x07E2, 0x0011, 0x03FC, 0x0405, 0x03FF, 0xFE00, 0x0036,
+    0x000D, 0x006B, 0x03FF, 0x0002, 0x0401, 0x03FF, 0x0406, 0xFE00,
+    0x07E2, 0x0010, 0x0064, 0x0402, 0x03FF, 0x0001, 0x0805, 0xFE00,
+    0x0628, 0x0006, 0x006B, 0x0402, 0x0401, 0x03FF, 0x03FF, 0x0006,
+    0xFE00, 0x05FC, 0x0008, 0x0063, 0x0401, 0x03FF, 0x0C06, 0x0BFF,
+    0xFE00, 0x07E6, 0x07F2, 0x03F9, 0x03FF, 0x13FF, 0xFE00, 0x07FC,
+    0x0001, 0x03FF, 0x0401, 0xFE00, 0x0474, 0x0002, 0x0042, 0x0009,
+    0x0001, 0x1004, 0x0001, 0xFE00, 0x046A, 0x0001, 0x003D, 0x0009,
+    0x0001, 0x1003, 0x0001, 0xFE00, 0x0400, 0xFE00, 0x0400, 0xFE00,
+    0x07AC, 0x0002, 0x000B, 0x03FD, 0x1402, 0xFE00, 0x0072, 0x03FC,
+    0x001E, 0x03E4, 0x03FF, 0x0C01, 0x0005, 0x03FE, 0xFE00, 0x047C,
+    0x03FE, 0x03EE, 0x03ED, 0x07FF, 0x03FF, 0x0BF6, 0x03FA, 0x27FE,
+    0x03FF, 0xFE00, 0x0490, 0x03FE, 0x03EB, 0x03EB, 0x07FE, 0x03FE,
+    0x0BF4, 0x03F9, 0x0FFF, 0x03FF, 0x13FD, 0x03FF, 0xFE00, 0x04C6,
+    0x03FC, 0x03D3, 0x0013, 0x0001, 0x03FF, 0x03FF, 0x03FF, 0x03FF,
+    0x0004, 0x03FA, 0x0FFF, 0x1401, 0x0001, 0x37FF, 0xFE00, 0x04E2,
+    0x03FC, 0x03CE, 0x0015, 0x0001, 0x03FE, 0x03FE, 0x03FF, 0x03FF,
+    0x0005, 0x03F9, 0x0FFF, 0x03FF, 0x1001, 0x0001, 0x37FF, 0xFE00,
+    0x000E, 0x07B6, 0x000A, 0x0003, 0x1006, 0x03FE, 0x27FF, 0x0001,
+    0xFE00, 0x079A, 0x0002, 0x03E9, 0x0003, 0x1402, 0xFE00, 0x0524,
+    0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x054A, 0x03F9, 0x0FFF, 0x03FF,
+    0x1FFE, 0xFE00, 0x0462, 0x03FE, 0x001B, 0x001C, 0x03FF, 0x03FF,
+    0x03FF, 0x07FF, 0x0010, 0x0009, 0x2404, 0x0002, 0xFE00, 0x046E,
+    0x03FE, 0x001F, 0x0020, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0012,
+    0x000A, 0x2404, 0x0002, 0x3401, 0xFE00, 0x007A, 0x03FD, 0x0019,
+    0x000A, 0x03FF, 0xFE00, 0x0770, 0x0003, 0x0019, 0x000A, 0x03FF,
+    0xFE00, 0x04B6, 0x03FC, 0x03A5, 0x0026, 0x0002, 0x03FF, 0x03FF,
+    0x07FF, 0x0009, 0x03F4, 0x0FFF, 0x1402, 0x0003, 0x37FD, 0xFE00,
+    0x04CA, 0x03FC, 0x039B, 0x002B, 0x0002, 0x03FF, 0x03FF, 0x07FF,
+    0x000A, 0x03F2, 0x0FFF, 0x1402, 0x0003, 0x37FD, 0xFE00, 0x0572,
+    0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x059A, 0x03F9,
+    0x0FFF, 0x03FF, 0x1FFE, 0xFE00, 0x004C, 0x03FE, 0x0026, 0x03E5,
+    0x03FE, 0x0C01, 0x0007, 0x03FE, 0xFE00, 0x006C, 0x03FC, 0x07EF,
+    0x1001, 0x03FF, 0x03FE, 0x27FF, 0xFE00, 0x04EE, 0x03FC, 0x0012,
+    0x0013, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x000B, 0x0006, 0x0FFE,
+    0x03FF, 0x1002, 0x0001, 0xFE00, 0x0508, 0x03FC, 0x0014, 0x0015,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x000C, 0x0006, 0x0FFF, 0x03FF,
+    0x1003, 0x0001, 0xFE00, 0x052E, 0x03FA, 0x03BD, 0x001C, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0006, 0x03F7, 0x0FFE, 0x03FF, 0x1001,
+    0x0002, 0x37FE, 0xFE00, 0x0552, 0x03FB, 0x03B3, 0x0020, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0007, 0x03F6, 0x0FFE, 0x03FF, 0x1002,
+    0x0002, 0x37FE, 0xFE00, 0x077E, 0x0004, 0x03D4, 0x0006, 0x0002,
+    0x1004, 0x03FF, 0xFE00, 0x0012, 0x0406, 0x03FF, 0xFE00, 0x05B8,
+    0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x05F0, 0x03F7,
+    0x0FFE, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x05F8, 0x03F4, 0x03F7,
+    0x03F6, 0x07FE, 0x03FD, 0x0BFB, 0x03FD, 0x0FFC, 0x03FF, 0x13FF,
+    0xFE00, 0x0638, 0x03F5, 0x03F6, 0x03F5, 0x07FF, 0x03FE, 0x0BFA,
+    0x03FD, 0x0FFD, 0x03FF, 0x13FF, 0xFE00, 0x073A, 0x0005, 0x003A,
+    0x0011, 0x03FD, 0x0FFF, 0x03FD, 0x03FD, 0x27FF, 0x0001, 0x3402,
+    0xFE00, 0x0098, 0x03FB, 0x03D3, 0x03F3, 0x0002, 0x1002, 0x0002,
+    0x2BFF, 0x37FF, 0xFE00, 0x05D6, 0x03F5, 0x0010, 0x002B, 0x03FF,
+    0x03FF, 0x03FE, 0x03FF, 0x03FF, 0x03CF, 0x0015, 0x0001, 0x0BFD,
+    0x03FF, 0x1004, 0x03F2, 0x340A, 0xFE00, 0x0612, 0x03F5, 0x0012,
+    0x0030, 0x07FF, 0x03FE, 0x03FF, 0x03FF, 0x03C8, 0x0018, 0x0001,
+    0x0BFD, 0x03FF, 0x1005, 0x03F0, 0x340B, 0xFE00, 0x0656, 0x03F2,
+    0x0FFF, 0x03FE, 0x1FFC, 0x03FF, 0xFE00, 0x06A2, 0x03F0, 0x13FE,
+    0x1FFC, 0x03FF, 0xFE00, 0x003E, 0x03FE, 0x03C9, 0x03FD, 0x0003,
+    0x13EE, 0x2BFB, 0x3BFE, 0xFE00, 0x013A, 0x03F7, 0x000A, 0x03EC,
+    0x1403, 0x03FD, 0x2401, 0x03FE, 0xFE00, 0x05B0, 0x03F6, 0x0032,
+    0x0028, 0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x000C, 0x13FD, 0x03FF,
+    0x13FD, 0x03FD, 0x37FF, 0xFE00, 0x05E4, 0x03F5, 0x003A, 0x002E,
+    0x03FE, 0x07FF, 0x07FF, 0x000E, 0x13FD, 0x17FD, 0x03FD, 0x37FE,
+    0xFE00, 0x0564, 0x03F8, 0x0012, 0x03F0, 0x03FF, 0x03FE, 0x03FE,
+    0x0804, 0x13FE, 0x03FF, 0x13FF, 0x0001, 0xFE00, 0x0590, 0x03F8,
+    0x0013, 0x03F0, 0x03FF, 0x07FF, 0x0804, 0x13FE, 0x17FF, 0x0001,
+    0xFE00, 0x002E, 0x03F4, 0x00B6, 0x03BF, 0x17FF, 0x0007, 0x27FC,
+    0x03FF, 0x3403, 0xFE00, 0x07CE, 0x000A, 0x0052, 0x03E1, 0x03F2,
+    0x0C05, 0x0402, 0x27FE, 0xFE00, 0x0526, 0x03FA, 0x0FFE, 0x03FE,
+    0x1FFE, 0x03FF, 0xFE00, 0x054E, 0x03FA, 0x0FFE, 0x03FE, 0x1FFE,
+    0x03FF, 0xFE00, 0x07A8, 0x000D, 0x02A0, 0x0382, 0x03F4, 0x0BFF,
+    0x03FC, 0x003A, 0x0039, 0x07FF, 0x03FF, 0x03FF, 0x1001, 0x0006,
+    0x03ED, 0x37EF, 0xFE00, 0x0764, 0x000A, 0x02D0, 0x039A, 0x03F6,
+    0x0401, 0x03FF, 0x03FC, 0x0039, 0x0034, 0x07FF, 0x03FF, 0x0BFF,
+    0x0C05, 0x03EE, 0x37F0, 0xFE00, 0x06F6, 0x03FA, 0x0391, 0x03D7,
+    0x03FC, 0x0FFE, 0x0007, 0x0006, 0x2402, 0x03FD, 0x37FC, 0xFE00,
+    0x0768, 0x0010, 0x03C0, 0x03E9, 0x000C, 0x0C05, 0x0004, 0x0003,
+    0x2401, 0x03FE, 0x37FE, 0xFE00, 0x010A, 0x0012, 0x0041, 0x0009,
+    0x0007, 0x0001, 0x0BFB, 0x03B8, 0x0054, 0x0001, 0x23C4, 0x002A,
+    0x37EC, 0xFE00, 0x00AC, 0x000E, 0x0024, 0x0020, 0x0006, 0x0402,
+    0x07FC, 0x03B1, 0x0054, 0x0001, 0x23C5, 0x0028, 0x37ED, 0xFE00,
+    0x03C8, 0x0001, 0x0018, 0x03FC, 0x03FE, 0x1001, 0xFE00, 0x03BE,
+    0x0002, 0x001D, 0x03FA, 0x03FE, 0x1002, 0x0BFF, 0xFE00, 0x06B8,
+    0x03F9, 0x0036, 0x0402, 0x1002, 0xFE00, 0x074A, 0x0014, 0x001D,
+    0x07FA, 0xFE00, 0x031C, 0x0008, 0x002B, 0x07FE, 0x1003, 0xFE00,
+    0x02F0, 0x0008, 0x0032, 0x03FF, 0x03FF, 0x0401, 0x03FF, 0x0403,
+    0x0BFF, 0x0BFF, 0xFE00, 0x0248, 0x0010, 0x002A, 0x07FE, 0x0401,
+    0x0803, 0xFE00, 0x01F4, 0x000D, 0x0032, 0x07FE, 0x1004, 0xFE00,
+    0x05A0, 0x03F2, 0x03FE, 0x000B, 0xFE00, 0x06A6, 0x0024, 0x0407,
+    0x0401, 0x0001, 0x07FF, 0xFE00, 0x0198, 0x0016, 0x001A, 0x0002,
+    0x03FF, 0x0401, 0x07FF, 0x0002, 0xFE00, 0x0124, 0x0013, 0x0020,
+    0x0003, 0x03FF, 0x0001, 0x0001, 0x07FF, 0x0002, 0xFE00, 0x00F8,
+    0x0017, 0x0039, 0x03FA, 0x0801, 0x0804, 0xFE00, 0x0084, 0x0013,
+    0x0034, 0x03FA, 0x0801, 0x0804, 0x03FF, 0xFE00, 0x06BE, 0x03F8,
+    0x03CB, 0x07FE, 0x13FE, 0xFE00, 0x0742, 0x0013, 0x03E2, 0x0406,
+    0x13FF, 0xFE00, 0x0788, 0x0010, 0x005A, 0x0402, 0x0401, 0x0805,
+    0xFE00, 0x0736, 0x000E, 0x0052, 0x0402, 0x0401, 0x0805, 0xFE00,
+    0x05CC, 0x0008, 0x005B, 0x0401, 0x0BFF, 0x0405, 0xFE00, 0x05A2,
+    0x0006, 0x0053, 0x0402, 0x1005, 0x0BFF, 0xFE00, 0x07EA, 0x07F4,
+    0x03FA, 0x17FF, 0xFE00, 0x07F0, 0x0001, 0x03F8, 0x03FC, 0x0001,
+    0xFE00, 0x0460, 0x0002, 0x0036, 0x0008, 0x0001, 0x1002, 0xFE00,
+    0x0458, 0x0001, 0x0031, 0x0007, 0x0001, 0x1003, 0x0001, 0xFE00,
+    0x0400, 0xFE00, 0x0400, 0xFE00, 0x0798, 0x0002, 0x000E, 0x03FC,
+    0x1403, 0xFE00, 0x008E, 0x03FD, 0x0025, 0x03DD, 0x03FE, 0x0C01,
+    0x0006, 0x03FD, 0xFE00, 0x049E, 0x03FE, 0x03E9, 0x03E8, 0x07FF,
+    0x03FF, 0x0BF3, 0x03F9, 0x27FD, 0x03FE, 0xFE00, 0x04AC, 0x03FE,
+    0x03E7, 0x03E6, 0x07FF, 0x03FF, 0x0BF1, 0x03F8, 0x27FC, 0x03FE,
+    0xFE00, 0x04FA, 0x03FC, 0x03C8, 0x0018, 0x0002, 0x03FF, 0x03FF,
+    0x03FF, 0x03FF, 0x0005, 0x03F8, 0x0FFF, 0x1401, 0x0001, 0x37FE,
+    0xFE00, 0x0514, 0x03FC, 0x03C0, 0x001B, 0x0002, 0x03FF, 0x03FF,
+    0x03FF, 0x03FF, 0x0006, 0x03F7, 0x0FFF, 0x1401, 0x0002, 0x37FE,
+    0xFE00, 0x0012, 0x07A4, 0x000D, 0x0004, 0x1008, 0x03FE, 0x27FF,
+    0x0001, 0xFE00, 0x0782, 0x0003, 0x03E4, 0x0003, 0x0001, 0x1002,
+    0xFE00, 0x056E, 0x03F8, 0x33FE, 0xFE00, 0x0596, 0x03F8, 0x33FE,
+    0xFE00, 0x047A, 0x03FE, 0x0022, 0x0024, 0x03FF, 0x03FF, 0x03FF,
+    0x07FF, 0x0014, 0x000B, 0x2405, 0x0002, 0x3401, 0xFE00, 0x0486,
+    0x03FE, 0x0027, 0x0028, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0016,
+    0x000D, 0x2406, 0x0003, 0x3401, 0xFE00, 0x0098, 0x03FD, 0x0020,
+    0x000C, 0x03FF, 0x17FF, 0xFE00, 0x074C, 0x0004, 0x0020, 0x000C,
+    0x03FF, 0x17FF, 0xFE00, 0x04E2, 0x03FC, 0x038E, 0x0030, 0x0003,
+    0x03FF, 0x03FF, 0x07FF, 0x000B, 0x03F1, 0x0FFF, 0x1403, 0x0003,
+    0x37FC, 0xFE00, 0x04FA, 0x03FC, 0x0382, 0x0035, 0x0003, 0x07FF,
+    0x07FF, 0x000C, 0x03EF, 0x0FFF, 0x1403, 0x0004, 0x37FC, 0xFE00,
+    0x05CC, 0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x05FE,
+    0x03F6, 0x13FF, 0x1FFE, 0xFE00, 0x005C, 0x03FF, 0x0030, 0x03DE,
+    0x03FE, 0x0C01, 0x0009, 0x03FE, 0xFE00, 0x0086, 0x03FE, 0x07EB,
+    0x17FE, 0x03FE, 0x27FF, 0x03FF, 0xFE00, 0x052A, 0x03FC, 0x0017,
+    0x0018, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x000D, 0x0007, 0x0FFE,
+    0x03FF, 0x1003, 0x0001, 0xFE00, 0x0546, 0x03FC, 0x001A, 0x001B,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x000F, 0x0008, 0x0FFF, 0x03FF,
+    0x1004, 0x0002, 0xFE00, 0x057C, 0x03FA, 0x03AB, 0x0024, 0x0002,
+    0x03FE, 0x03FE, 0x03FF, 0x03FF, 0x0008, 0x03F5, 0x0FFE, 0x03FF,
+    0x1002, 0x0002, 0x37FD, 0xFE00, 0x05A0, 0x03FB, 0x03A2, 0x0028,
+    0x0002, 0x03FE, 0x03FE, 0x07FF, 0x0009, 0x03F3, 0x0FFE, 0x03FF,
+    0x1002, 0x0003, 0x37FD, 0xFE00, 0x075C, 0x0003, 0x03C8, 0x0007,
+    0x0002, 0x1005, 0x03FF, 0xFE00, 0x0016, 0x0408, 0x03FF, 0xFE00,
+    0x062C, 0x03F5, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x0662,
+    0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x0678, 0x03F5,
+    0x03F5, 0x03F4, 0x07FF, 0x03FE, 0x0BFA, 0x03FD, 0x0FFD, 0x03FF,
+    0x13FF, 0xFE00, 0x06BA, 0x03F6, 0x03F3, 0x03F3, 0x07FE, 0x03FD,
+    0x0BF9, 0x03FC, 0x0FFD, 0x03FF, 0x13FE, 0x03FF, 0xFE00, 0x070A,
+    0x0005, 0x0048, 0x0015, 0x03FD, 0x0FFF, 0x03FC, 0x03FC, 0x27FF,
+    0x0002, 0x3402, 0xFE00, 0x00BA, 0x03FC, 0x03C8, 0x03F0, 0x0001,
+    0x1003, 0x0002, 0x2BFF, 0x37FE, 0xFE00, 0x064E, 0x03F4, 0x0015,
+    0x0036, 0x07FF, 0x03FE, 0x03FF, 0x03FE, 0x03C2, 0x001A, 0x0001,
+    0x0BFD, 0x03FF, 0x1006, 0x03EE, 0x340D, 0xFE00, 0x068A, 0x03F6,
+    0x0018, 0x003B, 0x07FF, 0x03FE, 0x03FF, 0x03FF, 0x03BC, 0x001D,
+    0x0001, 0x0BFD, 0x03FF, 0x1006, 0x03EC, 0x340E, 0xFE00, 0x06EE,
+    0x03EF, 0x0C01, 0x03FF, 0x1FFC, 0xFE00, 0x073A, 0x03F1, 0x13FF,
+    0x1FFC, 0xFE00, 0x004C, 0x03FF, 0x03BC, 0x03FD, 0x0003, 0x13EA,
+    0x2BF9, 0x3BFE, 0xFE00, 0x0184, 0x03F7, 0x000C, 0x03E7, 0x1001,
+    0x0003, 0x03FC, 0x2401, 0x03FE, 0xFE00, 0x061C, 0x03F4, 0x0040,
+    0x0033, 0x03FE, 0x07FF, 0x07FF, 0x000F, 0x13FD, 0x17FC, 0x03FC,
+    0x37FE, 0xFE00, 0x0650, 0x03F7, 0x0048, 0x0039, 0x03FE, 0x07FF,
+    0x07FF, 0x0011, 0x13FE, 0x17FC, 0x03FC, 0x37FE, 0xFE00, 0x05BE,
+    0x03F8, 0x0015, 0x03EF, 0x03FF, 0x07FF, 0x0805, 0x13FE, 0x17FF,
+    0x0001, 0xFE00, 0x05E8, 0x03F9, 0x0018, 0x03EC, 0x03FF, 0x03FF,
+    0x03FF, 0x0806, 0x13FE, 0x17FF, 0x0001, 0xFE00, 0x008E, 0x03F2,
+    0x00B5, 0x03C1, 0x0001, 0x0001, 0x07FF, 0x03FF, 0x03FE, 0x0007,
+    0x27FC, 0x03FE, 0x3403, 0xFE00, 0x07A0, 0x0477, 0x03D3, 0x03FD,
+    0x0C01, 0x0001, 0x0004, 0x27FD, 0x3801, 0xFE00, 0x0576, 0x03FA,
+    0x0FFE, 0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x0598, 0x03FB, 0x0FFF,
+    0x03FE, 0x1FFF, 0x03FF, 0xFE00, 0x073A, 0x0004, 0x02F3, 0x03AC,
+    0x03FA, 0x0002, 0x07FD, 0x03FE, 0x0038, 0x0030, 0x07FF, 0x07FF,
+    0x03FF, 0x1003, 0x03EE, 0x37F0, 0xFE00, 0x074A, 0x03F8, 0x02F6,
+    0x03B0, 0x0002, 0x0BFF, 0x043B, 0x0031, 0x2403, 0x03EE, 0x37F0,
+    0xFE00, 0x0724, 0x03F8, 0x03A5, 0x03DE, 0x03F9, 0x0002, 0x0401,
+    0x03FE, 0x0006, 0x0005, 0x2401, 0x03FD, 0x37FD, 0xFE00, 0x0716,
+    0x0002, 0x039F, 0x03DC, 0x0002, 0x1006, 0x0005, 0x2402, 0x03FD,
+    0x37FD, 0xFE00, 0x006C, 0x0006, 0x0010, 0x0030, 0x0002, 0x0003,
+    0x0401, 0x03FE, 0x03AD, 0x0054, 0x07FF, 0x1FC7, 0x0027, 0x37ED,
+    0xFE00, 0x0076, 0x03F8, 0x0013, 0x002D, 0x03FD, 0x0C01, 0x03AD,
+    0x0054, 0x27C7, 0x0027, 0x37ED, 0xFE00, 0x03B8, 0x0420, 0x03F9,
+    0x13FF, 0x0002, 0xFE00, 0x03B8, 0x0420, 0x03FA, 0x1402, 0xFE00,
+    0x06F6, 0x03F4, 0x002C, 0x0403, 0x0001, 0x0C01, 0xFE00, 0x06E4,
+    0x0004, 0x002F, 0x07FF, 0x0001, 0x0C02, 0xFE00, 0x02D2, 0x0002,
+    0x0039, 0x07FE, 0x0BFF, 0x03FF, 0x0004, 0xFE00, 0x02D6, 0x03FC,
+    0x0038, 0x0FFF, 0x0805, 0xFE00, 0x01BC, 0x0006, 0x0039, 0x0802,
+    0x0C05, 0xFE00, 0x01C2, 0x03FB, 0x0038, 0x0FFF, 0x0805, 0xFE00,
+    0x060C, 0x03EC, 0x0409, 0x03FF, 0x0004, 0xFE00, 0x05EC, 0x0007,
+    0x040A, 0x0401, 0xFE00, 0x00D4, 0x0008, 0x0022, 0x0003, 0x0402,
+    0x07FF, 0x0403, 0xFE00, 0x00DE, 0x03F8, 0x0022, 0x0003, 0x0BFF,
+    0x0803, 0x0001, 0xFE00, 0x0036, 0x0007, 0x0030, 0x03FB, 0x0403,
+    0x0C03, 0xFE00, 0x0042, 0x03F8, 0x0031, 0x03FB, 0x1403, 0xFE00,
+    0x06F2, 0x03F8, 0x03D4, 0x07FD, 0x0002, 0x0FFE, 0xFE00, 0x06E2,
+    0x0002, 0x03D2, 0x0401, 0x13FE, 0xFE00, 0x0700, 0x0004, 0x004D,
+    0x0801, 0x0C04, 0xFE00, 0x0708, 0x03FA, 0x004D, 0x07FF, 0x07FF,
+    0x0804, 0xFE00, 0x0588, 0x044D, 0x1805, 0xFE00, 0x058C, 0x03FD,
+    0x004D, 0x1805, 0xFE00, 0x07EC, 0x07F5, 0x03FB, 0x17FF, 0xFE00,
+    0x07EC, 0x07F5, 0x03FB, 0x17FF, 0xFE00, 0x0456, 0x042F, 0x0005,
+    0x1403, 0x0001, 0xFE00, 0x0456, 0x042F, 0x0005, 0x1403, 0x0001,
+    0xFE00, 0x0400, 0xFE00, 0x0400, 0xFE00, 0x0780, 0x0002, 0x0010,
+    0x03FC, 0x03FF, 0x1003, 0xFE00, 0x00AC, 0x03FD, 0x002D, 0x03D6,
+    0x03FE, 0x0C01, 0x0007, 0x03FC, 0x63FF, 0xFE00, 0x04C0, 0x03FE,
+    0x03E4, 0x03E3, 0x07FE, 0x03FE, 0x0BF0, 0x03F7, 0x0FFF, 0x03FF,
+    0x13FC, 0x03FE, 0x37FF, 0xFE00, 0x04D0, 0x03FE, 0x03E2, 0x03E1,
+    0x07FF, 0x03FF, 0x0BEF, 0x03F6, 0x27FC, 0x03FE, 0x37FF, 0xFE00,
+    0x0530, 0x03FC, 0x03BC, 0x001C, 0x0001, 0x03FF, 0x03FF, 0x03FF,
+    0x0406, 0x03F7, 0x0FFF, 0x1401, 0x0002, 0x37FE, 0xFE00, 0x0548,
+    0x03FC, 0x03B6, 0x0020, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0007,
+    0x03F6, 0x0FFF, 0x1401, 0x0002, 0x37FE, 0xFE00, 0x0014, 0x0791,
+    0x000F, 0x0004, 0x100A, 0x03FE, 0x27FF, 0x0001, 0xFE00, 0x0766,
+    0x0003, 0x03DF, 0x0004, 0x0001, 0x1003, 0xFE00, 0x05BC, 0x03FA,
+    0x13FF, 0x1FFF, 0xFE00, 0x05E2, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE,
+    0xFE00, 0x0496, 0x03FE, 0x0029, 0x002B, 0x03FF, 0x03FF, 0x03FF,
+    0x07FF, 0x0018, 0x000D, 0x0FFF, 0x1406, 0x0003, 0x3401, 0xFE00,
+    0x04A4, 0x03FE, 0x002D, 0x002F, 0x03FF, 0x03FF, 0x03FF, 0x07FF,
+    0x001A, 0x000F, 0x2407, 0x0003, 0x3401, 0xFE00, 0x00B6, 0x03FD,
+    0x0026, 0x000F, 0x03FF, 0x1001, 0x03FF, 0x27FF, 0xFE00, 0x072A,
+    0x0004, 0x0025, 0x000F, 0x03FF, 0x17FF, 0x27FF, 0xFE00, 0x0514,
+    0x03FC, 0x0378, 0x0039, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x000D,
+    0x03EE, 0x0FFF, 0x1403, 0x0004, 0x37FC, 0xFE00, 0x052C, 0x03FC,
+    0x036D, 0x003E, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x000E, 0x03EC,
+    0x0FFF, 0x1403, 0x0005, 0x37FB, 0xFE00, 0x062A, 0x03F9, 0x0FFF,
+    0x03FF, 0x1FFE, 0xFE00, 0x065A, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE,
+    0x03FF, 0xFE00, 0x0070, 0x03FE, 0x0039, 0x03D7, 0x03FE, 0x0C01,
+    0x000B, 0x03FD, 0xFE00, 0x00A4, 0x03FD, 0x07E7, 0x17FE, 0x03FD,
+    0x27FF, 0x03FF, 0xFE00, 0x0566, 0x03FC, 0x001B, 0x001C, 0x07FE,
+    0x03FE, 0x0810, 0x0009, 0x0FFE, 0x03FF, 0x1004, 0x0002, 0xFE00,
+    0x0582, 0x03FC, 0x001E, 0x0020, 0x03FF, 0x03FE, 0x03FE, 0x07FF,
+    0x0012, 0x000A, 0x0FFF, 0x03FF, 0x1004, 0x0002, 0x3401, 0xFE00,
+    0x05CA, 0x03FA, 0x039A, 0x002B, 0x0002, 0x03FE, 0x03FE, 0x03FF,
+    0x03FF, 0x000A, 0x03F2, 0x0FFE, 0x03FF, 0x1002, 0x0003, 0x37FD,
+    0xFE00, 0x05EC, 0x03FB, 0x0391, 0x002F, 0x0002, 0x03FE, 0x03FE,
+    0x07FF, 0x000B, 0x03F1, 0x0FFE, 0x03FF, 0x1002, 0x0003, 0x37FC,
+    0xFE00, 0x073C, 0x0003, 0x03BD, 0x0009, 0x0002, 0x1006, 0x03FF,
+    0xFE00, 0x001E, 0x040A, 0x03FF, 0x17FF, 0xFE00, 0x069C, 0x03F5,
+    0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x06D2, 0x03F8, 0x0FFF,
+    0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x06FC, 0x03F4, 0x03F2, 0x03F1,
+    0x07FE, 0x03FD, 0x0BF8, 0x03FC, 0x0FFC, 0x03FF, 0x13FE, 0x03FF,
+    0xFE00, 0x073C, 0x03F5, 0x03F0, 0x03F0, 0x07FF, 0x03FE, 0x0BF7,
+    0x03FB, 0x0FFD, 0x03FF, 0x13FE, 0x03FF, 0xFE00, 0x06DC, 0x0005,
+    0x0056, 0x0019, 0x03FD, 0x0FFF, 0x03FC, 0x03FC, 0x27FF, 0x0002,
+    0x3403, 0xFE00, 0x00DE, 0x03FC, 0x03BE, 0x03ED, 0x0002, 0x1003,
+    0x0003, 0x2401, 0x03FE, 0x37FE, 0xFE00, 0x06C8, 0x03F5, 0x001A,
+    0x0040, 0x03FF, 0x03FF, 0x03FE, 0x03FF, 0x03FF, 0x03B5, 0x0020,
+    0x0001, 0x0BFD, 0x03FF, 0x1007, 0x03EB, 0x3410, 0xFE00, 0x0702,
+    0x03F6, 0x001D, 0x0045, 0x07FF, 0x03FE, 0x07FF, 0x03AF, 0x0023,
+    0x0001, 0x0BFD, 0x03FF, 0x1007, 0x03E9, 0x3411, 0xFE00, 0x0786,
+    0x03F2, 0x0FFF, 0x03FE, 0x1FFC, 0x03FF, 0xFE00, 0x07CE, 0x03F2,
+    0x0FFF, 0x03FE, 0x1FFC, 0x03FF, 0xFE00, 0x005E, 0x03FE, 0x03AF,
+    0x03FB, 0x0002, 0x13E6, 0x2BF8, 0x3BFE, 0xFE00, 0x01CE, 0x03F7,
+    0x000E, 0x03E3, 0x1404, 0x03FB, 0x2401, 0x03FE, 0xFE00, 0x0688,
+    0x03F5, 0x004E, 0x003E, 0x03FF, 0x03FF, 0x03FE, 0x07FF, 0x0013,
+    0x13FD, 0x03FF, 0x13FB, 0x03FB, 0x37FE, 0xFE00, 0x06BE, 0x03F7,
+    0x0054, 0x0043, 0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x0014, 0x13FE,
+    0x17FB, 0x03FB, 0x37FD, 0xFE00, 0x0616, 0x03F6, 0x001A, 0x03EA,
+    0x03FE, 0x03FF, 0x03FE, 0x0806, 0x13FD, 0x03FF, 0x13FF, 0x0001,
+    0x37FF, 0xFE00, 0x0644, 0x03F9, 0x001C, 0x03E9, 0x07FF, 0x03FE,
+    0x0806, 0x13FE, 0x03FF, 0x13FE, 0x0001, 0x37FF, 0xFE00, 0x0124,
+    0x03EA, 0x0096, 0x03CF, 0x000A, 0x0FFB, 0x03FB, 0x0007, 0x27FD,
+    0x03FE, 0x3404, 0xFE00, 0x07A2, 0x0486, 0x03CE, 0x03FD, 0x0C01,
+    0x0001, 0x0004, 0x27FD, 0x3801, 0xFE00, 0x05BE, 0x03FA, 0x0FFE,
+    0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x05E6, 0x03FA, 0x0FFE, 0x03FE,
+    0x1FFE, 0x03FF, 0xFE00, 0x0776, 0x03F8, 0x02EF, 0x03B0, 0x0002,
+    0x03FF, 0x03FF, 0x083E, 0x0033, 0x03FF, 0x13FF, 0x0C03, 0x03ED,
+    0x37EF, 0xFE00, 0x07A6, 0x03F7, 0x02E6, 0x03B0, 0x0002, 0x03FF,
+    0x03FE, 0x0843, 0x0036, 0x03FF, 0x2003, 0x03EB, 0x37EE, 0xFE00,
+    0x079C, 0x03ED, 0x03D5, 0x03F1, 0x03F1, 0x0FFA, 0x0003, 0x0002,
+    0x2001, 0x07FF, 0x37FF, 0xFE00, 0x06FE, 0x0002, 0x0396, 0x03D9,
+    0x0002, 0x1007, 0x0006, 0x2402, 0x03FD, 0x37FC, 0xFE00, 0x009E,
+    0x03F8, 0x0020, 0x0024, 0x03FC, 0x03FF, 0x0801, 0x03B0, 0x0054,
+    0x27C6, 0x0028, 0x37ED, 0xFE00, 0x00CE, 0x03F8, 0x002E, 0x0018,
+    0x03FC, 0x03FF, 0x03FE, 0x0401, 0x03B4, 0x0054, 0x27C5, 0x0029,
+    0x37EC, 0xFE00, 0x03BE, 0x03FF, 0x001D, 0x03FB, 0x1402, 0x03FF,
+    0xFE00, 0x03C2, 0x03FF, 0x001B, 0x03FB, 0x1402, 0x03FF, 0xFE00,
+    0x078C, 0x03E8, 0x0013, 0x0406, 0x07FF, 0xFE00, 0x06C6, 0x0003,
+    0x0034, 0x07FF, 0x1002, 0xFE00, 0x02E8, 0x03FC, 0x0035, 0x1805,
+    0xFE00, 0x02FE, 0x03FC, 0x0031, 0x1804, 0xFE00, 0x01E6, 0x03F8,
+    0x0035, 0x1805, 0xFE00, 0x0210, 0x03F8, 0x0031, 0x0401, 0x1004,
+    0xFE00, 0x071C, 0x03D4, 0x0001, 0x0005, 0x0BFF, 0x0401, 0xFE00,
+    0x05B6, 0x0007, 0x03FE, 0x000B, 0xFE00, 0x0110, 0x03F6, 0x001F,
+    0x0003, 0x1403, 0xFE00, 0x014A, 0x03F4, 0x001D, 0x0003, 0x0BFF,
+    0x0803, 0xFE00, 0x0072, 0x03F6, 0x0032, 0x03FB, 0x03FF, 0x07FF,
+    0x0804, 0xFE00, 0x00AC, 0x03F4, 0x0035, 0x03FA, 0x0BFF, 0x0804,
+    0x03FF, 0xFE00, 0x0780, 0x03E8, 0x03EC, 0x07F9, 0x07FF, 0x0BFF,
+    0xFE00, 0x06CA, 0x0003, 0x03CD, 0x0401, 0x13FE, 0xFE00, 0x072A,
+    0x03F8, 0x0051, 0x07FF, 0x07FF, 0x0804, 0xFE00, 0x0754, 0x03F8,
+    0x0055, 0x07FF, 0x07FF, 0x0805, 0xFE00, 0x059E, 0x03FC, 0x0052,
+    0x07FF, 0x07FF, 0x0804, 0xFE00, 0x05B4, 0x03FC, 0x0055, 0x07FF,
+    0x1005, 0xFE00, 0x07F4, 0x03FF, 0x03FA, 0x03FE, 0x03FF, 0xFE00,
+    0x07EA, 0x07F5, 0x03FB, 0x17FF, 0xFE00, 0x0458, 0x0431, 0x0007,
+    0x1403, 0xFE00, 0x045E, 0x03FF, 0x0034, 0x0006, 0x03FF, 0x1003,
+    0x0001, 0xFE00, 0x0400, 0xFE00, 0x0400, 0xFE00, 0x076C, 0x0001,
+    0x0013, 0x03FB, 0x1404, 0xFE00, 0x00C6, 0x03FE, 0x0034, 0x03D0,
+    0x03FF, 0x0C01, 0x0008, 0x03FC, 0x63FF, 0xFE00, 0x04DE, 0x03FE,
+    0x03E0, 0x03DF, 0x07FF, 0x03FF, 0x0BED, 0x03F6, 0x27FB, 0x03FE,
+    0x37FF, 0xFE00, 0x04F2, 0x03FE, 0x03DD, 0x03DC, 0x07FE, 0x03FE,
+    0x0BEC, 0x03F5, 0x0FFF, 0x03FF, 0x13FB, 0x03FE, 0x37FF, 0xFE00,
+    0x0560, 0x03FC, 0x03B0, 0x0021, 0x0001, 0x03FF, 0x03FF, 0x03FF,
+    0x03FF, 0x0007, 0x03F5, 0x0FFF, 0x1402, 0x0002, 0x37FE, 0xFE00,
+    0x057C, 0x03FC, 0x03AB, 0x0024, 0x0001, 0x03FE, 0x03FE, 0x03FF,
+    0x03FF, 0x0008, 0x03F5, 0x0FFF, 0x03FF, 0x1002, 0x0002, 0x37FD,
+    0xFE00, 0x001A, 0x0780, 0x0012, 0x0003, 0x100C, 0x03FD, 0x27FF,
+    0x0001, 0xFE00, 0x0750, 0x0002, 0x03D9, 0x0005, 0x0001, 0x1003,
+    0x03FF, 0xFE00, 0x0604, 0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x062A,
+    0x03F9, 0x0FFF, 0x03FF, 0x1FFE, 0xFE00, 0x04AE, 0x03FE, 0x0030,
+    0x0032, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x001C, 0x0010, 0x2407,
+    0x0004, 0x3401, 0xFE00, 0x04BA, 0x03FE, 0x0034, 0x0036, 0x03FF,
+    0x03FF, 0x03FF, 0x07FF, 0x001E, 0x0011, 0x2408, 0x0004, 0x3401,
+    0xFE00, 0x00D6, 0x03FD, 0x002C, 0x0011, 0x03FF, 0x1001, 0x03FF,
+    0x27FF, 0xFE00, 0x0708, 0x0004, 0x002C, 0x0011, 0x03FF, 0x1001,
+    0x03FF, 0x27FF, 0xFE00, 0x0542, 0x03FC, 0x0361, 0x0044, 0x0002,
+    0x03FF, 0x03FF, 0x07FF, 0x000F, 0x03EA, 0x0FFF, 0x1404, 0x0005,
+    0x37FB, 0xFE00, 0x0558, 0x03FC, 0x0356, 0x0048, 0x0003, 0x07FF,
+    0x07FF, 0x0011, 0x03E9, 0x0FFF, 0x1404, 0x0005, 0x37FA, 0xFE00,
+    0x068A, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x06B6,
+    0x03F6, 0x13FF, 0x1FFE, 0xFE00, 0x0084, 0x03FD, 0x0043, 0x03D1,
+    0x03FE, 0x0C01, 0x000C, 0x03FC, 0xFE00, 0x00BC, 0x03FC, 0x07E3,
+    0x17FD, 0x03FC, 0x27FE, 0x03FF, 0x37FF, 0xFE00, 0x05A4, 0x03FC,
+    0x0020, 0x0021, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0013, 0x000A,
+    0x0FFE, 0x03FF, 0x1005, 0x0002, 0x3401, 0xFE00, 0x05C0, 0x03FC,
+    0x0022, 0x0024, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0014, 0x000B,
+    0x0FFF, 0x03FF, 0x1005, 0x0002, 0x3401, 0xFE00, 0x0614, 0x03FA,
+    0x038A, 0x0032, 0x0002, 0x03FE, 0x03FE, 0x07FF, 0x000B, 0x03F0,
+    0x0FFE, 0x03FF, 0x1003, 0x0004, 0x37FC, 0xFE00, 0x063A, 0x03FB,
+    0x0380, 0x0036, 0x0002, 0x03FE, 0x03FE, 0x07FF, 0x000C, 0x03EF,
+    0x0FFE, 0x03FF, 0x1003, 0x0004, 0x37FC, 0xFE00, 0x071E, 0x0004,
+    0x03B3, 0x000A, 0x0002, 0x1007, 0x03FE, 0x27FF, 0x0001, 0xFE00,
+    0x0022, 0x040C, 0x03FF, 0x17FF, 0xFE00, 0x0708, 0x03F7, 0x0FFE,
+    0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x0740, 0x03F7, 0x0FFE, 0x03FE,
+    0x1FFD, 0x03FF, 0xFE00, 0x077A, 0x03F5, 0x03F0, 0x03EF, 0x07FF,
+    0x03FE, 0x0BF6, 0x03FB, 0x0FFD, 0x03FF, 0x13FE, 0x03FF, 0xFE00,
+    0x07BA, 0x03F5, 0x03EE, 0x03EE, 0x07FF, 0x03FE, 0x0BF6, 0x03FA,
+    0x0FFD, 0x03FF, 0x13FE, 0x03FF, 0xFE00, 0x06AE, 0x0005, 0x0064,
+    0x001D, 0x03FD, 0x0FFF, 0x03FB, 0x03FB, 0x27FF, 0x0002, 0x3403,
+    0xFE00, 0x0102, 0x03FC, 0x03B3, 0x03EA, 0x0002, 0x1004, 0x0004,
+    0x2401, 0x03FE, 0x37FE, 0xFE00, 0x073E, 0x03F5, 0x001F, 0x004B,
+    0x07FF, 0x03FE, 0x03FF, 0x03FF, 0x03A9, 0x0025, 0x0001, 0x0BFD,
+    0x03FF, 0x1008, 0x03E7, 0x3412, 0xFE00, 0x0778, 0x03F5, 0x0022,
+    0x0050, 0x07FF, 0x03FE, 0x03FF, 0x03FF, 0x03A3, 0x0028, 0x0001,
+    0x0BFD, 0x03FF, 0x1008, 0x03E5, 0x3413, 0xFE00, 0x001A, 0x03F0,
+    0x13FE, 0x1FFC, 0x03FF, 0xFE00, 0x0062, 0x03F1, 0x13FF, 0x1FFC,
+    0xFE00, 0x0070, 0x03FF, 0x03A2, 0x03FA, 0x0003, 0x13E2, 0x03FF,
+    0x27F7, 0x3BFD, 0xFE00, 0x021A, 0x03F7, 0x0010, 0x03DE, 0x1001,
+    0x0005, 0x03FA, 0x2401, 0x03FD, 0xFE00, 0x06F4, 0x03F6, 0x005A,
+    0x0048, 0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x0016, 0x13FD, 0x03FF,
+    0x13FA, 0x03FB, 0x37FD, 0xFE00, 0x072A, 0x03F6, 0x0060, 0x004C,
+    0x03FE, 0x03FF, 0x03FE, 0x07FF, 0x0017, 0x13FD, 0x03FF, 0x13FA,
+    0x03FA, 0x37FD, 0xFE00, 0x0670, 0x03F8, 0x001F, 0x03E7, 0x07FF,
+    0x03FE, 0x0807, 0x13FE, 0x03FF, 0x13FE, 0x0001, 0x37FF, 0xFE00,
+    0x069C, 0x03F8, 0x0021, 0x03E5, 0x03FF, 0x03FF, 0x03FE, 0x0808,
+    0x13FE, 0x03FF, 0x13FE, 0x0002, 0x37FF, 0xFE00, 0x01CE, 0x03EC,
+    0x006C, 0x03E2, 0x0008, 0x07FF, 0x07FC, 0x03F9, 0x0006, 0x27FF,
+    0x03FD, 0x3404, 0xFE00, 0x07A2, 0x0495, 0x03C8, 0x03FD, 0x0C01,
+    0x0001, 0x0005, 0x27FD, 0x3801, 0xFE00, 0x060A, 0x03F9, 0x0FFF,
+    0x03FF, 0x1FFE, 0xFE00, 0x0630, 0x03FB, 0x0FFF, 0x03FE, 0x1FFF,
+    0x03FF, 0xFE00, 0x07D8, 0x03F7, 0x02DA, 0x03AE, 0x0002, 0x07FE,
+    0x03FF, 0x0447, 0x0038, 0x03FF, 0x2003, 0x03EA, 0x37ED, 0xFE00,
+    0x0004, 0x03FA, 0x02D2, 0x03AE, 0x0001, 0x03FF, 0x03FF, 0x084A,
+    0x003B, 0x2403, 0x03E9, 0x37EC, 0xFE00, 0x002E, 0x03F0, 0x0012,
+    0x0007, 0x03F3, 0x0FFB, 0x03FF, 0x03FF, 0xFE00, 0x06E6, 0x0002,
+    0x038B, 0x03D5, 0x0002, 0x1008, 0x0006, 0x2402, 0x03FC, 0x37FC,
+    0xFE00, 0x00FE, 0x03F7, 0x003D, 0x000C, 0x03FC, 0x07FE, 0x0402,
+    0x03B8, 0x0054, 0x27C4, 0x002A, 0x37EC, 0xFE00, 0x0126, 0x03FA,
+    0x004A, 0x0002, 0x03FD, 0x07FF, 0x0401, 0x03BB, 0x0054, 0x27C3,
+    0x002A, 0x37EB, 0xFE00, 0x03C6, 0x03FF, 0x0018, 0x03FC, 0x1402,
+    0xFE00, 0x03CA, 0x0415, 0x03FC, 0x1402, 0x03FF, 0xFE00, 0x003A,
+    0x03EC, 0x03F7, 0x0406, 0xFE00, 0x06A8, 0x0003, 0x0039, 0x07FF,
+    0x1002, 0xFE00, 0x0314, 0x03FC, 0x002D, 0x0FFF, 0x0804, 0xFE00,
+    0x0326, 0x03FD, 0x0029, 0x1804, 0xFE00, 0x023A, 0x03F8, 0x002D,
+    0x0FFF, 0x0804, 0xFE00, 0x025E, 0x03FA, 0x002A, 0x1804, 0xFE00,
+    0x0066, 0x03DB, 0x0001, 0x03FE, 0x07FF, 0x03FF, 0x0401, 0xFE00,
+    0x057E, 0x0006, 0x03FE, 0x000C, 0xFE00, 0x0184, 0x03F6, 0x001A,
+    0x0003, 0x0BFE, 0x0802, 0xFE00, 0x01B8, 0x03F8, 0x0018, 0x0002,
+    0x0BFF, 0x0802, 0x0001, 0xFE00, 0x00E8, 0x03F6, 0x0038, 0x03FB,
+    0x0BFF, 0x0804, 0x03FF, 0xFE00, 0x011A, 0x03F8, 0x003B, 0x03F9,
+    0x0BFF, 0x0804, 0x03FF, 0xFE00, 0x002E, 0x03ED, 0x0009, 0x07FA,
+    0xFE00, 0x06AA, 0x0003, 0x03C8, 0x1BFE, 0xFE00, 0x077E, 0x03F8,
+    0x0059, 0x0FFF, 0x0805, 0xFE00, 0x07A2, 0x03FB, 0x005D, 0x0FFF,
+    0x0806, 0xFE00, 0x05CA, 0x03FC, 0x0059, 0x07FF, 0x07FF, 0x0805,
+    0xFE00, 0x05DC, 0x03FD, 0x005D, 0x1806, 0xFE00, 0x0000, 0x03FF,
+    0x0BFF, 0xFE00, 0x07E8, 0x07F4, 0x03FA, 0x17FF, 0xFE00, 0x0462,
+    0x03FF, 0x0036, 0x0007, 0x1403, 0x0001, 0xFE00, 0x0466, 0x03FF,
+    0x0039, 0x0008, 0x03FE, 0x0FFF, 0x0003, 0xFE00, 0x0400, 0xFE00,
+    0x0400, 0xFE00, 0x0756, 0x0003, 0x0016, 0x03FA, 0x1405, 0xFE00,
+    0x00E6, 0x03FC, 0x003C, 0x03C8, 0x03FE, 0x0C01, 0x000A, 0x03FB,
+    0x63FF, 0xFE00, 0x0500, 0x03FE, 0x03DB, 0x03D9, 0x07FF, 0x03FF,
+    0x0BEA, 0x03F4, 0x27FB, 0x03FD, 0x37FF, 0xFE00, 0x0514, 0x03FE,
+    0x03D8, 0x03D7, 0x07FE, 0x03FE, 0x0BE9, 0x03F3, 0x0FFF, 0x03FF,
+    0x13FA, 0x03FD, 0x37FF, 0xFE00, 0x0594, 0x03FC, 0x03A6, 0x0026,
+    0x0002, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x0009, 0x03F4, 0x0FFF,
+    0x1402, 0x0003, 0x37FD, 0xFE00, 0x05B0, 0x03FC, 0x03A1, 0x0028,
+    0x0001, 0x03FF, 0x03FF, 0x03FF, 0x0409, 0x03F3, 0x0FFF, 0x1402,
+    0x0003, 0x37FD, 0xFE00, 0x001E, 0x076C, 0x0014, 0x0004, 0x100E,
+    0x03FD, 0x27FE, 0x0002, 0xFE00, 0x0734, 0x0003, 0x03D4, 0x0006,
+    0x0001, 0x1004, 0x03FF, 0xFE00, 0x064E, 0x03F8, 0x33FE, 0xFE00,
+    0x0674, 0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x04C6, 0x03FE, 0x0038,
+    0x003A, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0020, 0x0012, 0x2408,
+    0x0004, 0x3401, 0xFE00, 0x04D8, 0x03FE, 0x003A, 0x003C, 0x03FF,
+    0x03FF, 0x03FF, 0x07FF, 0x0022, 0x0013, 0x0FFF, 0x1409, 0x0004,
+    0x3402, 0xFE00, 0x00F4, 0x03FC, 0x0032, 0x0014, 0x03FF, 0x1001,
+    0x03FF, 0x27FF, 0xFE00, 0x06E4, 0x0004, 0x0032, 0x0014, 0x03FF,
+    0x1001, 0x03FF, 0x27FF, 0xFE00, 0x056E, 0x03FC, 0x034B, 0x004D,
+    0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0012, 0x03E7, 0x0FFF, 0x1404,
+    0x0006, 0x37FA, 0xFE00, 0x0588, 0x03FC, 0x0341, 0x0051, 0x0002,
+    0x03FF, 0x03FF, 0x07FF, 0x0013, 0x03E6, 0x0FFF, 0x03FF, 0x1005,
+    0x0006, 0x37FA, 0xFE00, 0x06E2, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE,
+    0xFE00, 0x0712, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF, 0xFE00,
+    0x0098, 0x03FD, 0x004C, 0x03CA, 0x03FE, 0x0C01, 0x000E, 0x03FC,
+    0x63FF, 0xFE00, 0x00DA, 0x03FD, 0x07DE, 0x1001, 0x03FD, 0x03FC,
+    0x27FE, 0x03FF, 0x37FF, 0xFE00, 0x05E0, 0x03FC, 0x0025, 0x0026,
+    0x07FE, 0x03FE, 0x0815, 0x000C, 0x0FFE, 0x03FF, 0x1005, 0x0002,
+    0x3401, 0xFE00, 0x0602, 0x03FC, 0x0026, 0x0028, 0x03FF, 0x03FE,
+    0x03FE, 0x07FF, 0x0016, 0x000C, 0x0FFE, 0x03FF, 0x1006, 0x0003,
+    0x3401, 0xFE00, 0x0662, 0x03FA, 0x0378, 0x0039, 0x0002, 0x03FE,
+    0x03FE, 0x07FF, 0x000D, 0x03EE, 0x0FFE, 0x03FF, 0x1003, 0x0004,
+    0x37FC, 0xFE00, 0x068A, 0x03FA, 0x0371, 0x003C, 0x0002, 0x03FE,
+    0x03FE, 0x07FF, 0x000E, 0x03ED, 0x0FFE, 0x03FF, 0x1003, 0x0004,
+    0x37FB, 0xFE00, 0x0700, 0x0004, 0x03A8, 0x000C, 0x0002, 0x1008,
+    0x03FE, 0x27FF, 0x0001, 0xFE00, 0x0028, 0x040E, 0x03FE, 0x17FF,
+    0xFE00, 0x0778, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF, 0xFE00,
+    0x07B0, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x07FC,
+    0x03F6, 0x03ED, 0x03EC, 0x07FD, 0x03FC, 0x0BF5, 0x03FA, 0x0FFD,
+    0x03FF, 0x13FD, 0x03FF, 0xFE00, 0x003C, 0x03F6, 0x03EC, 0x03EC,
+    0x07FD, 0x03FC, 0x0BF4, 0x03FA, 0x0FFD, 0x03FF, 0x13FD, 0x03FF,
+    0xFE00, 0x07D8, 0x03F9, 0x000B, 0x0003, 0x0004, 0x03F6, 0x03FC,
+    0x0006, 0x0001, 0x0C01, 0x0002, 0x03FF, 0x03FF, 0x0001, 0x2BFF,
+    0xFE00, 0x0008, 0x000C, 0x03FD, 0x03FF, 0x03F9, 0x0007, 0x0003,
+    0x03FC, 0x03FE, 0x13FF, 0x0001, 0x0001, 0xFE00, 0x06B2, 0x0018,
+    0x0116, 0x0396, 0x03D4, 0x0019, 0x0015, 0x03D9, 0x0019, 0x03D7,
+    0x002B, 0x03F8, 0x0016, 0x03E6, 0x000D, 0x0008, 0x03EE, 0x000E,
+    0x03F9, 0x07FA, 0x03F3, 0x0001, 0x07FB, 0x0007, 0x03F8, 0x0002,
+    0x07FD, 0x0003, 0x03FE, 0x0401, 0x03FE, 0x000C, 0x0001, 0x03FE,
+    0x0001, 0x07FF, 0x0001, 0x03FF, 0x13FF, 0x0001, 0x0001, 0x03FF,
+    0xFE00, 0x06CC, 0x03F5, 0x0137, 0x0383, 0x0001, 0x03FE, 0x03FD,
+    0x03FF, 0x03FE, 0x03D9, 0x002E, 0x0001, 0x0BFE, 0x03FF, 0x03FF,
+    0x0FF9, 0x03F3, 0x340C, 0xFE00, 0x04A2, 0x007A, 0x0C6C, 0x0060,
+    0x1C40, 0x0025, 0x2C0C, 0x0004, 0xFE00, 0x0468, 0x03EE, 0x0FFF,
+    0x03FD, 0x1FFF, 0x03FF, 0xFE00, 0x07FE, 0x0006, 0x03F6, 0x07FC,
+    0x0003, 0x0001, 0x03FA, 0x07FD, 0x07FF, 0x07FE, 0x0BFF, 0x07FF,
+    0x07FF, 0x0FFF, 0x1BFF, 0xFE00, 0x000C, 0x001D, 0x0001, 0x03FF,
+    0x040F, 0x0006, 0x07FC, 0x0FFF, 0x0402, 0x0002, 0x2C01, 0xFE00,
+    0x0478, 0x0054, 0x000C, 0x0009, 0x0010, 0x004E, 0x0045, 0x000D,
+    0x0009, 0x0003, 0x0403, 0x0008, 0x0009, 0x002E, 0x001A, 0x0006,
+    0x0005, 0x0002, 0x07FF, 0x0C01, 0x0002, 0x0002, 0x0009, 0x0002,
+    0x0001, 0x0001, 0x0FFF, 0xFE00, 0x0460, 0x03F0, 0x0001, 0x0001,
+    0x03FF, 0x03FF, 0x03FC, 0x23FF, 0xFE00, 0x0464, 0x0044, 0x0005,
+    0x03FB, 0x0004, 0x0041, 0x0039, 0x0004, 0x03FE, 0x0001, 0x0BFD,
+    0x0002, 0x0026, 0x0016, 0x0002, 0x03FF, 0x1FFF, 0x0407, 0x0002,
+    0xFE00, 0x0454, 0x03F2, 0x0008, 0x03F9, 0x03FF, 0x07FD, 0x0802,
+    0x13FF, 0x03FF, 0xFE00, 0x0090, 0x0005, 0x03D5, 0x0012, 0x000E,
+    0x000D, 0x0005, 0x0006, 0x03FB, 0x03FE, 0x03FF, 0x07FF, 0x0002,
+    0x0002, 0x0001, 0x0001, 0x0C01, 0x1401, 0x0001, 0xFE00, 0x072C,
+    0x0001, 0x006D, 0x03D4, 0x000B, 0x0003, 0x0001, 0x07FC, 0x0002,
+    0x0003, 0x27FD, 0xFE00, 0x0450, 0x003A, 0x0C36, 0x0030, 0x1C1E,
+    0x0012, 0x2C06, 0x0001, 0xFE00, 0x0438, 0x03F4, 0x13FF, 0x1FFD,
+    0xFE00, 0x0794, 0x0016, 0x025E, 0x0354, 0x000C, 0x0007, 0x000D,
+    0x0017, 0x0009, 0x0030, 0x003B, 0x0004, 0x000C, 0x000B, 0x0006,
+    0x0004, 0x0008, 0x0006, 0x0003, 0x040A, 0x03F0, 0x0802, 0x0003,
+    0x0004, 0x0001, 0x0401, 0x0001, 0x0001, 0x07FF, 0x03FF, 0x03F0,
+    0x03FF, 0x03FF, 0xFE00, 0x0724, 0x0017, 0x028C, 0x0365, 0x03EA,
+    0x03FF, 0x0002, 0x07F8, 0x0029, 0x0034, 0x0BFF, 0x03FF, 0x1002,
+    0x0009, 0x03F2, 0x37F2, 0xFE00, 0x0086, 0x03F7, 0x0039, 0x0014,
+    0x03F9, 0x0001, 0x0BFD, 0x03FC, 0x03FD, 0x27FF, 0x0001, 0x3402,
+    0xFE00, 0x06EE, 0x03F8, 0x038E, 0x03D6, 0x03F9, 0x0001, 0x0401,
+    0x03FD, 0x0008, 0x0006, 0x2402, 0x03FD, 0x37FC, 0xFE00, 0x0132,
+    0x000A, 0x004D, 0x0404, 0x03F8, 0x0001, 0x03FD, 0x03FE, 0x03BB,
+    0x0054, 0x0402, 0x07FF, 0x17C3, 0x002A, 0x37EB, 0xFE00, 0x00B2,
+    0x0020, 0x0025, 0x001F, 0x000D, 0x03FF, 0x0003, 0x03FF, 0x03F8,
+    0x03B2, 0x0054, 0x0002, 0x23C5, 0x0028, 0x37ED, 0xFE00, 0x03CC,
+    0x0414, 0x03FC, 0x1402, 0xFE00, 0x03C0, 0x0003, 0x001D, 0x03FA,
+    0x03FE, 0x0FFF, 0x0002, 0x03FF, 0xFE00, 0x00A4, 0x03F6, 0x03E4,
+    0x0403, 0x0001, 0x0FFF, 0xFE00, 0x06B2, 0x03F6, 0x0038, 0x0402,
+    0x0001, 0x0C02, 0xFE00, 0x0328, 0x0003, 0x0028, 0x07FE, 0x03FE,
+    0x0001, 0x0804, 0xFE00, 0x02EE, 0x000E, 0x0032, 0x07FC, 0x03FF,
+    0x0C04, 0xFE00, 0x0264, 0x0008, 0x0028, 0x03FF, 0x03FE, 0x03FB,
+    0x0401, 0x0404, 0x0BFF, 0xFE00, 0x01F6, 0x001B, 0x0033, 0x07FD,
+    0x03FF, 0x0002, 0x0805, 0xFE00, 0x0138, 0x03EC, 0x03FF, 0x03FB,
+    0x0401, 0xFE00, 0x0592, 0x03ED, 0x03FE, 0x000C, 0x0403, 0xFE00,
+    0x01C2, 0x000B, 0x0015, 0x0002, 0x03FF, 0x03F8, 0x0001, 0x0802,
+    0x0BFF, 0xFE00, 0x0128, 0x0026, 0x001E, 0x0003, 0x03FE, 0x03FF,
+    0x0003, 0x0802, 0xFE00, 0x0128, 0x000B, 0x003A, 0x03FB, 0x0001,
+    0x03F9, 0x0001, 0x03FF, 0x0404, 0x0BFF, 0xFE00, 0x008E, 0x0027,
+    0x0034, 0x03FB, 0x0001, 0x03FE, 0x0002, 0x0803, 0x03FF, 0xFE00,
+    0x00A4, 0x03F4, 0x001C, 0x07FC, 0x1001, 0xFE00, 0x06B6, 0x03F6,
+    0x03CA, 0x07FD, 0x0001, 0x0FFE, 0xFE00, 0x07AA, 0x0008, 0x005D,
+    0x03FF, 0x07FC, 0x0C06, 0xFE00, 0x073A, 0x001C, 0x0053, 0x0404,
+    0x0401, 0x0806, 0xFE00, 0x05E0, 0x0004, 0x005D, 0x0BFE, 0x07FF,
+    0x03FF, 0x0006, 0xFE00, 0x05A6, 0x000E, 0x0052, 0x0404, 0x0401,
+    0x0806, 0xFE00, 0x0006, 0x03FF, 0x0004, 0x0002, 0x03FF, 0xFE00,
+    0x07EA, 0x07F4, 0x03FA, 0x17FF, 0xFE00, 0x0466, 0x0001, 0x0039,
+    0x0008, 0x0001, 0x03FF, 0x07FF, 0x0404, 0x0001, 0xFE00, 0x0458,
+    0x0003, 0x0032, 0x0005, 0x0002, 0x1003, 0x0001, 0xFE00, 0x0400,
+    0xFE00, 0x0400, 0xFE00, 0x0740, 0x0002, 0x0018, 0x03FA, 0x1405,
+    0xFE00, 0x00FE, 0x03FD, 0x0043, 0x03C2, 0x03FF, 0x0C01, 0x000B,
+    0x03FB, 0x27FF, 0x3BFF, 0xFE00, 0x0524, 0x03FE, 0x03D6, 0x03D4,
+    0x07FF, 0x03FF, 0x0BE7, 0x03F2, 0x27FA, 0x03FD, 0x37FF, 0xFE00,
+    0x0532, 0x03FE, 0x03D4, 0x03D2, 0x07FF, 0x03FF, 0x0BE6, 0x03F2,
+    0x27F9, 0x03FD, 0x37FF, 0xFE00, 0x05C8, 0x03FC, 0x039C, 0x002A,
+    0x0002, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x000A, 0x03F2, 0x0FFF,
+    0x1402, 0x0003, 0x37FD, 0xFE00, 0x05E0, 0x03FC, 0x0396, 0x002D,
+    0x0001, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x000A, 0x03F2, 0x0FFF,
+    0x1402, 0x0003, 0x37FD, 0xFE00, 0x0022, 0x075C, 0x0017, 0x0003,
+    0x100F, 0x03FC, 0x27FE, 0x0002, 0xFE00, 0x071E, 0x0002, 0x03CF,
+    0x0006, 0x1404, 0x03FF, 0xFE00, 0x0696, 0x03F8, 0x33FE, 0xFE00,
+    0x06BC, 0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x04E4, 0x03FE, 0x003D,
+    0x0040, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0024, 0x0014, 0x2409,
+    0x0005, 0x3402, 0xFE00, 0x04F0, 0x03FE, 0x0042, 0x0044, 0x03FF,
+    0x03FF, 0x03FF, 0x07FF, 0x0026, 0x0015, 0x240A, 0x0005, 0x3402,
+    0xFE00, 0x0114, 0x03FD, 0x0038, 0x0016, 0x03FF, 0x1001, 0x03FF,
+    0x27FF, 0xFE00, 0x06C2, 0x0004, 0x0038, 0x0016, 0x03FF, 0x1001,
+    0x03FF, 0x27FF, 0xFE00, 0x059E, 0x03FC, 0x0336, 0x0056, 0x0003,
+    0x07FF, 0x07FF, 0x0014, 0x03E5, 0x0FFF, 0x1405, 0x0006, 0x37F9,
+    0xFE00, 0x05B6, 0x03FC, 0x032A, 0x005B, 0x0003, 0x07FF, 0x07FF,
+    0x0015, 0x03E3, 0x0FFF, 0x1405, 0x0007, 0x37F9, 0xFE00, 0x073E,
+    0x03F6, 0x13FF, 0x1FFE, 0xFE00, 0x076E, 0x03F6, 0x13FF, 0x1FFE,
+    0xFE00, 0x00AA, 0x03FF, 0x0056, 0x03C3, 0x03FE, 0x0C01, 0x0010,
+    0x03FB, 0x63FF, 0xFE00, 0x00F6, 0x03FE, 0x03FF, 0x03DB, 0x17FD,
+    0x03FB, 0x27FE, 0x03FF, 0x37FF, 0xFE00, 0x061E, 0x03FC, 0x0029,
+    0x002A, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0018, 0x000D, 0x0FFF,
+    0x03FF, 0x1006, 0x0003, 0x3401, 0xFE00, 0x0638, 0x03FC, 0x002C,
+    0x002D, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0019, 0x000E, 0x0FFF,
+    0x03FF, 0x1006, 0x0003, 0x3401, 0xFE00, 0x06B0, 0x03FB, 0x0368,
+    0x0040, 0x0002, 0x03FE, 0x03FE, 0x07FF, 0x000F, 0x03EB, 0x0FFE,
+    0x03FF, 0x1004, 0x0005, 0x37FB, 0xFE00, 0x06D4, 0x03FB, 0x035E,
+    0x0044, 0x0002, 0x03FE, 0x03FE, 0x07FF, 0x0010, 0x03EA, 0x0FFE,
+    0x03FF, 0x1004, 0x0005, 0x37FB, 0xFE00, 0x06E0, 0x0004, 0x039D,
+    0x000D, 0x0002, 0x1009, 0x03FE, 0x27FF, 0x0001, 0xFE00, 0x002E,
+    0x0410, 0x03FE, 0x17FF, 0xFE00, 0x07E8, 0x03F7, 0x0FFE, 0x03FE,
+    0x1FFD, 0x03FF, 0xFE00, 0x0020, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD,
+    0x03FF, 0xFE00, 0x007A, 0x03F6, 0x03EB, 0x03EA, 0x07FD, 0x03FC,
+    0x0BF4, 0x03F9, 0x0FFD, 0x03FF, 0x13FD, 0x03FF, 0xFE00, 0x00BA,
+    0x03F6, 0x03EA, 0x03E9, 0x07FD, 0x03FC, 0x0BF3, 0x03F9, 0x0FFD,
+    0x03FF, 0x13FD, 0x03FF, 0xFE00, 0x07B8, 0x0006, 0x0015, 0x0006,
+    0x03FC, 0x0FFF, 0x03FF, 0x03FF, 0xFE00, 0x07E6, 0x0002, 0x0007,
+    0x0002, 0x03FF, 0xFE00, 0x070A, 0x03F6, 0x0135, 0x038C, 0x07FE,
+    0x03FD, 0x07FE, 0x03D1, 0x0030, 0x0001, 0x0BFE, 0x03FF, 0x03FF,
+    0x0FFA, 0x03F1, 0x340E, 0xFE00, 0x0748, 0x03F6, 0x0134, 0x0394,
+    0x07FE, 0x03FD, 0x03FF, 0x03FE, 0x03CA, 0x0033, 0x0001, 0x0BFE,
+    0x03FF, 0x03FF, 0x0FFB, 0x03EF, 0x340F, 0xFE00, 0x04C2, 0x03EF,
+    0x0FFE, 0x03FE, 0x23FF, 0xFE00, 0x051E, 0x03F0, 0x0FFF, 0x03FC,
+    0x23FF, 0xFE00, 0x07E8, 0x0001, 0x03F0, 0x0001, 0x0003, 0x13FB,
+    0x2BFF, 0xFE00, 0x07A4, 0x0008, 0x0002, 0x0005, 0xFE00, 0x04B6,
+    0x03F0, 0x0003, 0x0002, 0x0BFE, 0x03FF, 0x1FFF, 0xFE00, 0x050A,
+    0x03F0, 0x0005, 0x0004, 0x07FF, 0x03FE, 0x0801, 0x17FF, 0xFE00,
+    0x04A0, 0x03F1, 0x000E, 0x03F4, 0x03FF, 0x07FE, 0x0803, 0x13FF,
+    0x03FF, 0x13FF, 0xFE00, 0x04EC, 0x03F1, 0x0015, 0x03EF, 0x03FF,
+    0x03FF, 0x03FE, 0x03FF, 0x0405, 0x17FF, 0x13FF, 0x0001, 0xFE00,
+    0x00F0, 0x03EF, 0x03B2, 0x0021, 0x0007, 0x0FFC, 0x03FD, 0x03FE,
+    0x2402, 0x03FF, 0xFE00, 0x074E, 0x03FA, 0x0039, 0x03E8, 0x000E,
+    0x0FFA, 0x0002, 0x0001, 0x27FF, 0xFE00, 0x0470, 0x03F4, 0x13FF,
+    0x1FFD, 0xFE00, 0x04A8, 0x03F4, 0x13FF, 0x1FFD, 0xFE00, 0x0690,
+    0x0020, 0x02E8, 0x0391, 0x03E4, 0x0402, 0x07F6, 0x0024, 0x0029,
+    0x0001, 0x07FF, 0x03FF, 0x1003, 0x0006, 0x03F4, 0x37F4, 0xFE00,
+    0x05EC, 0x001C, 0x034C, 0x03C1, 0x03E7, 0x0404, 0x07F7, 0x001F,
+    0x001E, 0x0BFE, 0x03FF, 0x0BFF, 0x0402, 0x0003, 0x03F6, 0x37F7,
+    0xFE00, 0x00EC, 0x03F1, 0x0061, 0x0024, 0x03F5, 0x0FFC, 0x03F9,
+    0x03FB, 0x27FE, 0x0003, 0x3403, 0xFE00, 0x0754, 0x03F1, 0x03B7,
+    0x03E6, 0x03F5, 0x0FFC, 0x0005, 0x0004, 0x2401, 0x03FE, 0x37FE,
+    0xFE00, 0x07F0, 0x002A, 0x03EB, 0x004F, 0x0012, 0x0402, 0x07F5,
+    0x03A3, 0x0054, 0x0003, 0x23C9, 0x0024, 0x3001, 0x03EF, 0xFE00,
+    0x071A, 0x0026, 0x03A8, 0x0084, 0x0010, 0x0405, 0x07F6, 0x0393,
+    0x0054, 0x0003, 0x0401, 0x0FFF, 0x0BCE, 0x0020, 0x37F1, 0xFE00,
+    0x03AA, 0x0004, 0x0028, 0x03FB, 0x03FD, 0x1003, 0x03FF, 0xFE00,
+    0x0394, 0x0003, 0x0034, 0x03FA, 0x03FD, 0x0401, 0x03FF, 0x0404,
+    0x03FE, 0xFE00, 0x011C, 0x03EE, 0x03D1, 0x0405, 0x13FF, 0xFE00,
+    0x072C, 0x03EE, 0x0024, 0x0405, 0x1001, 0xFE00, 0x0298, 0x0011,
+    0x0043, 0x07FB, 0x0401, 0x0806, 0xFE00, 0x0236, 0x0011, 0x0055,
+    0x07FB, 0x0401, 0x0807, 0xFE00, 0x014E, 0x0022, 0x0042, 0x03FF,
+    0x03FB, 0x0402, 0x0806, 0x1401, 0xFE00, 0x0092, 0x0023, 0x0055,
+    0x07FC, 0x0402, 0x0807, 0xFE00, 0x021A, 0x03DF, 0x03FF, 0x03F6,
+    0x0BFF, 0x0401, 0xFE00, 0x0670, 0x03DF, 0x03FE, 0x0007, 0x0BFF,
+    0x0401, 0xFE00, 0x003C, 0x0031, 0x0028, 0x0004, 0x03FD, 0x0403,
+    0x0804, 0xFE00, 0x0734, 0x0030, 0x0035, 0x0004, 0x03FD, 0x0403,
+    0x07FF, 0x0004, 0xFE00, 0x07A2, 0x0030, 0x0029, 0x03FC, 0x0003,
+    0x0404, 0x0803, 0x03FF, 0xFE00, 0x069A, 0x0030, 0x001D, 0x03FC,
+    0x0002, 0x0404, 0x07FF, 0x0002, 0xFE00, 0x0120, 0x03EE, 0x0030,
+    0x07FB, 0x1002, 0xFE00, 0x072E, 0x03EE, 0x03DD, 0x07FA, 0x13FF,
+    0xFE00, 0x0694, 0x0022, 0x0042, 0x03FF, 0x0004, 0x03FF, 0x0002,
+    0x0804, 0xFE00, 0x05D6, 0x0023, 0x002E, 0x0404, 0x0402, 0x0803,
+    0xFE00, 0x0550, 0x0012, 0x0041, 0x0404, 0x0401, 0x0804, 0xFE00,
+    0x04EE, 0x0011, 0x002F, 0x0404, 0x0401, 0x0803, 0xFE00, 0x0010,
+    0x03FF, 0x0008, 0x0004, 0x03FF, 0xFE00, 0x07F2, 0x03FF, 0x03F9,
+    0x03FD, 0x03FF, 0x13FF, 0xFE00, 0x0442, 0x0003, 0x0028, 0x0005,
+    0x0002, 0x0401, 0x0802, 0x0001, 0xFE00, 0x042E, 0x0004, 0x001B,
+    0x0004, 0x0002, 0x1002, 0xFE00, 0x0400, 0xFE00, 0x0400, 0xFE00,
+    0x072E, 0x0002, 0x001C, 0x03F9, 0x1406, 0xFE00, 0x011A, 0x03FC,
+    0x004A, 0x03BB, 0x03FF, 0x0C01, 0x000C, 0x03FA, 0x27FF, 0x3BFF,
+    0xFE00, 0x0546, 0x03FE, 0x03D1, 0x03D0, 0x07FE, 0x03FE, 0x0BE5,
+    0x03F1, 0x0FFF, 0x03FF, 0x13F9, 0x03FC, 0x37FF, 0xFE00, 0x0554,
+    0x03FE, 0x03CF, 0x03CD, 0x07FF, 0x03FF, 0x0BE3, 0x03F0, 0x27F9,
+    0x03FC, 0x37FF, 0xFE00, 0x05FE, 0x03FC, 0x0390, 0x002F, 0x0001,
+    0x03FE, 0x03FE, 0x03FF, 0x03FF, 0x000B, 0x03F1, 0x0FFF, 0x03FF,
+    0x1002, 0x0003, 0x37FC, 0xFE00, 0x0616, 0x03FC, 0x038A, 0x0032,
+    0x0002, 0x03FF, 0x03FF, 0x07FF, 0x000B, 0x03F0, 0x0FFF, 0x1403,
+    0x0004, 0x37FC, 0xFE00, 0x0026, 0x03FF, 0x0348, 0x001A, 0x0004,
+    0x1011, 0x03FC, 0x27FE, 0x0002, 0xFE00, 0x0704, 0x0002, 0x03C9,
+    0x0007, 0x1405, 0x03FF, 0xFE00, 0x06E2, 0x03F9, 0x0FFF, 0x03FF,
+    0x1FFE, 0xFE00, 0x070A, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE, 0xFE00,
+    0x04FC, 0x03FE, 0x0045, 0x0048, 0x03FF, 0x03FF, 0x03FF, 0x07FF,
+    0x0028, 0x0017, 0x240A, 0x0005, 0x3402, 0xFE00, 0x0508, 0x03FE,
+    0x0049, 0x004C, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x002B, 0x0018,
+    0x240B, 0x0006, 0x3402, 0xFE00, 0x0134, 0x03FC, 0x003E, 0x0019,
+    0x03FF, 0x1001, 0x03FF, 0x27FF, 0xFE00, 0x06A0, 0x0004, 0x003E,
+    0x0018, 0x03FF, 0x1001, 0x03FF, 0x27FF, 0xFE00, 0x05CC, 0x03FC,
+    0x031F, 0x0060, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0016, 0x03E1,
+    0x0FFF, 0x1406, 0x0007, 0x37F8, 0xFE00, 0x05E4, 0x03FC, 0x0312,
+    0x0065, 0x0003, 0x03FF, 0x03FF, 0x07FF, 0x0017, 0x03E0, 0x0FFF,
+    0x1406, 0x0008, 0x37F8, 0xFE00, 0x079A, 0x03F9, 0x0FFF, 0x03FF,
+    0x1FFE, 0xFE00, 0x07CC, 0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF,
+    0xFE00, 0x00BE, 0x03FE, 0x0060, 0x03BD, 0x03FE, 0x0C01, 0x0012,
+    0x03FB, 0x63FF, 0xFE00, 0x0112, 0x03FD, 0x03FF, 0x03D6, 0x17FC,
+    0x03FB, 0x27FD, 0x03FE, 0x37FF, 0xFE00, 0x0658, 0x03FC, 0x002D,
+    0x002F, 0x07FE, 0x03FE, 0x081A, 0x000F, 0x0FFE, 0x03FF, 0x1007,
+    0x0003, 0x3401, 0xFE00, 0x067C, 0x03FC, 0x0030, 0x0031, 0x03FF,
+    0x03FE, 0x03FE, 0x07FF, 0x001C, 0x0010, 0x0FFE, 0x03FF, 0x1007,
+    0x0004, 0x3401, 0xFE00, 0x06FC, 0x03FA, 0x0356, 0x0048, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0010, 0x03E9, 0x0FFE, 0x03FF, 0x1004,
+    0x0005, 0x37FA, 0xFE00, 0x0724, 0x03FA, 0x034F, 0x004B, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0011, 0x03E8, 0x0FFE, 0x03FF, 0x1004,
+    0x0006, 0x37FA, 0xFE00, 0x06C0, 0x0004, 0x0392, 0x000F, 0x0002,
+    0x100A, 0x03FE, 0x27FF, 0x0001, 0xFE00, 0x0034, 0x0411, 0x03FE,
+    0x17FF, 0xFE00, 0x0058, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF,
+    0xFE00, 0x0090, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF, 0xFE00,
+    0x00FA, 0x03F4, 0x03E8, 0x03E8, 0x07FE, 0x03FD, 0x0BF2, 0x03F8,
+    0x0FFC, 0x03FF, 0x13FD, 0x03FE, 0xFE00, 0x013E, 0x03F6, 0x03E7,
+    0x03E6, 0x07FD, 0x03FC, 0x0BF1, 0x03F8, 0x0FFD, 0x03FF, 0x13FC,
+    0x03FE, 0xFE00, 0x077C, 0x0007, 0x0027, 0x000B, 0x03FC, 0x0FFF,
+    0x03FE, 0x03FE, 0x2801, 0x3401, 0xFE00, 0x07D8, 0x0002, 0x000B,
+    0x0003, 0x03FF, 0xFE00, 0x078A, 0x03F4, 0x0132, 0x039C, 0x07FE,
+    0x03FD, 0x07FF, 0x03C3, 0x0036, 0x0001, 0x0BFD, 0x03FF, 0x03FF,
+    0x0FFB, 0x03ED, 0x3411, 0xFE00, 0x07CC, 0x03F5, 0x0131, 0x03A5,
+    0x0001, 0x03FE, 0x03FD, 0x07FE, 0x03BC, 0x0039, 0x0001, 0x0BFE,
+    0x03FF, 0x03FF, 0x0FFC, 0x03EB, 0x3412, 0xFE00, 0x057C, 0x03F0,
+    0x0FFE, 0x03FC, 0x1FFF, 0x03FF, 0xFE00, 0x05DA, 0x03EE, 0x0FFD,
+    0x03FD, 0x23FF, 0xFE00, 0x07D8, 0x0002, 0x03E3, 0x0002, 0x0003,
+    0x13F7, 0x2BFD, 0x3BFF, 0xFE00, 0x075A, 0x0009, 0x0004, 0x000A,
+    0x13FF, 0x0001, 0x0001, 0xFE00, 0x0560, 0x03F0, 0x0008, 0x0006,
+    0x07FF, 0x03FC, 0x0802, 0x17FF, 0xFE00, 0x05B6, 0x03F0, 0x000A,
+    0x0008, 0x07FE, 0x03FE, 0x0802, 0x17FF, 0xFE00, 0x0538, 0x03F2,
+    0x001A, 0x03EA, 0x03FF, 0x07FD, 0x0806, 0x13FF, 0x03FF, 0x13FF,
+    0x0001, 0x37FF, 0xFE00, 0x0586, 0x03F1, 0x0022, 0x03E5, 0x03FF,
+    0x03FF, 0x03FE, 0x03FF, 0x0408, 0x17FF, 0x13FE, 0x0002, 0x37FF,
+    0xFE00, 0x016A, 0x03F4, 0x039A, 0x002C, 0x0002, 0x03FF, 0x0401,
+    0x03FF, 0x03FB, 0x03FE, 0x2403, 0x03FF, 0xFE00, 0x0778, 0x03FF,
+    0x0007, 0x03FB, 0x0009, 0x03FF, 0x0BFC, 0x0002, 0xFE00, 0x04E0,
+    0x03F4, 0x13FF, 0x1FFD, 0xFE00, 0x0518, 0x03F4, 0x13FF, 0x1FFD,
+    0xFE00, 0x055A, 0x0016, 0x03A7, 0x03ED, 0x03EB, 0x0001, 0x0001,
+    0x03FF, 0x03F8, 0x001A, 0x0013, 0x0BFF, 0x03FF, 0x1002, 0x07F8,
+    0x37F9, 0xFE00, 0x0502, 0x0002, 0x03E6, 0x000C, 0x03F8, 0x0004,
+    0x07FD, 0x03FC, 0x0019, 0x000C, 0x07FF, 0x0FFF, 0x0FFF, 0x03F9,
+    0x37FB, 0xFE00, 0x014E, 0x03F8, 0x0088, 0x0034, 0x03FA, 0x03FF,
+    0x07FF, 0x03FE, 0x03F6, 0x03F8, 0x27FE, 0x0004, 0x3405, 0xFE00,
+    0x07B6, 0x03F8, 0x03DF, 0x03F5, 0x03F9, 0x03FF, 0x07FF, 0x03FD,
+    0x0002, 0x0001, 0x2BFF, 0x37FF, 0xFE00, 0x065A, 0x001E, 0x036C,
+    0x00B3, 0x000D, 0x0001, 0x0001, 0x07F8, 0x0385, 0x0054, 0x0002,
+    0x23D1, 0x001C, 0x37F3, 0xFE00, 0x05E0, 0x0006, 0x0346, 0x00D0,
+    0x0002, 0x0005, 0x0402, 0x03FE, 0x037C, 0x0054, 0x07FF, 0x1FD4,
+    0x001A, 0x37F5, 0xFE00, 0x037E, 0x0003, 0x003E, 0x03F9, 0x03FE,
+    0x1005, 0x03FF, 0x27FF, 0xFE00, 0x0370, 0x0445, 0x03F8, 0x03FF,
+    0x0BFF, 0x0405, 0x03FF, 0xFE00, 0x0194, 0x03F6, 0x03BD, 0x0402,
+    0x03FE, 0x0FFE, 0xFE00, 0x07A6, 0x03F4, 0x0010, 0x0403, 0x03FF,
+    0xFE00, 0x01DE, 0x000E, 0x0067, 0x03FF, 0x03FC, 0x0401, 0x0808,
+    0xFE00, 0x01A8, 0x0002, 0x0071, 0x0802, 0x07FF, 0x0407, 0xFE00,
+    0x07EA, 0x001C, 0x0066, 0x07FB, 0x0001, 0x0002, 0x03FF, 0x0407,
+    0x0BFF, 0xFE00, 0x0780, 0x0006, 0x0071, 0x0804, 0x0C08, 0xFE00,
+    0x02F8, 0x03EE, 0x03FF, 0x03F2, 0x07FC, 0xFE00, 0x0750, 0x03EC,
+    0x0403, 0x07FE, 0xFE00, 0x0646, 0x0027, 0x003F, 0x0006, 0x03FE,
+    0x0001, 0x0003, 0x0805, 0x0001, 0xFE00, 0x05B4, 0x0007, 0x0047,
+    0x0006, 0x03FF, 0x0006, 0x0001, 0x0806, 0x0001, 0xFE00, 0x05AE,
+    0x0026, 0x0012, 0x03FF, 0x0001, 0x0001, 0x0003, 0x07FF, 0x0001,
+    0x03FF, 0xFE00, 0x051C, 0x0008, 0x000A, 0x03FE, 0x0406, 0x0C01,
+    0xFE00, 0x0198, 0x03F6, 0x0044, 0x07FD, 0x03FE, 0x0C02, 0xFE00,
+    0x07A6, 0x03F7, 0x03F0, 0x07FD, 0x03FE, 0xFE00, 0x052E, 0x001B,
+    0x001F, 0x0403, 0x0402, 0x0803, 0xFE00, 0x04C6, 0x0006, 0x0015,
+    0x0804, 0x0C02, 0xFE00, 0x0498, 0x000E, 0x001F, 0x0402, 0x0401,
+    0x0802, 0xFE00, 0x0462, 0x0002, 0x0013, 0x0801, 0x0BFF, 0x0001,
+    0xFE00, 0x001A, 0x040E, 0x0006, 0x1401, 0xFE00, 0x07FA, 0x07FE,
+    0x03FF, 0xFE00, 0x041A, 0x0002, 0x0010, 0x0004, 0x0002, 0x1001,
+    0xFE00, 0x0410, 0x0409, 0x1801, 0xFE00, 0x0400, 0xFE00, 0x0400,
+    0xFE00, 0x0718, 0x0003, 0x001E, 0x03F8, 0x1407, 0xFE00, 0x0136,
+    0x03FD, 0x0052, 0x03B4, 0x03FF, 0x0C01, 0x000D, 0x03F9, 0x27FF,
+    0x3BFF, 0xFE00, 0x0562, 0x03FE, 0x03CD, 0x03CB, 0x07FF, 0x03FF,
+    0x0BE2, 0x03EF, 0x27F8, 0x03FC, 0x37FF, 0xFE00, 0x0576, 0x03FE,
+    0x03CA, 0x03C8, 0x07FE, 0x03FE, 0x0BE0, 0x03EE, 0x0FFF, 0x03FF,
+    0x13F8, 0x03FC, 0x37FF, 0xFE00, 0x062E, 0x03FC, 0x0383, 0x0034,
+    0x0002, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x000C, 0x03EF, 0x0FFF,
+    0x1403, 0x0004, 0x37FC, 0xFE00, 0x064A, 0x03FC, 0x037F, 0x0037,
+    0x0001, 0x03FF, 0x03FF, 0x03FF, 0x040C, 0x03EF, 0x0FFF, 0x1403,
+    0x0004, 0x37FC, 0xFE00, 0x0028, 0x0736, 0x001C, 0x0004, 0x1013,
+    0x03FB, 0x27FE, 0x0002, 0xFE00, 0x06EC, 0x0003, 0x03C3, 0x0008,
+    0x0001, 0x1005, 0x03FF, 0xFE00, 0x072E, 0x03F8, 0x33FE, 0xFE00,
+    0x0754, 0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x0514, 0x03FE, 0x004D,
+    0x0050, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x002D, 0x0019, 0x240C,
+    0x0006, 0x3402, 0xFE00, 0x0524, 0x03FE, 0x0050, 0x0053, 0x03FF,
+    0x03FF, 0x03FF, 0x07FF, 0x002E, 0x001A, 0x0FFF, 0x140C, 0x0006,
+    0x3402, 0xFE00, 0x0150, 0x03FD, 0x0044, 0x001B, 0x03FF, 0x1001,
+    0x03FE, 0x27FE, 0xFE00, 0x067E, 0x0004, 0x0044, 0x001B, 0x03FF,
+    0x1001, 0x03FE, 0x27FE, 0xFE00, 0x05FA, 0x03FC, 0x0306, 0x006A,
+    0x0002, 0x03FF, 0x03FF, 0x07FF, 0x0018, 0x03DE, 0x0FFF, 0x1406,
+    0x0008, 0x37F7, 0xFE00, 0x0614, 0x03FC, 0x02FD, 0x006F, 0x0002,
+    0x03FF, 0x03FF, 0x07FF, 0x0019, 0x03DD, 0x0FFF, 0x03FF, 0x1006,
+    0x0008, 0x37F7, 0xFE00, 0x07FA, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE,
+    0xFE00, 0x002A, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF, 0xFE00,
+    0x00D0, 0x03FE, 0x0069, 0x03B6, 0x03FE, 0x0C01, 0x0014, 0x03FA,
+    0x63FF, 0xFE00, 0x012A, 0x03FC, 0x03FF, 0x03D2, 0x17FC, 0x03FA,
+    0x27FD, 0x03FE, 0x37FF, 0xFE00, 0x0696, 0x03FC, 0x0032, 0x0034,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x001D, 0x0010, 0x0FFF, 0x03FF,
+    0x1007, 0x0004, 0x3401, 0xFE00, 0x06B8, 0x03FC, 0x0035, 0x0036,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x001F, 0x0011, 0x0FFE, 0x03FF,
+    0x1008, 0x0004, 0x3401, 0xFE00, 0x074A, 0x03FB, 0x0346, 0x004F,
+    0x0002, 0x03FE, 0x03FE, 0x07FF, 0x0012, 0x03E7, 0x0FFE, 0x03FF,
+    0x1004, 0x0006, 0x37FA, 0xFE00, 0x0772, 0x03FA, 0x033E, 0x0053,
+    0x0002, 0x03FE, 0x03FE, 0x07FF, 0x0013, 0x03E6, 0x0FFE, 0x03FF,
+    0x1005, 0x0006, 0x37F9, 0xFE00, 0x06A0, 0x0004, 0x0388, 0x0010,
+    0x0002, 0x100B, 0x03FD, 0x27FF, 0x0001, 0xFE00, 0x0038, 0x0413,
+    0x03FD, 0x17FE, 0xFE00, 0x00C8, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD,
+    0x03FF, 0xFE00, 0x0100, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF,
+    0xFE00, 0x017E, 0x03F6, 0x03E6, 0x03E5, 0x07FD, 0x03FC, 0x0BF1,
+    0x03F8, 0x0FFD, 0x03FF, 0x13FC, 0x03FE, 0x37FF, 0xFE00, 0x01BC,
+    0x03F6, 0x03E5, 0x03E4, 0x07FD, 0x03FC, 0x0BF0, 0x03F7, 0x0FFD,
+    0x03FF, 0x13FC, 0x03FE, 0x37FF, 0xFE00, 0x0742, 0x0006, 0x0038,
+    0x0010, 0x03FC, 0x0FFF, 0x03FD, 0x03FD, 0x27FF, 0x0001, 0x3402,
+    0xFE00, 0x07C4, 0x0002, 0x0011, 0x0005, 0x03FF, 0x13FF, 0xFE00,
+    0x000C, 0x03F5, 0x012F, 0x03AE, 0x0001, 0x03FE, 0x03FD, 0x03FF,
+    0x03FE, 0x03B4, 0x003B, 0x0001, 0x0BFE, 0x03FF, 0x03FF, 0x0FFD,
+    0x03E9, 0x3414, 0xFE00, 0x004A, 0x03F6, 0x012D, 0x03B6, 0x07FE,
+    0x03FD, 0x07FE, 0x03AD, 0x003E, 0x0001, 0x0BFE, 0x03FF, 0x03FF,
+    0x0FFE, 0x03E7, 0x3415, 0xFE00, 0x0638, 0x03EE, 0x0FFF, 0x03FD,
+    0x1FFF, 0x03FF, 0xFE00, 0x0692, 0x03EF, 0x0FFE, 0x03FE, 0x23FF,
+    0xFE00, 0x07C8, 0x0002, 0x03D6, 0x0004, 0x0002, 0x13F3, 0x2BFC,
+    0x3BFF, 0xFE00, 0x0712, 0x0008, 0x0006, 0x000E, 0x1402, 0x0002,
+    0x2801, 0xFE00, 0x060E, 0x03EF, 0x000C, 0x0009, 0x07FF, 0x03FD,
+    0x0803, 0x13FF, 0x03FF, 0x13FF, 0xFE00, 0x0662, 0x03F0, 0x000E,
+    0x000C, 0x0BFE, 0x03FF, 0x0403, 0x17FF, 0x13FF, 0x03FF, 0xFE00,
+    0x05D4, 0x03F0, 0x0028, 0x03E0, 0x03FE, 0x07FE, 0x0809, 0x13FF,
+    0x03FF, 0x13FE, 0x0002, 0x37FF, 0xFE00, 0x061E, 0x03F2, 0x002F,
+    0x03DA, 0x03FE, 0x07FE, 0x0401, 0x000B, 0x13FF, 0x03FF, 0x13FD,
+    0x0002, 0x37FF, 0xFE00, 0x0196, 0x03FF, 0x03AA, 0x0027, 0x03F8,
+    0x0C02, 0x03FA, 0x03FF, 0x2402, 0x03FE, 0x3401, 0xFE00, 0x0792,
+    0x03FA, 0x03DE, 0x000A, 0x000C, 0x0FFB, 0x0001, 0x03FF, 0x63FF,
+    0xFE00, 0x0550, 0x03F4, 0x13FF, 0x1FFD, 0xFE00, 0x0586, 0x03F6,
+    0x13FF, 0x1FFE, 0xFE00, 0x052E, 0x03F0, 0x03E4, 0x0010, 0x0004,
+    0x07FF, 0x081D, 0x000E, 0x03FF, 0x13FF, 0x0BFF, 0x03FE, 0x03F8,
+    0x37FA, 0xFE00, 0x0582, 0x03EF, 0x03CE, 0x000B, 0x0006, 0x03FF,
+    0x0801, 0x0024, 0x0013, 0x03FF, 0x13FF, 0x0BFF, 0x03FE, 0x03F6,
+    0x37F8, 0xFE00, 0x014A, 0x0005, 0x0086, 0x0033, 0x0004, 0x0C01,
+    0x03F6, 0x03F8, 0x27FE, 0x0004, 0x3405, 0xFE00, 0x0004, 0x03F3,
+    0x0BF6, 0x0FFC, 0xFE00, 0x0602, 0x03F0, 0x0351, 0x00C8, 0x03FA,
+    0x07FF, 0x0404, 0x037E, 0x0054, 0x03FF, 0x23D3, 0x001A, 0x37F4,
+    0xFE00, 0x0658, 0x03EF, 0x036C, 0x00B3, 0x03F8, 0x03FF, 0x0804,
+    0x0385, 0x0054, 0x03FF, 0x23D1, 0x001C, 0x37F3, 0xFE00, 0x0372,
+    0x03FE, 0x0045, 0x03F7, 0x0001, 0x0FFF, 0x0004, 0x03FF, 0xFE00,
+    0x037A, 0x03FE, 0x0040, 0x03F8, 0x0001, 0x0BFF, 0x0404, 0x03FF,
+    0xFE00, 0x018E, 0x0006, 0x03BF, 0x07FE, 0x13FE, 0xFE00, 0x0008,
+    0x03F1, 0x0804, 0x07FF, 0xFE00, 0x01B6, 0x03F9, 0x006D, 0x0401,
+    0x0FFF, 0x0006, 0xFE00, 0x01DC, 0x03F8, 0x0066, 0x0FFF, 0x03FF,
+    0x0405, 0xFE00, 0x07A0, 0x03F4, 0x006D, 0x0401, 0x07FF, 0x0807,
+    0xFE00, 0x07E8, 0x03F1, 0x0066, 0x0401, 0x07FF, 0x0806, 0x0BFF,
+    0xFE00, 0x02EC, 0x000C, 0x03FE, 0x03F2, 0x07FF, 0xFE00, 0x0004,
+    0x03E3, 0x0C01, 0x03FF, 0x0401, 0xFE00, 0x05E2, 0x03EF, 0x0047,
+    0x0008, 0x0BFF, 0x0804, 0xFE00, 0x0648, 0x03EA, 0x0041, 0x0007,
+    0x0BFF, 0x0805, 0x0001, 0xFE00, 0x0544, 0x03EE, 0x000C, 0x07FE,
+    0x07FF, 0x0FFF, 0xFE00, 0x05AC, 0x03EB, 0x0012, 0x03FD, 0x03FE,
+    0x07FE, 0x0801, 0xFE00, 0x0190, 0x0007, 0x0042, 0x0401, 0x1002,
+    0xFE00, 0x07FC, 0x03F1, 0x0BFB, 0x07FF, 0xFE00, 0x04E4, 0x03F4,
+    0x0018, 0x07FE, 0x07FF, 0x03FF, 0x0401, 0xFE00, 0x052E, 0x03F0,
+    0x0020, 0x07FE, 0x07FF, 0x0803, 0xFE00, 0x046E, 0x03F9, 0x0018,
+    0x07FE, 0x1001, 0xFE00, 0x0494, 0x03F8, 0x0020, 0x07FE, 0x07FF,
+    0x0803, 0xFE00, 0x001A, 0x040E, 0x0006, 0x1401, 0xFE00, 0x07FE,
+    0x0FFF, 0xFE00, 0x0410, 0x040A, 0x0001, 0x03FF, 0x0FFF, 0x0001,
+    0xFE00, 0x041A, 0x03FE, 0x0010, 0x0004, 0x03FF, 0x1001, 0xFE00,
+    0x0400, 0xFE00, 0x0400, 0xFE00, 0x0702, 0x0001, 0x0020, 0x03F8,
+    0x1407, 0x2801, 0xFE00, 0x0154, 0x03FD, 0x0059, 0x03AE, 0x03FF,
+    0x0C01, 0x000F, 0x03F8, 0x27FF, 0x3BFE, 0xFE00, 0x0586, 0x03FE,
+    0x03C8, 0x03C6, 0x07FF, 0x03FF, 0x0BDF, 0x03EE, 0x27F8, 0x03FC,
+    0x37FE, 0xFE00, 0x0594, 0x03FE, 0x03C6, 0x03C4, 0x07FF, 0x03FF,
+    0x0BDE, 0x03ED, 0x27F7, 0x03FB, 0x37FE, 0xFE00, 0x0662, 0x03FC,
+    0x037A, 0x0039, 0x0002, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x000D,
+    0x03EE, 0x0FFF, 0x1403, 0x0004, 0x37FC, 0xFE00, 0x067A, 0x03FC,
+    0x0373, 0x003C, 0x0001, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x000E,
+    0x03ED, 0x0FFF, 0x1403, 0x0004, 0x37FB, 0xFE00, 0x002E, 0x03FF,
+    0x0325, 0x001E, 0x0004, 0x1015, 0x03FB, 0x27FE, 0x0003, 0xFE00,
+    0x06D2, 0x0002, 0x03BF, 0x0009, 0x0001, 0x1006, 0x03FF, 0xFE00,
+    0x0776, 0x03F8, 0x33FE, 0xFE00, 0x079C, 0x03FA, 0x13FF, 0x1FFF,
+    0xFE00, 0x0532, 0x03FE, 0x0053, 0x0056, 0x03FF, 0x03FF, 0x03FF,
+    0x07FF, 0x0030, 0x001B, 0x240D, 0x0006, 0x3402, 0xFE00, 0x053C,
+    0x03FE, 0x0057, 0x005A, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0032,
+    0x001C, 0x240D, 0x0007, 0x3403, 0xFE00, 0x0172, 0x03FC, 0x004A,
+    0x001D, 0x03FF, 0x1002, 0x03FE, 0x27FE, 0xFE00, 0x0658, 0x0004,
+    0x004A, 0x001D, 0x03FF, 0x1001, 0x03FE, 0x27FE, 0xFE00, 0x062C,
+    0x03FC, 0x02F1, 0x0074, 0x0002, 0x03FF, 0x03FF, 0x07FF, 0x001B,
+    0x03DB, 0x0FFF, 0x1407, 0x0009, 0x37F7, 0xFE00, 0x0642, 0x03FC,
+    0x02E6, 0x0078, 0x0003, 0x07FF, 0x07FF, 0x001C, 0x03DA, 0x0FFF,
+    0x1407, 0x0009, 0x37F6, 0xFE00, 0x005A, 0x03F8, 0x0FFF, 0x03FE,
+    0x1FFE, 0x03FF, 0xFE00, 0x0086, 0x03F6, 0x13FF, 0x1FFE, 0xFE00,
+    0x00E2, 0x03FF, 0x0073, 0x03B0, 0x03FE, 0x0C01, 0x0016, 0x03FA,
+    0x63FF, 0xFE00, 0x0146, 0x03FE, 0x03FF, 0x03CE, 0x17FB, 0x03FA,
+    0x27FD, 0x03FE, 0x37FF, 0xFE00, 0x06D4, 0x03FC, 0x0038, 0x0039,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0020, 0x0012, 0x0FFF, 0x03FF,
+    0x1008, 0x0004, 0x3401, 0xFE00, 0x06F4, 0x03FC, 0x0039, 0x003B,
+    0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0021, 0x0013, 0x0FFE, 0x03FF,
+    0x1008, 0x0004, 0x3401, 0xFE00, 0x0796, 0x03FB, 0x0335, 0x0057,
+    0x0002, 0x03FE, 0x03FE, 0x07FF, 0x0014, 0x03E4, 0x0FFE, 0x03FF,
+    0x1005, 0x0007, 0x37F9, 0xFE00, 0x07BE, 0x03FA, 0x032D, 0x005A,
+    0x0002, 0x03FE, 0x03FE, 0x07FF, 0x0015, 0x03E3, 0x0FFE, 0x03FF,
+    0x1005, 0x0007, 0x37F9, 0xFE00, 0x0682, 0x0004, 0x037C, 0x0012,
+    0x0002, 0x100C, 0x03FD, 0x27FF, 0x0001, 0xFE00, 0x0040, 0x0415,
+    0x03FD, 0x17FE, 0xFE00, 0x0138, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD,
+    0x03FF, 0xFE00, 0x0170, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF,
+    0xFE00, 0x01FE, 0x03F4, 0x03E3, 0x03E3, 0x07FE, 0x03FD, 0x0BEF,
+    0x03F7, 0x0FFC, 0x03FF, 0x13FC, 0x03FE, 0x37FF, 0xFE00, 0x023C,
+    0x03F7, 0x03E3, 0x03E2, 0x07FE, 0x03FE, 0x0BEF, 0x03F7, 0x0FFD,
+    0x03FF, 0x13FC, 0x03FE, 0x37FF, 0xFE00, 0x070A, 0x0007, 0x0049,
+    0x0015, 0x03FC, 0x0FFF, 0x03FC, 0x03FC, 0x27FF, 0x0002, 0x3402,
+    0xFE00, 0x07B6, 0x0002, 0x0016, 0x0006, 0x03FF, 0x13FF, 0x03FF,
+    0xFE00, 0x008C, 0x03F4, 0x012C, 0x03BF, 0x07FE, 0x03FD, 0x07FE,
+    0x03A6, 0x0041, 0x0001, 0x0BFD, 0x03FF, 0x03FF, 0x0FFF, 0x03E5,
+    0x3417, 0xFE00, 0x00CA, 0x03F6, 0x012A, 0x03C8, 0x0001, 0x03FE,
+    0x03FE, 0x07FE, 0x039F, 0x0043, 0x0001, 0x0BFE, 0x03FF, 0x03FF,
+    0x13E3, 0x3418, 0xFE00, 0x06F0, 0x03EE, 0x0FFF, 0x03FD, 0x1FFF,
+    0x03FF, 0xFE00, 0x074A, 0x03EF, 0x0FFE, 0x03FE, 0x23FF, 0xFE00,
+    0x07BC, 0x0002, 0x03C9, 0x0004, 0x0003, 0x13EE, 0x2BFB, 0x3BFE,
+    0xFE00, 0x06C8, 0x0008, 0x0008, 0x0013, 0x13FF, 0x0002, 0x0003,
+    0x2801, 0xFE00, 0x06B8, 0x03F0, 0x0010, 0x000C, 0x03FF, 0x03FF,
+    0x03FC, 0x0804, 0x17FF, 0x13FF, 0x03FF, 0xFE00, 0x070E, 0x03F0,
+    0x0012, 0x000E, 0x0BFE, 0x03FF, 0x0404, 0x17FF, 0x13FF, 0x03FF,
+    0xFE00, 0x066C, 0x03F2, 0x0035, 0x03D4, 0x03FE, 0x03FF, 0x03FD,
+    0x080D, 0x13FF, 0x03FF, 0x13FD, 0x0003, 0x37FF, 0xFE00, 0x06B8,
+    0x03F2, 0x003C, 0x03D0, 0x03FE, 0x07FE, 0x080E, 0x13FF, 0x03FF,
+    0x13FC, 0x0003, 0x37FE, 0xFE00, 0x01A0, 0x07CC, 0x001B, 0x03F8,
+    0x0C03, 0x03FA, 0x2802, 0x03FE, 0x3401, 0xFE00, 0x07CC, 0x03F8,
+    0x03A3, 0x0020, 0x000E, 0x0FFA, 0x07FD, 0x2402, 0x3BFF, 0xFE00,
+    0x05BC, 0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x05F4,
+    0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x05E0, 0x03EF,
+    0x03B3, 0x0004, 0x0006, 0x03FF, 0x03FE, 0x082B, 0x0018, 0x03FF,
+    0x13FF, 0x0BFF, 0x03FE, 0x03F3, 0x37F6, 0xFE00, 0x063C, 0x03F0,
+    0x039B, 0x03FF, 0x0005, 0x07FF, 0x0832, 0x001D, 0x03FF, 0x0BFE,
+    0x13FF, 0x03FF, 0x03F1, 0x37F4, 0xFE00, 0x0118, 0x0006, 0x0072,
+    0x002B, 0x0004, 0x0C01, 0x03F8, 0x03FA, 0x27FE, 0x0003, 0x3404,
+    0xFE00, 0x0080, 0x03F0, 0x0031, 0x0013, 0x03F4, 0x0FFB, 0x03FD,
+    0x03FD, 0x27FF, 0x0001, 0x3401, 0xFE00, 0x06BA, 0x03EF, 0x038A,
+    0x009C, 0x03F8, 0x03FF, 0x03FE, 0x0404, 0x038C, 0x0054, 0x03FF,
+    0x0801, 0x17CF, 0x001E, 0x37F2, 0xFE00, 0x0716, 0x03F0, 0x03A7,
+    0x0084, 0x03F9, 0x07FF, 0x0404, 0x0393, 0x0054, 0x03FF, 0x0BFF,
+    0x17CE, 0x0020, 0x37F1, 0xFE00, 0x0386, 0x03FE, 0x003A, 0x03F8,
+    0x0001, 0x1003, 0x03FF, 0xFE00, 0x038E, 0x03FF, 0x0035, 0x03F8,
+    0x1403, 0x03FF, 0xFE00, 0x0152, 0x0007, 0x03C8, 0x07FE, 0x13FE,
+    0xFE00, 0x009A, 0x03EC, 0x03E8, 0x0405, 0x13FF, 0xFE00, 0x0208,
+    0x03F7, 0x005E, 0x0402, 0x07FF, 0x07FF, 0x0005, 0xFE00, 0x0234,
+    0x03F8, 0x0057, 0x0402, 0x07FF, 0x03FF, 0x0405, 0xFE00, 0x003C,
+    0x03F0, 0x005E, 0x0401, 0x07FF, 0x0806, 0x17FF, 0xFE00, 0x008E,
+    0x03F1, 0x0056, 0x0401, 0x03FF, 0x03FF, 0x07FF, 0x0004, 0xFE00,
+    0x027A, 0x000E, 0x03FE, 0x03F4, 0xFE00, 0x0116, 0x03DB, 0x0001,
+    0x03FA, 0x0BFF, 0x0401, 0xFE00, 0x06BE, 0x03E9, 0x003B, 0x0006,
+    0x0BFE, 0x0804, 0xFE00, 0x0732, 0x03EB, 0x0035, 0x0005, 0x0001,
+    0x03FF, 0x03FF, 0x0803, 0xFE00, 0x0622, 0x03EA, 0x0018, 0x03FE,
+    0x03FF, 0x07FE, 0x0802, 0xFE00, 0x0696, 0x03EB, 0x001D, 0x03FD,
+    0x03FF, 0x03FF, 0x03FF, 0x0802, 0x03FF, 0xFE00, 0x0154, 0x0007,
+    0x0039, 0x0402, 0x1002, 0xFE00, 0x0090, 0x03EC, 0x0018, 0x07FB,
+    0x07FF, 0xFE00, 0x0582, 0x03F0, 0x0027, 0x07FD, 0x07FF, 0x0803,
+    0xFE00, 0x05D4, 0x03F1, 0x0030, 0x07FE, 0x07FF, 0x0804, 0xFE00,
+    0x04C0, 0x03F7, 0x0026, 0x07FE, 0x07FF, 0x0803, 0xFE00, 0x04EC,
+    0x03F8, 0x0030, 0x07FE, 0x07FF, 0x0803, 0xFE00, 0x0014, 0x040B,
+    0x0005, 0x1401, 0xFE00, 0x0006, 0x03FF, 0x0004, 0x0002, 0x03FF,
+    0xFE00, 0x0422, 0x03FE, 0x0015, 0x0004, 0x03FE, 0x1001, 0xFE00,
+    0x042C, 0x03FE, 0x001A, 0x0004, 0x03FF, 0x1001, 0xFE00, 0x0400,
+    0xFE00, 0x0400, 0xFE00, 0x06EE, 0x0001, 0x0024, 0x03F7, 0x1408,
+    0x2801, 0xFE00, 0x0170, 0x03FC, 0x0061, 0x03A7, 0x03FF, 0x0C01,
+    0x0010, 0x03F8, 0x27FF, 0x3BFE, 0xFE00, 0x05A8, 0x03FE, 0x03C3,
+    0x03C0, 0x07FE, 0x03FE, 0x0BDC, 0x03EC, 0x0FFF, 0x03FF, 0x13F7,
+    0x03FB, 0x37FE, 0xFE00, 0x05B6, 0x03FE, 0x03C0, 0x03BF, 0x07FF,
+    0x03FF, 0x0BDB, 0x03EB, 0x27F6, 0x03FB, 0x37FE, 0xFE00, 0x0698,
+    0x03FC, 0x036E, 0x003E, 0x0001, 0x03FE, 0x03FE, 0x03FF, 0x03FF,
+    0x000E, 0x03EC, 0x0FFF, 0x03FF, 0x1003, 0x0005, 0x37FB, 0xFE00,
+    0x06AE, 0x03FC, 0x036A, 0x0040, 0x0001, 0x03FF, 0x03FF, 0x03FF,
+    0x03FF, 0x000F, 0x03EC, 0x0FFF, 0x1404, 0x0005, 0x37FB, 0xFE00,
+    0x0030, 0x03FF, 0x0312, 0x0021, 0x0004, 0x1016, 0x03FA, 0x27FD,
+    0x0003, 0xFE00, 0x06B8, 0x0002, 0x03B9, 0x0009, 0x0001, 0x1006,
+    0x03FF, 0x27FF, 0xFE00, 0x07C2, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE,
+    0xFE00, 0x07E4, 0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x0548, 0x03FE,
+    0x005A, 0x005E, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0035, 0x001E,
+    0x240E, 0x0007, 0x3403, 0xFE00, 0x0554, 0x03FE, 0x005D, 0x0061,
+    0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0037, 0x001F, 0x240E, 0x0007,
+    0x3403, 0xFE00, 0x0190, 0x03FD, 0x0050, 0x0020, 0x03FE, 0x1002,
+    0x03FE, 0x27FE, 0xFE00, 0x0636, 0x0004, 0x0050, 0x0020, 0x03FF,
+    0x1002, 0x03FE, 0x27FE, 0xFE00, 0x0658, 0x03FC, 0x02DB, 0x007D,
+    0x0002, 0x03FF, 0x03FF, 0x07FF, 0x001D, 0x03D8, 0x0FFF, 0x1407,
+    0x000A, 0x37F6, 0xFE00, 0x066E, 0x03FC, 0x02CF, 0x0082, 0x0002,
+    0x03FF, 0x03FF, 0x07FF, 0x001E, 0x03D6, 0x0FFF, 0x1408, 0x000A,
+    0x37F5, 0xFE00, 0x00B2, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE, 0xFE00,
+    0x00E2, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF, 0xFE00, 0x00F6,
+    0x03FE, 0x007C, 0x03A9, 0x03FE, 0x0C01, 0x0017, 0x03F9, 0x63FF,
+    0xFE00, 0x0162, 0x03FD, 0x03FE, 0x03CA, 0x17FB, 0x03F9, 0x27FC,
+    0x03FE, 0x37FE, 0xFE00, 0x0710, 0x03FC, 0x003B, 0x003E, 0x03FF,
+    0x03FE, 0x03FE, 0x07FF, 0x0023, 0x0013, 0x0FFF, 0x03FF, 0x1009,
+    0x0005, 0x3402, 0xFE00, 0x0730, 0x03FC, 0x003D, 0x0040, 0x03FF,
+    0x03FE, 0x03FE, 0x07FF, 0x0024, 0x0014, 0x0FFE, 0x03FF, 0x1009,
+    0x0005, 0x3402, 0xFE00, 0x07E4, 0x03FB, 0x0323, 0x005E, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0016, 0x03E2, 0x0FFE, 0x03FF, 0x1005,
+    0x0007, 0x37F8, 0xFE00, 0x000C, 0x03FA, 0x031C, 0x0061, 0x0002,
+    0x03FE, 0x03FE, 0x07FF, 0x0016, 0x03E1, 0x0FFE, 0x03FF, 0x1006,
+    0x0007, 0x37F8, 0xFE00, 0x0662, 0x0004, 0x0371, 0x0013, 0x0002,
+    0x100D, 0x03FD, 0x27FE, 0x0001, 0xFE00, 0x0046, 0x0417, 0x03FD,
+    0x17FE, 0xFE00, 0x01A8, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF,
+    0xFE00, 0x01E0, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF, 0xFE00,
+    0x027C, 0x03F4, 0x03E2, 0x03E0, 0x07FE, 0x03FD, 0x0BEE, 0x03F6,
+    0x0FFC, 0x03FF, 0x13FC, 0x03FE, 0x37FF, 0xFE00, 0x02BC, 0x03F5,
+    0x03E0, 0x03E0, 0x07FF, 0x03FE, 0x0BEE, 0x03F6, 0x0FFD, 0x03FF,
+    0x13FB, 0x03FE, 0x37FF, 0xFE00, 0x06CE, 0x0007, 0x005A, 0x001A,
+    0x03FC, 0x0FFF, 0x03FB, 0x03FC, 0x27FF, 0x0002, 0x3403, 0xFE00,
+    0x07A4, 0x0002, 0x001B, 0x0008, 0x03FF, 0x13FF, 0x03FF, 0xFE00,
+    0x010A, 0x03F4, 0x0128, 0x03D0, 0x07FE, 0x03FD, 0x07FE, 0x0398,
+    0x0046, 0x0001, 0x0BFD, 0x03FF, 0x03FF, 0x13E0, 0x341A, 0xFE00,
+    0x014A, 0x03F4, 0x0125, 0x03D9, 0x07FF, 0x03FE, 0x07FE, 0x0391,
+    0x0049, 0x0001, 0x0BFE, 0x03FF, 0x03FF, 0x0C01, 0x03DE, 0x341B,
+    0xFE00, 0x07A8, 0x03EE, 0x0FFF, 0x03FD, 0x1FFF, 0x03FF, 0xFE00,
+    0x0008, 0x03EE, 0x0FFF, 0x03FD, 0x1FFF, 0x03FF, 0xFE00, 0x07AA,
+    0x0002, 0x03BC, 0x0005, 0x0003, 0x13EA, 0x0001, 0x27F9, 0x3BFE,
+    0xFE00, 0x0680, 0x0009, 0x000B, 0x0017, 0x13FF, 0x0003, 0x0004,
+    0x2401, 0x0002, 0xFE00, 0x0762, 0x03EF, 0x0015, 0x0010, 0x0BFD,
+    0x03FF, 0x0405, 0x17FF, 0x13FF, 0x03FF, 0xFE00, 0x07BA, 0x03F0,
+    0x0017, 0x0012, 0x0BFE, 0x0805, 0x13FF, 0x03FF, 0x13FF, 0x03FF,
+    0xFE00, 0x0704, 0x03F0, 0x0041, 0x03CC, 0x03FE, 0x07FE, 0x080F,
+    0x13FF, 0x03FF, 0x13FC, 0x0004, 0x37FE, 0xFE00, 0x0752, 0x03F1,
+    0x0048, 0x03C6, 0x03FF, 0x07FE, 0x0811, 0x13FF, 0x03FF, 0x13FC,
+    0x0004, 0x37FE, 0xFE00, 0x01AE, 0x03FE, 0x03EC, 0x0010, 0x03F9,
+    0x0C02, 0x03FA, 0x0001, 0x2401, 0x03FE, 0x3402, 0xFE00, 0x07F4,
+    0x0770, 0x0034, 0x0007, 0x03FF, 0x0401, 0x03FD, 0x07FB, 0x2403,
+    0x3BFE, 0xFE00, 0x0630, 0x03F4, 0x13FF, 0x1FFD, 0xFE00, 0x0668,
+    0x03F4, 0x13FF, 0x1FFD, 0xFE00, 0x068C, 0x03F2, 0x0388, 0x03FC,
+    0x0002, 0x03FF, 0x03FE, 0x0838, 0x0022, 0x03FF, 0x0BFF, 0x13FF,
+    0x03FF, 0x03EF, 0x37F3, 0xFE00, 0x06B4, 0x0789, 0x07FA, 0x03FF,
+    0x0401, 0x03FD, 0x003C, 0x0023, 0x0FFF, 0x17FE, 0x03EE, 0x37F2,
+    0xFE00, 0x00EE, 0x0004, 0x0060, 0x0025, 0x0004, 0x03FF, 0x07FF,
+    0x0001, 0x03F9, 0x03FB, 0x27FF, 0x0003, 0x3403, 0xFE00, 0x00E2,
+    0x03FA, 0x005A, 0x0023, 0x03FA, 0x03FF, 0x07FF, 0x03FF, 0x03FA,
+    0x03FB, 0x27FF, 0x0002, 0x3403, 0xFE00, 0x0766, 0x03F3, 0x03C0,
+    0x0071, 0x03FA, 0x03FF, 0x03FE, 0x0402, 0x0399, 0x0054, 0x03FF,
+    0x23CC, 0x0021, 0x37F0, 0xFE00, 0x0784, 0x0004, 0x03C9, 0x0069,
+    0x0002, 0x03FE, 0x07FF, 0x03FF, 0x039B, 0x0054, 0x27CB, 0x0022,
+    0x37F0, 0xFE00, 0x0396, 0x0430, 0x03FB, 0x1403, 0xFE00, 0x0396,
+    0x0430, 0x03FB, 0x1403, 0xFE00, 0x0120, 0x0004, 0x03D1, 0x07FE,
+    0x13FF, 0xFE00, 0x0112, 0x03F9, 0x03D4, 0x0402, 0x03FE, 0x0FFF,
+    0xFE00, 0x0258, 0x03FA, 0x0050, 0x13FF, 0x0404, 0x0BFF, 0xFE00,
+    0x0262, 0x044D, 0x0BFF, 0x0C04, 0xFE00, 0x00D4, 0x03F4, 0x0050,
+    0x0FFF, 0x0805, 0xFE00, 0x00EC, 0x0002, 0x004D, 0x0BFF, 0x0C05,
+    0xFE00, 0x0216, 0x000C, 0x07F6, 0x03FF, 0x03FE, 0x0001, 0xFE00,
+    0x01F8, 0x03F0, 0x07F6, 0x07FC, 0xFE00, 0x0792, 0x03F1, 0x0032,
+    0x0006, 0x07FF, 0x03FF, 0x07FF, 0x0003, 0xFE00, 0x07B8, 0x0004,
+    0x0030, 0x0005, 0x03FF, 0x03FF, 0x0C03, 0x0001, 0xFE00, 0x06F6,
+    0x03F0, 0x0022, 0x03FC, 0x03FF, 0x03FF, 0x03FF, 0x0802, 0x0BFF,
+    0xFE00, 0x071E, 0x0004, 0x0022, 0x03FC, 0x07FD, 0x03FF, 0x0803,
+    0x03FF, 0x0FFF, 0xFE00, 0x011C, 0x0007, 0x002F, 0x0402, 0x03FF,
+    0x0C01, 0xFE00, 0x010A, 0x03F7, 0x002C, 0x07FD, 0x03FF, 0x0C01,
+    0xFE00, 0x061A, 0x03F5, 0x0036, 0x0FFF, 0x07FF, 0x0004, 0xFE00,
+    0x0636, 0x0002, 0x0039, 0x0BFE, 0x0C04, 0xFE00, 0x050E, 0x03FA,
+    0x0036, 0x07FF, 0x03FF, 0x03FF, 0x07FF, 0x0004, 0x1BFF, 0xFE00,
+    0x051E, 0x0439, 0x0BFF, 0x07FF, 0x0404, 0xFE00, 0x0010, 0x0409,
+    0x0004, 0x1401, 0xFE00, 0x0010, 0x0408, 0x0004, 0x1401, 0xFE00,
+    0x0434, 0x03FF, 0x0020, 0x0006, 0x03FF, 0x0BFF, 0x03FF, 0x0002,
+    0xFE00, 0x0436, 0x0420, 0x0005, 0x1403, 0xFE00, 0x0400, 0xFE00,
+    0x0400, 0xFE00, 0x06D6, 0x0002, 0x0026, 0x03F7, 0x03FF, 0x1009,
+    0x2801, 0xFE00, 0x018E, 0x03FD, 0x0068, 0x03A0, 0x03FF, 0x0C01,
+    0x0011, 0x03F7, 0x27FF, 0x3BFE, 0xFE00, 0x05CA, 0x03FE, 0x03BE,
+    0x03BC, 0x07FE, 0x03FE, 0x0BD9, 0x03EA, 0x0FFF, 0x03FF, 0x13F6,
+    0x03FB, 0x37FE, 0xFE00, 0x05DA, 0x03FE, 0x03BC, 0x03B9, 0x07FF,
+    0x03FF, 0x0BD8, 0x03EA, 0x27F6, 0x03FB, 0x37FE, 0xFE00, 0x06CC,
+    0x03FC, 0x0365, 0x0042, 0x0001, 0x03FE, 0x03FE, 0x03FF, 0x03FF,
+    0x000F, 0x03EB, 0x0FFF, 0x03FF, 0x1004, 0x0005, 0x37FB, 0xFE00,
+    0x06E4, 0x03FC, 0x035F, 0x0044, 0x0002, 0x03FF, 0x03FF, 0x03FF,
+    0x03FF, 0x0010, 0x03EA, 0x0FFF, 0x1404, 0x0005, 0x37FB, 0xFE00,
+    0x0038, 0x0700, 0x0024, 0x0004, 0x1018, 0x03FA, 0x27FD, 0x0003,
+    0xFE00, 0x06A0, 0x0003, 0x03B4, 0x000A, 0x0001, 0x1007, 0x03FE,
+    0x27FF, 0x0001, 0xFE00, 0x000A, 0x03F9, 0x0FFF, 0x03FF, 0x1FFE,
+    0xFE00, 0x002E, 0x03F8, 0x33FE, 0xFE00, 0x0566, 0x03FE, 0x0061,
+    0x0064, 0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x0038, 0x0020, 0x0FFF,
+    0x140F, 0x0008, 0x3403, 0xFE00, 0x0572, 0x03FE, 0x0064, 0x0068,
+    0x03FF, 0x03FF, 0x03FF, 0x07FF, 0x003A, 0x0021, 0x240F, 0x0008,
+    0x3403, 0xFE00, 0x01B2, 0x03FD, 0x0057, 0x0022, 0x03FF, 0x1002,
+    0x03FE, 0x27FE, 0xFE00, 0x0612, 0x0004, 0x0056, 0x0022, 0x03FF,
+    0x1002, 0x03FE, 0x27FE, 0xFE00, 0x068A, 0x03FC, 0x02C5, 0x0087,
+    0x0002, 0x03FF, 0x03FF, 0x07FF, 0x001F, 0x03D5, 0x0FFF, 0x03FF,
+    0x1008, 0x000A, 0x37F5, 0xFE00, 0x06A0, 0x03FC, 0x02BA, 0x008B,
+    0x0003, 0x07FF, 0x07FF, 0x0020, 0x03D4, 0x0FFF, 0x1408, 0x000B,
+    0x37F5, 0xFE00, 0x0112, 0x03F8, 0x0FFF, 0x03FE, 0x1FFE, 0x03FF,
+    0xFE00, 0x013E, 0x03F6, 0x13FF, 0x1FFE, 0xFE00, 0x010C, 0x03FE,
+    0x0086, 0x03A2, 0x03FE, 0x0C01, 0x0019, 0x03F8, 0x63FF, 0xFE00,
+    0x0180, 0x03FD, 0x03FF, 0x03C6, 0x17FA, 0x03F8, 0x27FC, 0x03FE,
+    0x37FE, 0xFE00, 0x0752, 0x03FC, 0x0040, 0x0042, 0x03FF, 0x03FE,
+    0x03FE, 0x07FF, 0x0025, 0x0015, 0x0FFE, 0x03FF, 0x100A, 0x0005,
+    0x3402, 0xFE00, 0x076E, 0x03FC, 0x0042, 0x0044, 0x03FF, 0x03FE,
+    0x03FE, 0x07FF, 0x0026, 0x0016, 0x0FFF, 0x03FF, 0x100A, 0x0005,
+    0x3402, 0xFE00, 0x0034, 0x03FA, 0x0315, 0x0064, 0x0002, 0x03FE,
+    0x03FE, 0x07FF, 0x0017, 0x03E0, 0x0FFE, 0x03FF, 0x1006, 0x0008,
+    0x37F8, 0xFE00, 0x0058, 0x03FA, 0x030D, 0x0068, 0x0002, 0x03FE,
+    0x03FE, 0x03FF, 0x03FF, 0x0018, 0x03DF, 0x0FFE, 0x03FF, 0x1006,
+    0x0008, 0x37F8, 0xFE00, 0x0644, 0x0003, 0x0367, 0x0015, 0x0002,
+    0x100E, 0x03FC, 0x27FE, 0x0002, 0xFE00, 0x004C, 0x0419, 0x03FD,
+    0x17FE, 0xFE00, 0x0218, 0x03F7, 0x0FFE, 0x03FE, 0x1FFD, 0x03FF,
+    0xFE00, 0x024C, 0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00,
+    0x02FE, 0x03F6, 0x03DF, 0x03DE, 0x07FD, 0x03FC, 0x0BED, 0x03F5,
+    0x0FFD, 0x03FF, 0x13FB, 0x03FE, 0x37FF, 0xFE00, 0x033C, 0x03F6,
+    0x03DE, 0x03DC, 0x07FE, 0x03FD, 0x0BEC, 0x03F5, 0x0FFD, 0x03FF,
+    0x13FB, 0x03FE, 0x37FF, 0xFE00, 0x0694, 0x0006, 0x006C, 0x001F,
+    0x03FC, 0x0FFF, 0x03FA, 0x03FB, 0x27FF, 0x0003, 0x3403, 0xFE00,
+    0x0792, 0x0001, 0x0020, 0x0009, 0x03FF, 0x13FF, 0x03FF, 0x6001,
+    0xFE00, 0x018C, 0x03F6, 0x0124, 0x03E2, 0x07FE, 0x03FD, 0x03FF,
+    0x03FE, 0x0389, 0x004C, 0x0001, 0x0BFE, 0x03FF, 0x03FF, 0x0C02,
+    0x03DC, 0x341D, 0xFE00, 0x01CC, 0x03F6, 0x0122, 0x03EA, 0x07FE,
+    0x03FD, 0x03FF, 0x03FE, 0x0382, 0x004E, 0x0001, 0x0BFE, 0x03FF,
+    0x03FF, 0x0C02, 0x03DA, 0x341E, 0xFE00, 0x0066, 0x03F0, 0x0FFF,
+    0x03FC, 0x23FF, 0xFE00, 0x00C0, 0x03EF, 0x0FFF, 0x03FE, 0x1FFF,
+    0x03FF, 0xFE00, 0x079A, 0x0001, 0x03AF, 0x0006, 0x0002, 0x13E6,
+    0x0001, 0x27F8, 0x3BFE, 0xFE00, 0x0638, 0x0008, 0x000D, 0x001C,
+    0x13FF, 0x0004, 0x0005, 0x2401, 0x0002, 0xFE00, 0x0010, 0x03F0,
+    0x001A, 0x0014, 0x07FF, 0x03FE, 0x0806, 0x17FF, 0x13FF, 0x03FF,
+    0x37FF, 0xFE00, 0x0064, 0x03F0, 0x001C, 0x0016, 0x07FF, 0x03FC,
+    0x0806, 0x17FF, 0x13FE, 0x03FF, 0x37FF, 0xFE00, 0x079E, 0x03F1,
+    0x004E, 0x03C1, 0x03FF, 0x03FF, 0x03FE, 0x03FF, 0x0413, 0x17FF,
+    0x13FB, 0x0005, 0x37FE, 0xFE00, 0x07EA, 0x03F2, 0x0055, 0x03BC,
+    0x03FF, 0x07FD, 0x0814, 0x13FF, 0x03FF, 0x13FB, 0x0005, 0x37FD,
+    0xFE00, 0x0182, 0x000C, 0x0029, 0x03F9, 0x03EB, 0x03FF, 0x0808,
+    0x03FA, 0x0003, 0x2BFE, 0x3402, 0xFE00, 0x07E8, 0x0001, 0x0359,
+    0x003B, 0x0005, 0x0FFE, 0x07FA, 0x2404, 0x0001, 0x37FE, 0xFE00,
+    0x06A0, 0x03F4, 0x13FF, 0x1FFD, 0xFE00, 0x06D8, 0x03F4, 0x13FF,
+    0x1FFD, 0xFE00, 0x06A2, 0x0004, 0x03A6, 0x0011, 0x03F7, 0x0FFC,
+    0x003E, 0x0021, 0x17FF, 0x0FFD, 0x03EE, 0x37F2, 0xFE00, 0x0680,
+    0x0006, 0x03CA, 0x0026, 0x03F6, 0x0001, 0x0BFC, 0x003F, 0x001E,
+    0x0BFF, 0x0BFF, 0x0FFB, 0x03EE, 0x37F2, 0xFE00, 0x0078, 0x0017,
+    0x002D, 0x0012, 0x0011, 0x03FF, 0x0807, 0x03FD, 0x03FE, 0x03FF,
+    0x1FFF, 0x0401, 0x3401, 0xFE00, 0x00FE, 0x03FC, 0x0067, 0x0027,
+    0x03FD, 0x0FFF, 0x03F9, 0x03FA, 0x27FE, 0x0003, 0x3404, 0xFE00,
+    0x075E, 0x0009, 0x03BD, 0x0073, 0x0004, 0x0FFE, 0x0398, 0x0054,
+    0x27CC, 0x0021, 0x37F0, 0xFE00, 0x0724, 0x000A, 0x03AB, 0x0081,
+    0x0004, 0x0001, 0x0001, 0x07FD, 0x0394, 0x0054, 0x27CD, 0x0020,
+    0x37F1, 0xFE00, 0x0394, 0x0431, 0x03FA, 0x03FF, 0x1003, 0xFE00,
+    0x038E, 0x0435, 0x03F9, 0x03FF, 0x1003, 0x03FF, 0xFE00, 0x0094,
+    0x001C, 0x03E9, 0x07F8, 0x03FF, 0x0FFF, 0xFE00, 0x0134, 0x03FC,
+    0x03CD, 0x0401, 0x13FE, 0xFE00, 0x0252, 0x0004, 0x0050, 0x07FF,
+    0x1004, 0xFE00, 0x0238, 0x0004, 0x0055, 0x07FE, 0x1005, 0xFE00,
+    0x00CC, 0x0008, 0x0050, 0x07FE, 0x1004, 0xFE00, 0x009C, 0x0009,
+    0x0055, 0x07FE, 0x1005, 0xFE00, 0x0108, 0x0033, 0x0001, 0x03FB,
+    0x07FE, 0x0001, 0x07FE, 0xFE00, 0x0242, 0x03F8, 0x03FF, 0x03F5,
+    0xFE00, 0x0788, 0x000A, 0x0032, 0x0006, 0x0801, 0x0803, 0xFE00,
+    0x0744, 0x000D, 0x0035, 0x0005, 0x03FF, 0x0401, 0x0803, 0xFE00,
+    0x06F0, 0x000B, 0x0020, 0x03FC, 0x1402, 0x03FF, 0xFE00, 0x06AC,
+    0x000D, 0x001C, 0x03FC, 0x0801, 0x0802, 0x03FF, 0xFE00, 0x0086,
+    0x001C, 0x0017, 0x0408, 0x03FF, 0x0001, 0xFE00, 0x0136, 0x03FC,
+    0x0034, 0x07FF, 0x1002, 0xFE00, 0x0616, 0x0008, 0x0036, 0x1805,
+    0xFE00, 0x05E4, 0x0009, 0x0032, 0x0401, 0x1004, 0xFE00, 0x050E,
+    0x0004, 0x0036, 0x0401, 0x1005, 0xFE00, 0x04F4, 0x0004, 0x0032,
+    0x1805, 0xFE00, 0x0006, 0x0001, 0x0003, 0x0002, 0x0001, 0xFE00,
+    0x0012, 0x040A, 0x0005, 0x1401, 0xFE00, 0x0432, 0x041E, 0x0004,
+    0x1402, 0xFE00, 0x042E, 0x041B, 0x0004, 0x1402, 0xFE00, 0x0400,
+    0xFE00, 0x0400, 0xFE00, 0x06C2, 0x0002, 0x0029, 0x03F6, 0x1409,
+    0x2801, 0x0001, 0xFE00, 0x01AC, 0x03FD, 0x0070, 0x0399, 0x03FF,
+    0x0C01, 0x0013, 0x03F6, 0x27FF, 0x3BFE, 0xFE00, 0x05E8, 0x03FE,
+    0x03BA, 0x03B7, 0x07FF, 0x03FF, 0x0BD7, 0x03E9, 0x27F5, 0x03FB,
+    0x37FE, 0xFE00, 0x05F6, 0x03FE, 0x03B8, 0x03B5, 0x07FF, 0x03FF,
+    0x0BD6, 0x03E8, 0x27F5, 0x03FA, 0x37FE, 0xFE00, 0x06FC, 0x03FC,
+    0x0358, 0x0047, 0x0001, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x0010,
+    0x03E9, 0x0FFF, 0x1404, 0x0005, 0x37FA, 0xFE00, 0x0712, 0x03FC,
+    0x0353, 0x0049, 0x0002, 0x03FF, 0x03FF, 0x03FF, 0x03FF, 0x0011,
+    0x03E8, 0x0FFF, 0x1404, 0x0005, 0x37FA, 0xFE00, 0x0038, 0x03FF,
+    0x02EF, 0x0026, 0x0004, 0x101A, 0x03F9, 0x27FD, 0x0003, 0xFE00,
+    0x0688, 0x0002, 0x03AF, 0x000B, 0x1407, 0x03FE, 0x27FF, 0x0001,
+    0xFE00, 0x0054, 0x03FA, 0x13FF, 0x1FFF, 0xFE00, 0x0076, 0x03F8,
+    0x33FE, 0xFE00, 0x057E, 0x03FE, 0x0068, 0x006C, 0x03FF, 0x03FF,
+    0x03FF, 0x07FF, 0x003C, 0x0022, 0x2410, 0x0008, 0x3403, 0xFE00,
+    0x058A, 0x03FE, 0x006B, 0x006F, 0x03FF, 0x03FF, 0x03FF, 0x07FF,
+    0x003E, 0x0023, 0x2410, 0x0008, 0x3403, 0xFE00, 0x01CE, 0x03FD,
+    0x005C, 0x0025, 0x03FF, 0x1002, 0x03FE, 0x27FE, 0xFE00, 0x05F2,
+    0x0004, 0x005C, 0x0024, 0x03FF, 0x1002, 0x03FE, 0x27FE, 0xFE00,
+    0x06B6, 0x03FC, 0x02AF, 0x0090, 0x0002, 0x03FF, 0x03FF, 0x07FF,
+    0x0021, 0x03D2, 0x0FFF, 0x1409, 0x000B, 0x37F4, 0xFE00, 0x06CE,
+    0x03FC, 0x02A2, 0x0095, 0x0003, 0x03FF, 0x03FF, 0x07FF, 0x0022,
+    0x03D0, 0x0FFF, 0x1409, 0x000C, 0x37F4, 0xFE00, 0x016A, 0x03F9,
+    0x0FFF, 0x03FF, 0x1FFE, 0xFE00, 0x019C, 0x03F6, 0x0FFF, 0x03FE,
+    0x1FFD, 0x03FF, 0xFE00, 0x011A, 0x03FF, 0x008F, 0x039C, 0x03FE,
+    0x0C01, 0x001B, 0x03F8, 0x63FE, 0xFE00, 0x0196, 0x03FE, 0x03FE,
+    0x03C2, 0x17FA, 0x03F8, 0x27FC, 0x03FE, 0x37FE, 0xFE00, 0x0788,
+    0x03FC, 0x0045, 0x0047, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0028,
+    0x0016, 0x0FFF, 0x03FF, 0x100A, 0x0005, 0x3402, 0xFE00, 0x07AA,
+    0x03FC, 0x0047, 0x0049, 0x03FF, 0x03FE, 0x03FE, 0x07FF, 0x0029,
+    0x0017, 0x0FFE, 0x03FF, 0x100B, 0x0005, 0x3402, 0xFE00, 0x007C,
+    0x03FA, 0x0303, 0x006C, 0x0002, 0x03FE, 0x03FE, 0x03FF, 0x03FF,
+    0x0019, 0x03DE, 0x0FFE, 0x03FF, 0x1006, 0x0008, 0x37F7, 0xFE00,
+    0x00A4, 0x03FA, 0x02FB, 0x006F, 0x0002, 0x03FE, 0x03FE, 0x03FF,
+    0x03FF, 0x001A, 0x03DC, 0x0FFE, 0x03FF, 0x1006, 0x0008, 0x37F7,
+    0xFE00, 0x0624, 0x0003, 0x035C, 0x0016, 0x0002, 0x100F, 0x03FC,
+    0x27FE, 0x0002, 0xFE00, 0x0052, 0x041B, 0x03FC, 0x17FE, 0xFE00,
+    0x0284, 0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x02BC,
+    0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00, 0x037C, 0x03F6,
+    0x03DC, 0x03DB, 0x07FE, 0x03FD, 0x0BEB, 0x03F4, 0x0FFD, 0x03FF,
+    0x13FB, 0x03FD, 0x37FF, 0xFE00, 0x03BA, 0x03F6, 0x03DB, 0x03DA,
+    0x07FE, 0x03FD, 0x0BEB, 0x03F4, 0x0FFD, 0x03FF, 0x13FB, 0x03FD,
+    0x37FF, 0xFE00, 0x065A, 0x0007, 0x007D, 0x0024, 0x03FC, 0x0FFF,
+    0x03FA, 0x03FA, 0x27FE, 0x0003, 0x3404, 0xFE00, 0x0784, 0x0002,
+    0x0024, 0x000A, 0x03FF, 0x13FE, 0x03FE, 0x2801, 0x3401, 0xFE00,
+    0x020A, 0x03F6, 0x0120, 0x03F3, 0x07FE, 0x03FD, 0x07FF, 0x037B,
+    0x0051, 0x0001, 0x0BFE, 0x03FF, 0x03FF, 0x0C03, 0x03D8, 0x3420,
+    0xFE00, 0x024A, 0x03F5, 0x011F, 0x03FC, 0x0001, 0x03FE, 0x03FD,
+    0x03FF, 0x03FE, 0x0374, 0x0054, 0x0001, 0x0BFE, 0x03FF, 0x03FF,
+    0x0C04, 0x03D6, 0x3421, 0xFE00, 0x011C, 0x03F0, 0x0FFE, 0x03FC,
+    0x1FFF, 0x03FF, 0xFE00, 0x0178, 0x03EE, 0x0FFF, 0x03FD, 0x1FFF,
+    0x03FF, 0xFE00, 0x078C, 0x0002, 0x03A2, 0x0007, 0x0003, 0x13E2,
+    0x0001, 0x27F7, 0x3BFD, 0xFE00, 0x05EE, 0x0008, 0x000F, 0x0021,
+    0x13FF, 0x0004, 0x0006, 0x2401, 0x0003, 0xFE00, 0x00BA, 0x03F0,
+    0x001D, 0x0018, 0x0BFE, 0x03FF, 0x0407, 0x13FF, 0x03FF, 0x13FE,
+    0x03FE, 0x37FF, 0xFE00, 0x010E, 0x03EF, 0x0020, 0x0019, 0x07FF,
+    0x03FD, 0x0807, 0x17FF, 0x13FE, 0x03FE, 0x37FF, 0xFE00, 0x0036,
+    0x03F2, 0x005B, 0x03B7, 0x03FE, 0x07FE, 0x0401, 0x0016, 0x13FF,
+    0x03FF, 0x13FA, 0x0005, 0x37FD, 0xFE00, 0x0082, 0x03F1, 0x0062,
+    0x03B2, 0x03FE, 0x03FF, 0x03FE, 0x0401, 0x0017, 0x17FF, 0x13FA,
+    0x0006, 0x37FD, 0xFE00, 0x0130, 0x0007, 0x007C, 0x03D9, 0x03EF,
+    0x0BFF, 0x0007, 0x03FB, 0x0006, 0x27FE, 0x03FE, 0x3403, 0xFE00,
+    0x07DE, 0x0001, 0x0343, 0x0043, 0x0004, 0x0FFF, 0x0001, 0x03F9,
+    0x2404, 0x0001, 0x37FD, 0xFE00, 0x070E, 0x03F6, 0x13FF, 0x1FFE,
+    0xFE00, 0x0744, 0x03F6, 0x0FFF, 0x03FE, 0x1FFD, 0x03FF, 0xFE00,
+    0x065E, 0x0005, 0x03EE, 0x003A, 0x03F8, 0x0FFC, 0x0040, 0x001B,
+    0x17FF, 0x0FFA, 0x03EE, 0x37F3, 0xFE00, 0x0646, 0x0003, 0x000C,
+    0x004C, 0x03F8, 0x0FFC, 0x0041, 0x0018, 0x17FF, 0x0FF9, 0x03EE,
+    0x37F3, 0xFE00, 0x07CC, 0x0011, 0x03E8, 0x03F8, 0x000E, 0x0001,
+    0x0805, 0x0001, 0x0001, 0xFE00, 0x011E, 0x03FD, 0x0074, 0x002C,
+    0x03FE, 0x13F8, 0x03FA, 0x27FE, 0x0003, 0x3404, 0xFE00, 0x06EA,
+    0x000A, 0x0399, 0x008F, 0x0004, 0x0FFE, 0x0390, 0x0054, 0x27CE,
+    0x001F, 0x37F2, 0xFE00, 0x06BE, 0x0007, 0x038B, 0x009A, 0x0002,
+    0x0FFE, 0x038C, 0x0054, 0x27CF, 0x001E, 0x37F2, 0xFE00, 0x0388,
+    0x0438, 0x03F8, 0x1404, 0x03FF, 0xFE00, 0x0384, 0x043A, 0x03F8,
+    0x1404, 0x03FF, 0xFE00, 0x07C6, 0x0015, 0x000A, 0x07F9, 0x0001,
+    0xFE00, 0x0158, 0x03FC, 0x03C8, 0x0401, 0x13FE, 0xFE00, 0x0220,
+    0x0004, 0x005A, 0x07FF, 0x1005, 0xFE00, 0x0208, 0x0003, 0x005D,
+    0x03FF, 0x03FF, 0x1005, 0xFE00, 0x006A, 0x0009, 0x0059, 0x07FE,
+    0x1005, 0xFE00, 0x0040, 0x0006, 0x005D, 0x07FF, 0x1005, 0xFE00,
+    0x0786, 0x002A, 0x0001, 0x0002, 0x0402, 0x0001, 0x07FF, 0xFE00,
+    0x0284, 0x03F9, 0x03FE, 0x03F4, 0x07FF, 0xFE00, 0x06FE, 0x000C,
+    0x0038, 0x0005, 0x03FF, 0x1004, 0xFE00, 0x06C6, 0x0008, 0x003A,
+    0x0006, 0x1404, 0xFE00, 0x0666, 0x000C, 0x0018, 0x03FC, 0x0001,
+    0x1002, 0x03FF, 0xFE00, 0x062C, 0x0008, 0x0017, 0x03FD, 0x1402,
+    0x03FF, 0xFE00, 0x07BA, 0x0016, 0x03F5, 0x0407, 0x0001, 0xFE00,
+    0x0158, 0x03FC, 0x003A, 0x07FF, 0x1002, 0xFE00, 0x05B2, 0x0008,
+    0x002D, 0x0401, 0x1004, 0xFE00, 0x0588, 0x0006, 0x0029, 0x1804,
+    0xFE00, 0x04DA, 0x0004, 0x002D, 0x0401, 0x1004, 0xFE00, 0x04C6,
+    0x0002, 0x0028, 0x1804, 0xFE00, 0x07FA, 0x07FD, 0x03FF, 0xFE00,
+    0x0014, 0x040B, 0x0005, 0x1401, 0xFE00, 0x0426, 0x0418, 0x0004,
+    0x1402, 0xFE00, 0x0424, 0x0414, 0x0003, 0x1402, 0xFE00, 0x0400,
+    0xFE00, 0x0400, 0xFE00, 0x07EA, 0x03F4, 0x0002, 0x03FF, 0x0003,
+    0x03F7, 0x03FC, 0x0002, 0x1BFF, 0x03FF, 0xFE00, 0x001A, 0x0010,
+    0x0007, 0x03FA, 0x0008, 0x000B, 0x0005, 0x0006, 0x03F8, 0x0001,
+    0x0401, 0x03FD, 0x0002, 0x0002, 0x0001, 0x0001, 0x03FF, 0x1FFF,
+    0x0401, 0x0BFF, 0xFE00, 0x043E, 0x003A, 0x03F7, 0x03F7, 0x03F4,
+    0x002E, 0x0029, 0x03F6, 0x03F7, 0x03FB, 0x03FD, 0x03FA, 0x03F8,
+    0x03F9, 0x001C, 0x0010, 0x03FC, 0x03FB, 0x03FC, 0x03FC, 0x03FF,
+    0x07FF, 0x03FE, 0x03FD, 0x03FE, 0x03FE, 0x0005, 0x0001, 0x07FF,
+    0x03FF, 0x03FF, 0x03FF, 0x17FF, 0x03FF, 0xFE00, 0x0400, 0xFE00,
+    0x0462, 0x005B, 0x03EA, 0x0008, 0x03E4, 0x0049, 0x0042, 0x03E8,
+    0x0009, 0x0002, 0x03FD, 0x0002, 0x0007, 0x03F0, 0x002C, 0x0019,
+    0x03F5, 0x0005, 0x0001, 0x03FC, 0x0FFE, 0x0001, 0x0002, 0x03FB,
+    0x0008, 0x0003, 0x03FE, 0x0001, 0x07FF, 0x0FFF, 0x0BFF, 0x23FF,
+    0x03FF, 0xFE00, 0x0400, 0xFE00, 0x0002, 0x0001, 0x03EF, 0x0002,
+    0x03EC, 0x0001, 0x07F0, 0x0003, 0x0001, 0x0401, 0x0001, 0x03FB,
+    0x0BFD, 0x27FE, 0x0BFF, 0xFE00, 0x07E6, 0x03F1, 0x03FB, 0x07FA,
+    0x03F6, 0x03FC, 0x03FC, 0x17FF, 0x03FE, 0x03FF, 0x2FFF, 0xFE00,
+    0x0490, 0x0086, 0x0C6C, 0x0061, 0x1C40, 0x0025, 0x2C0C, 0x0004,
+    0xFE00, 0x0400, 0xFE00, 0x0430, 0x002D, 0x000D, 0x000D, 0x0011,
+    0x0024, 0x0021, 0x000E, 0x000D, 0x0008, 0x0004, 0x0009, 0x000B,
+    0x000A, 0x0015, 0x000C, 0x0006, 0x0007, 0x0006, 0x0005, 0x0002,
+    0x0001, 0x0001, 0x0003, 0x0004, 0x0003, 0x0003, 0x0004, 0x0001,
+    0x0001, 0x0001, 0x0002, 0x0002, 0x0001, 0x0001, 0x0801, 0x0001,
+    0x0001, 0x0001, 0xFE00, 0x0400, 0xFE00, 0x001C, 0x0011, 0x0005,
+    0x0002, 0x0006, 0x000C, 0x0005, 0x0005, 0x0002, 0x0C01, 0x0001,
+    0x0002, 0x0001, 0x0001, 0x2801, 0xFE00, 0x07DE, 0x03EC, 0x0005,
+    0x0002, 0x0006, 0x03F1, 0x03FA, 0x0005, 0x0002, 0x0C01, 0x0001,
+    0x03FE, 0x03FE, 0x0001, 0x2BFF, 0xFE00, 0x045A, 0x0053, 0x03D3,
+    0x0012, 0x03C8, 0x0043, 0x003C, 0x03CF, 0x0012, 0x0004, 0x03FA,
+    0x0005, 0x000F, 0x03DF, 0x0028, 0x0017, 0x03EA, 0x000A, 0x0003,
+    0x03F8, 0x0001, 0x0001, 0x0001, 0x03FC, 0x0002, 0x0004, 0x03F5,
+    0x0007, 0x0002, 0x03FC, 0x0002, 0x0001, 0x03FD, 0x0401, 0x03FF,
+    0x03FE, 0x0001, 0x07FE, 0x0401, 0x03FF, 0x0BFF, 0x0401, 0x03FE,
+    0x03FF, 0x1FFF, 0xFE00, 0x0400, 0xFE00, 0x04B6, 0x00A8, 0x0C87,
+    0x007A, 0x1C50, 0x002F, 0x2C10, 0x0005, 0xFE00, 0x0400, 0xFE00,
+    0x0012, 0x000A, 0x0008, 0x03FA, 0x000A, 0x0007, 0x0003, 0x0008,
+    0x03F8, 0x0001, 0x0401, 0x03FD, 0x0002, 0x0001, 0x0001, 0x0001,
+    0x03FF, 0x0001, 0x1BFF, 0x0001, 0x0FFF, 0xFE00, 0x0018, 0x000F,
+    0x07FC, 0x040B, 0x0005, 0x07FB, 0x0FFE, 0x0402, 0x0001, 0x07FF,
+    0x2401, 0xFE00, 0x0476, 0x006C, 0x0008, 0x0008, 0x000B, 0x0058,
+    0x004E, 0x0009, 0x0009, 0x0005, 0x0003, 0x0005, 0x0007, 0x0006,
+    0x0034, 0x001E, 0x0004, 0x0005, 0x0004, 0x0003, 0x0001, 0x0401,
+    0x0002, 0x0003, 0x0002, 0x0002, 0x000A, 0x0003, 0x0401, 0x0001,
+    0x0001, 0x0001, 0x1401, 0x0001, 0xFE00, 0x0400, 0xFE00, 0x0494,
+    0x008A, 0x03DF, 0x000D, 0x03D6, 0x006F, 0x0064, 0x03DC, 0x000D,
+    0x0003, 0x03FC, 0x0003, 0x000B, 0x03E7, 0x0042, 0x0027, 0x03F0,
+    0x0007, 0x0002, 0x03FB, 0x0401, 0x07FD, 0x0001, 0x0003, 0x03F8,
+    0x000D, 0x0004, 0x03FD, 0x0002, 0x07FE, 0x0401, 0x03FF, 0x03FF,
+    0x0001, 0x07FF, 0x0BFF, 0x0BFF, 0x0BFF, 0x03FF, 0x1FFF, 0xFE00,
+    0x0400, 0xFE00, 0x07E0, 0x03ED, 0x03F6, 0x0001, 0x03F4, 0x03F3,
+    0x03FA, 0x03F7, 0x0001, 0x13FD, 0x03FE, 0x03FF, 0x03FF, 0x27FF,
+    0x03FF, 0x07FF, 0xFE00, 0x0004, 0x0002, 0x0001, 0x0401, 0x0002,
+    0x0401, 0xFE00, 0x04DA, 0x00C9, 0x0CA2, 0x0092, 0x1C60, 0x0039,
+    0x2C13, 0x0006, 0xFE00, 0x0400, 0xFE00, 0x04F8, 0x00E6, 0x03FB,
+    0x03FC, 0x03FA, 0x00B9, 0x00A7, 0x03FB, 0x03FB, 0x03FD, 0x03FF,
+    0x03FD, 0x03FC, 0x03FD, 0x006E, 0x0041, 0x03FE, 0x03FE, 0x03FE,
+    0x03FE, 0x0FFF, 0x03FF, 0x03FF, 0x03FF, 0x0015, 0x0007, 0xFE00,
+    0x0400, 0xFE00, 0x07E2, 0x03EF, 0x0008, 0x0002, 0x0009, 0x03F4,
+    0x03FB, 0x0007, 0x0003, 0x0C01, 0x0002, 0x03FE, 0x03FF, 0x0001,
+    0x2401, 0x03FF, 0xFE00, 0x07F6, 0x03FB, 0x0002, 0x0403, 0x03FC,
+    0x03FF, 0x0002, 0xFE00, 0x04CA, 0x00BC, 0x0022, 0x042C, 0x0098,
+    0x0088, 0x0026, 0x07EE, 0x000A, 0x03EB, 0x041A, 0x005A, 0x0035,
+    0x0011, 0x07F1, 0x000D, 0x07FB, 0x0408, 0x03F5, 0x0408, 0x0011,
+    0x0006, 0x0003, 0x07FB, 0x0005, 0x07FA, 0x0004, 0x0006, 0x03FB,
+    0x0403, 0x03FD, 0x0401, 0x07FF, 0x0002, 0x07FC, 0x0006, 0x0004,
+    0x03FE, 0x0401, 0x0FFF, 0x0003, 0x0002, 0x03FF, 0x0801, 0xFE00,
+    0x0400, 0xFE00, 0x04B4, 0x00A6, 0x0C86, 0x0078, 0x1C50, 0x002F,
+    0x2C0F, 0x0005, 0xFE00, 0x0400, 0xFE00, 0x07F6, 0x03FB, 0x03FA,
+    0x07F9, 0x03FC, 0x03FF, 0x03FB, 0x07FE, 0x07FE, 0x07FE, 0x0BFF,
+    0x07FF, 0x17FF, 0xFE00, 0x07DC, 0x03EB, 0x0402, 0x07F1, 0x03F9,
+    0x0402, 0x0C01, 0x07FE, 0x03FE, 0x2FFF, 0xFE00, 0x04A6, 0x009A,
+    0x0003, 0x0002, 0x0004, 0x007C, 0x0070, 0x0004, 0x0002, 0x0C02,
+    0x0002, 0x004A, 0x002B, 0x0001, 0x0001, 0x240E, 0x0005, 0xFE00,
+    0x0400, 0xFE00, 0x0494, 0x0089, 0x000C, 0x03F6, 0x0010, 0x006E,
+    0x0064, 0x000E, 0x03F6, 0x0003, 0x0403, 0x03F8, 0x0009, 0x0042,
+    0x0027, 0x0006, 0x03FB, 0x0002, 0x07FF, 0x0C01, 0x03FE, 0x0003,
+    0x000D, 0x0004, 0x0001, 0x03FF, 0xFE00, 0x0400, 0xFE00, 0x0010,
+    0x000A, 0x0009, 0x03FD, 0x000B, 0x0007, 0x0003, 0x0008, 0x03FC,
+    0x0FFF, 0x0003, 0x0001, 0x0001, 0x0001, 0x2401, 0xFE00, 0x07FC,
+    0x03FF, 0x03F4, 0x0004, 0x03F2, 0x03FF, 0x07F5, 0x0005, 0x0C02,
+    0x03FC, 0x0BFE, 0x0001, 0x1C01, 0x03FF, 0x0BFF, 0xFE00, 0x046A,
+    0x0062, 0x0C4F, 0x0047, 0x1C2F, 0x001C, 0x2C09, 0x0003, 0xFE00,
+    0x0400, 0xFE00, 0x0446, 0x0042, 0x0002, 0x0009, 0x0003, 0x0035,
+    0x0030, 0x0003, 0x0009, 0x0008, 0x0003, 0x0009, 0x0008, 0x0001,
+    0x001F, 0x0012, 0x0001, 0x0005, 0x0007, 0x0003, 0x03FF, 0x03FE,
+    0x03FF, 0x0002, 0x0005, 0x0002, 0x0406, 0x0002, 0x0401, 0x0002,
+    0x0001, 0x07FE, 0x03FF, 0x03FE, 0x03FE, 0x0401, 0x0001, 0x1BFF,
+    0x03FE, 0x03FE, 0x03FF, 0x1BFF, 0xFE00, 0x0400, 0xFE00, 0x07F8,
+    0x03FC, 0x03FD, 0x03FF, 0x03FC, 0x03FD, 0x03FF, 0x03FD, 0x03FF,
+    0x13FF, 0xFE00, 0x0012, 0x000B, 0x0007, 0x0002, 0x0008, 0x0007,
+    0x0003, 0x0006, 0x0003, 0x0C01, 0x0002, 0x0001, 0x0001, 0x0001,
+    0x2401, 0xFE00, 0x0454, 0x004E, 0x03F1, 0x0013, 0x03EC, 0x003F,
+    0x0039, 0x03EF, 0x0013, 0x03F1, 0x000A, 0x03EF, 0x0010, 0x03F5,
+    0x0025, 0x0016, 0x03F9, 0x000A, 0x03F3, 0x000D, 0x03FA, 0x0003,
+    0x03FB, 0x0008, 0x03F7, 0x0005, 0x03FD, 0x0007, 0x0002, 0x03FF,
+    0x0002, 0x03FC, 0x0005, 0x03FB, 0x0004, 0x03FF, 0x03FE, 0x0003,
+    0x03FD, 0x0003, 0x03FE, 0x0001, 0x0BFF, 0x0002, 0x03FE, 0x0002,
+    0x03FE, 0x03FE, 0x0001, 0x03FF, 0x0001, 0x0BFF, 0x0001, 0x03FF,
+    0xFE00, 0x0400, 0xFE00, 0x04F0, 0x00DE, 0x0007, 0x0409, 0x00B2,
+    0x00A1, 0x0008, 0x03FF, 0x0FFF, 0x0005, 0x006A, 0x003E, 0x0003,
+    0x2401, 0x0015, 0x0007, 0xFE00, 0x0400, 0xFE00, 0x07F6, 0x03FB,
+    0x0001, 0x0401, 0x03FC, 0x03FF, 0x0001, 0xFE00, 0x0016, 0x000D,
+    0x03FC, 0x07FC, 0x0009, 0x0004, 0x03FD, 0x17FF, 0x0001, 0x0001,
+    0xFE00, 0x04C0, 0x00B1, 0x000B, 0x040E, 0x008E, 0x0080, 0x000C,
+    0x1408, 0x0055, 0x0032, 0x0005, 0x2402, 0x0010, 0x0005, 0x0001,
+    0xFE00, 0x0400, 0xFE00, 0x0486, 0x007B, 0x000B, 0x040E, 0x0063,
+    0x0059, 0x000C, 0x1408, 0x003B, 0x0023, 0x0005, 0x2402, 0x000B,
+    0x0004, 0x0001, 0xFE00, 0x0400, 0xFE00, 0x07EE, 0x03F6, 0x0FF9,
+    0x03FD, 0x1FFF, 0x03FF, 0xFE00, 0x0028, 0x0019, 0x07FF, 0x0412,
+    0x0008, 0x07FF, 0x1403, 0x0002, 0x2C01, 0xFE00, 0x0456, 0x004F,
+    0x0007, 0x0409, 0x0040, 0x0039, 0x0008, 0x1405, 0x0026, 0x0016,
+    0x0003, 0x2401, 0x0007, 0x0002, 0xFE00, 0x0400, 0xFE00, 0x0442,
+    0x003E, 0x0002, 0x0403, 0x0032, 0x002D, 0x0002, 0x1401, 0x001D,
+    0x0011, 0x0001, 0x2805, 0x0002, 0xFE00, 0x0400, 0xFE00, 0x07F6,
+    0x03FB, 0x03FF, 0x07FF, 0x03FC, 0x03FF, 0x03FF, 0xFE00, 0x0016,
+    0x000D, 0x0003, 0x0404, 0x0009, 0x0004, 0x0003, 0x1401, 0x0001,
+    0x0001, 0xFE00, 0x042E, 0x002C, 0x0004, 0x0406, 0x0023, 0x0020,
+    0x0004, 0x1403, 0x0014, 0x000C, 0x0002, 0x2401, 0x0004, 0x0001,
+    0xFE00, 0x0400, 0xFE00, 0x0418, 0x0016, 0x0004, 0x0406, 0x0011,
+    0x0010, 0x0004, 0x1403, 0x000A, 0x0006, 0x0002, 0x2401, 0x0002,
+    0xFE00, 0x0400, 0xFE00, 0x07FE, 0xFE00, 0x0000, 0xFE00, 0x0404,
+    0x0004, 0x0002, 0x0402, 0x0002, 0x0002, 0x0002, 0x1401, 0x0001,
+    0x0001, 0x0001, 0xFE00, 0x0400, 0xFE00, 0x0400, 0xFE00, 0x0400,
+    0xFE00
+};
+
+class TestMasher : public Oddlib::Masher
+{
+public:
+    using Oddlib::Masher::Masher;
+
+    bool CompareDecodedFrameData(const std::vector<u16>& expected)
+    {
+        return mDecodedVideoFrameData == expected;
+    }
+
+    using Oddlib::Masher::decode_audio_frame;
+};
+
+// Audio and video test
+TEST(Masher, all_colours_low_compression_15fps_8bit_mono_high_compression_5_frames_interleave)
+{
+    TestMasher masher(std::make_unique<Oddlib::MemoryStream>(get_all_colours_low_compression_15fps_8bit_mono_high_compression_5_frames_interleave()));
+    ASSERT_TRUE(masher.HasVideo());
+    ASSERT_TRUE(masher.HasAudio());
+    ASSERT_EQ(258u, masher.Width());
+    ASSERT_EQ(200u, masher.Height());
+    std::vector<u32> pixelBuffer(masher.Width() * masher.Height());
+    ASSERT_TRUE(masher.Update(pixelBuffer.data(), nullptr));
+    ASSERT_FALSE(masher.Update(pixelBuffer.data(), nullptr));
+    const std::vector<u16> expected(std::begin(kLowCompression16BitExpected), std::end(kLowCompression16BitExpected));
+    ASSERT_TRUE(masher.CompareDecodedFrameData(expected));
+}
+
+// All video compression tests
+TEST(Masher, all_colours_high_compression_30_fps)
+{
+    /*
+    TestMasher masher(get_all_colours_high_compression_30_fps());
+    ASSERT_EQ(true, masher.HasVideo());
+    ASSERT_EQ(false, masher.HasAudio());
+    ASSERT_EQ(258, masher.Width());
+    ASSERT_EQ(200, masher.Height());
+    std::vector<u32> pixelBuffer(masher.Width() * masher.Height());
+    ASSERT_EQ(true, masher.Update(pixelBuffer.data(), nullptr));
+    ASSERT_EQ(false, masher.Update(pixelBuffer.data(), nullptr));
+    const std::vector<u16> expected(std::begin(kLowCompression16BitExpected), std::end(kLowCompression16BitExpected));
+    ASSERT_EQ(true, masher.CompareDecodedFrameData(expected));
+    */
+}
+
+TEST(Masher, all_colours_low_compression_30_fps)
+{
+    TestMasher masher(std::make_unique<Oddlib::MemoryStream>(get_all_colours_low_compression_30_fps()));
+    ASSERT_TRUE(masher.HasVideo());
+    ASSERT_FALSE(masher.HasAudio());
+    ASSERT_EQ(258u, masher.Width());
+    ASSERT_EQ(200u, masher.Height());
+    std::vector<u32> pixelBuffer(masher.Width() * masher.Height());
+    ASSERT_TRUE(masher.Update(pixelBuffer.data(), nullptr));
+    ASSERT_FALSE(masher.Update(pixelBuffer.data(), nullptr));
+    const std::vector<u16> expected(std::begin(kLowCompression16BitExpected), std::end(kLowCompression16BitExpected));
+    ASSERT_TRUE(masher.CompareDecodedFrameData(expected));
+}
+
+TEST(Masher, all_colours_max_compression_30_fps)
+{
+    //Oddlib::Masher masher(get_all_colours_max_compression_30_fps());
+    //while (masher.Update());
+}
+
+TEST(Masher, all_colours_medium_compression_30_fps)
+{
+    //Oddlib::Masher masher(get_all_colours_medium_compression_30_fps());
+    //while (masher.Update());
+}
+
+TEST(Masher, all_colours_min_compression_30_fps)
+{
+    //Oddlib::Masher masher(get_all_colours_min_compression_30_fps());
+    //while (masher.Update());
+}
+
+// All audio compression and formats tests
+TEST(Masher, mono_8_high_compression_all_samples)
+{
+    //Oddlib::Masher masher(get_mono_8_high_compression_all_samples());
+    //while (masher.Update());
+}
+
+TEST(Masher, mono_8_low_compression_all_samples)
+{
+    //Oddlib::Masher masher(get_mono_8_low_compression_all_samples());
+    //while (masher.Update());
+}
+
+TEST(Masher, mono_16_high_compression_all_samples)
+{
+    //Oddlib::Masher masher(get_mono_16_high_compression_all_samples());
+    //while (masher.Update());
+}
+
+TEST(Masher, mono_16_low_compression_all_samples)
+{
+    // Oddlib::Masher masher(get_mono_16_low_compression_all_samples());
+    //while (masher.Update());
+}
+
+TEST(Masher, stereo_8_high_compression_all_samples)
+{
+    // Oddlib::Masher masher(get_stereo_8_high_compression_all_samples());
+    // while (masher.Update());
+}
+
+TEST(Masher, stereo_8_low_compression_all_samples)
+{
+    // Oddlib::Masher masher(get_stereo_8_low_compression_all_samples());
+    // while (masher.Update());
+}
+
+TEST(Masher, stereo_16_high_compression_all_samples)
+{
+    TestMasher masher(std::make_unique<Oddlib::MemoryStream>(get_stereo_16_high_compression_all_samples()));
+    ASSERT_FALSE(masher.HasVideo());
+    ASSERT_TRUE(masher.HasAudio());
+
+    // Valgrind reported we read/write 2 bytes out of bounds, so must be off by one error?
+    std::vector<u8> audioBuffer((masher.SingleAudioFrameSizeSamples() * 4));
+
+    // TODO: Verify the content of each decoded audio frame / VS \ on each channel
+    while (masher.Update(nullptr, audioBuffer.data()));
+
+
+}
+
+TEST(Masher, stereo_16_low_compression_all_samples)
+{
+    // Oddlib::Masher masher(get_stereo_16_low_compression_all_samples());
+    // while (masher.Update());
+}
+
+// TODO: Has a bug in release
+#ifdef _DEBUG
+TEST(Masher, Decode16BitStereoAudio)
+{
+    TestMasher masher;
+    Oddlib::SetAudioFrameSizeBytesAndBits(2);
+    Oddlib::init_Snd_tbl();
+
+    std::vector<u16> rawFrameBuffer;
+    const std::vector<u8> kCompressed = get_compressed_audio();
+    rawFrameBuffer.resize(kCompressed.size() / sizeof(u16));
+    memcpy(rawFrameBuffer.data(), kCompressed.data(), kCompressed.size());
+
+
+    const signed int numSamplesPerFrame = 2940; //  sampleRate / fps, i.e 44100/15=2940
+    const int numChannels = 2;
+    std::vector<u16> outPtr(numSamplesPerFrame*numChannels);
+
+    masher.decode_audio_frame(rawFrameBuffer.data(), outPtr.data(), numSamplesPerFrame);
+
+    const std::vector<u8> kExpected = get_expected_audio();
+    ASSERT_EQ(kExpected.size()/sizeof(u16), outPtr.size());
+
+    std::vector<u16> expected16(outPtr.size());
+    memcpy(expected16.data(), kExpected.data(), kExpected.size());
+    ASSERT_TRUE(memcmp(expected16.data(), outPtr.data(), outPtr.size()*sizeof(u16)) == 0);
+}
+#endif
